Control.Coapplicative

Documentation

type Applicative' f = (Apply f, Applicative f)

Coapplicative

class Coapply f => Coapplicative f where

coapply . fmap Right = Right
either (f . copure) (g . copure) . coapply = either f g . copure

Methods

copure :: f a -> a

Coapply

class Functor f => Coapply f where

coapply . fmap Left = Left
coeval . fmap Left = fmap Left
coeval . fmap Right = fmap Right

Methods

coapply :: f (a + b) -> f a + f b

apply :: Apply f => (f a, f b) -> f (a, b)

select :: Functor f => f (a, b) -> (f a, f b)

coselect :: Functor f => (f a + f b) -> f (a + b)

coeval :: Coapply f => f (a + b) -> f (a + b)

Evaluate a coapplicative expression, exiting on the first Left if it exists.

>>> coeval $ Left "foo" :| [Right "bar"]
Left "foo" :| []
>>> coeval $ Right "foo" :| [Right "bar"]
Right "foo" :| [Right "bar"]

Select

type Select f = (Coapply f, Apply f)

eitherS :: Select f => f (a + b) -> f (a -> c) -> f (b -> c) -> f c

branch :: Select f => f Bool -> f a -> f a -> f a

Branch on a boolean value, skipping unnecessary effects.

>>> head $ branch (True :| []) (print "foo" :| []) (print "bar" :| [])
"foo"

bindBool :: Select f => f Bool -> (Bool -> f a) -> f a

Selective

type Selective f = (Coapply f, Applicative' f)

apS :: Selective f => f (a -> b) -> f a -> f b

Recover the application operator <*> from <*?. Rigid selective functors satisfy the law <*> = apS and furthermore, the resulting applicative functor satisfies all laws of Applicative:

Identity:

pure id <*> v = v

Homomorphism:

pure f <*> pure x = pure (f x)

Interchange:

u <*> pure y = pure ($y) <*> u

Composition:

(.) <$> u <*> v <*> w = u <*> (v <*> w)

(<*?) :: Selective f => f (a + b) -> f (a -> b) -> f b infixl 4

eliminate :: Eq a => Selective f => a -> f b -> f (a + b) -> f (a + b)

Eliminate the a from f (a + b) and replace with the provided f b.

>>> eliminate 1.0 ["foo","bar"] [Left 1.0] :: [Float + String]
[Right "foo",Right "bar"]
>>> eliminate 1.0 ["foo","bar"] [Left 3.4] :: [Float + String]
[Left 3.4]

whenS :: Selective f => f Bool -> f () -> f ()

Conditionally perform an effect.

whileS :: Selective f => f Bool -> f ()

Keep checking an effectful condition while it holds.

fromMaybeS :: Selective f => f a -> f (Maybe a) -> f a

A lifted version of fromMaybe.

untilRight :: Monoid a => Selective f => f (a + b) -> f (a, b)

Keep running an effectful computation until it returns a Right value, collecting the Left's using a supplied Monoid instance.

(<||>) :: Selective f => f Bool -> f Bool -> f Bool

A lifted version of lazy Boolean OR.

(<&&>) :: Selective f => f Bool -> f Bool -> f Bool

A lifted version of lazy Boolean AND.

anyS :: Selective f => (a -> f Bool) -> [a] -> f Bool

A lifted version of any. Retains the short-circuiting behaviour.

allS :: Selective f => (a -> f Bool) -> [a] -> f Bool

A lifted version of all. Retains the short-circuiting behaviour.

foldS :: (Selective f, Foldable t, Monoid b) => t (f (a + b)) -> f (a + b)

Generalised folding with the short-circuiting behaviour.

orElse :: Semigroup a => Selective f => f (a + b) -> f (a + b) -> f (a + b)

Return the first Right value. If both are Left's, accumulate errors.

andAlso :: Semigroup b => Selective f => f (a + b) -> f (a + b) -> f (a + b)

Accumulate the Right values, or return the first Left.

>>> andAlso [Right "foo", Right "bar", Right "baz"] [Right "!"]
[Right "foo!",Right "bar!",Right "baz!"]
>>> andAlso [Right "foo", Right "bar", Left 'e'] [Right "!"]
[Right "foo!",Right "bar!"]
>>> andAlso [Right "foo", Right "bar", Left 'e'] [Left 'f']
[Left 'f',Left 'f']

appendl :: Semigroup a => a -> (a + b) -> a + b

Append two semigroup values or return the Right one.