Data.Functor.Coapply

Types

type (+) = Either

Coapply

class Functor f => Coapply f where

coapply . fmap Left = Left
coeval . fmap Left = fmap Left
coeval . fmap Right = fmap Right

Methods

coapply :: f (a + b) -> f a + f b

apply :: Apply f => (f a, f b) -> f (a, b)

select :: Functor f => f (a, b) -> (f a, f b)

coselect :: Functor f => (f a + f b) -> f (a + b)

coeval :: Coapply f => f (a + b) -> f (a + b)

Evaluate a coapplicative expression, exiting on the first Left if it exists.

>>> coeval $ Left "foo" :| [Right "bar"]
Left "foo" :| []
>>> coeval $ Right "foo" :| [Right "bar"]
Right "foo" :| [Right "bar"]

appendl :: Semigroup a => a -> (a + b) -> a + b

Append two semigroup values or return the Right one.

concatl :: [a + b] -> [a]

concatr :: [a + b] -> [b]

concatl' :: [a + b] -> [a]

concatr' :: [a + b] -> [b]

Select

type Select f = (Coapply f, Apply f)

eitherS :: Select f => f (a + b) -> f (a -> c) -> f (b -> c) -> f c

branch :: Select f => f Bool -> f a -> f a -> f a

Branch on a boolean value, skipping unnecessary effects.

>>> head $ branch (True :| []) (print "foo" :| []) (print "bar" :| [])
"foo"

bindBool :: Select f => f Bool -> (Bool -> f a) -> f a

Distributive1

class Functor g => Distributive1 g where

Minimal complete definition

distribute1 | collect1

Methods

distribute1 :: Coapply f => f (g a) -> g (f a)

The dual of sequence1

distribute1 = collect1 id
distribute1 . distribute1 = id

collect1 :: Coapply f => (a -> g b) -> f a -> g (f b)

collect1 f = distribute1 . fmap f
fmap f = runIdentity . collect1 (Identity . f)
fmap distribute1 . collect1 f = getCompose . collect1 (Compose . f)

cotraverse1 :: Coapply f => Distributive1 g => (f a -> b) -> f (g a) -> g b

The dual of traverse1

cotraverse1 f = fmap f . distribute1