Data.Map.NonEmpty.Optic

Description

Left and right carriers are characterized by their representation functors (Rep and Corep). Iso sits at the top where both are trivial (Identity). Adjoint sits at the bottom where they form a full adjunction (Corep p ⊣ Rep p). A left-indexed (Ix) optic threads an index through the left adjoint functor (,) k, while a right-indexed (Cx) optic threads a coindex through the right adjoint functor (->) k.

Documentation

fromIxfold1 :: Ord k => Monoid k => AIxfold (Map k a) k s a -> s -> Map k a

O(1). Create a NEMap from an Ixfold1.

at :: Ord k => k -> Traversal0' (NEMap k a) a

O(log n). Affine traversal into the value at a key of a NEMap.

>>> Map.fromList ((0,"hello") :| [(1,"world")]) ^. at 0
"hello"
>>> Map.fromList ((0,"hello") :| [(1,"world")]) ^. at 1
"world"
>>> Map.fromList ((67,'C') :| []) & at 99 ..~ toLower
fromList ((67,'C') :| [])
>>> Map.fromList ((67,'C') :| []) & at 67 ..~ toLower
fromList ((67,'c') :| [])

iat :: Ord k => k -> Ixtraversal0' k (NEMap k a) a

O(log n). Indexed affine traversal into the value at a key of a NEMap.

>>> Map.fromList ((0,"hello") :| [(1,"world")]) ^% iat (0 :: Int)
(Just 0,"hello")
>>> Map.fromList ((1,"world") :| []) ^% iat (1 :: Int)
(Just 1,"world")
>>> Map.fromList ((67,'C') :| []) & iat 99 %~ chr
fromList ((67,'C') :| [])
>>> Map.fromList ((67,'C') :| []) & iat 67 %%~ const toLower
fromList ((67,'c') :| [])
>>> iat 2 %%~ (\i x -> if odd i then not x else x) $ Map.fromList ((1,True) :| [(2,False)])
fromList ((1,True) :| [(2,False)])

values :: Fold1 (NEMap k a) a

O(n). Fold1 over the values of a NEMap, in ascending key order.

Subject to list fusion.

>>> nelists values (Map.fromList ((5,'a') :| [(3,'b')]))
'b' :| "a"

imapped :: Semigroup k => Ixsetter k (NEMap k a) (NEMap k b) a b

O(n). Ixsetter over the values of a NEMap.

ifiltered :: Semigroup k => Ixsetter k (NEMap k a) (Map k a) a Bool

O(n). Ixsetter filtering the values of a NEMap.

itraversed :: (Ord k, Semigroup k) => Ixtraversal1 k (NEMap k a) (NEMap k b) a b

O(n). Ixtraversal1 over the values of a NEMap.

ifolded :: Semigroup k => Ixfold k (NEMap k a) a

O(n). Ixfold1 over the values of a NEMap.

altered :: Ord k => k -> Setter (NEMap k a) (Map k a) (Maybe a) (Maybe a)

O(log n). Alter the value at a specific key.

altered can be used to insert, delete, or update a value in a Map.

>>> let f _ = Nothing
>>> over (altered 7) f $ Map.fromList ((5,'a') :| [(3,'b')])
fromList [(3,'b'),(5,'a')] 
>>> over (altered 5) f $ Map.fromList ((5,'a') :| [(3,'b')])
fromList [(3,'b')]

>>> let f _ = Just "c"
>>> over (altered 7) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"a"),(7,"c")]
>>> over (altered 5) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"c")]

See also alter.

ialtered :: Ord k => Ixsetter k (NEMap k a) (Map k a) (Maybe a) (Maybe a)

O(log n). Alter the value at a specific key.

ialtered can be used to insert, delete, or update a value in a Map.

>>> let f _ _ = Nothing
>>> iover (ialtered 7) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"a")]
>>> iover (ialtered 5) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b")]

>>> let f i _ = if i == 7 then Just "c" else Nothing
>>> iover (ialtered 7) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"a"),(7,"c")]
>>> iover (ialtered 5) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b")]

See also alter.

alteredF :: Ord k => k -> Lens (NEMap k a) (Map k a) (Maybe a) (Maybe a)

O(log n). Lens into Maybe of a value at a key of a NEMap.

ialteredF :: Ord k => Ixlens k (NEMap k a) (Map k a) (Maybe a) (Maybe a)

O(log n). Alter the value at a specific key.

>>> let f _ _ = Just "c"
>>> iover ialteredF f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"c")]

See also alterF.

adjusted :: Ord k => Ixsetter k (NEMap k a) (NEMap k a) a a

O(log n). Adjust a value at a specific key.

>>> let f key x = (show key) ++ ":new " ++ x
>>> iover (adjusted 5) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList ((3,"b") :| [(5,"5:new a")])
>>> iover (adjusted 7) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList ((3,"b") :| [(5,"a")])

See also adjustWithKey.

updated :: Ord k => Ixsetter k (NEMap k a) (Map k a) a (Maybe a)

O(log n). Update a value at a specific key.

If (f k x) is Nothing, the element is deleted. If it is (Just y), the key k is bound to the new value y.

>>> let f k x = if x == "a" then Just ((show k) ++ ":new a") else Nothing
>>> iover (updated 5) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"5:new a")]
>>> iover (updated 7) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(3,"b"),(5,"a")]
>>> iover (updated 3) f $ Map.fromList ((5,"a") :| [(3,"b")])
fromList [(5,"a")]

See also updateWithKey.

foundMin :: Ixview k (NEMap k a) a

O(1). Ixview into the value at the minimal key of a NEMap.

This function is asymptotically more efficient than lookedMin if you have a non-empty map.

foundMax :: Ixview k (NEMap k a) a

O(1). Ixview into the value at the maximal key of a NEMap.

This function is asymptotically more efficient than lookedMax if you have a non-empty map.

lookedLT :: Ord k => k -> Ixtraversal0' k (NEMap k a) a

O(log n). Indexed affine traversal into the value at the largest key smaller than the given one.

>>> ipreview (lookedLT 3) $ Map.fromList ((3,'a') :| [(5,'b')])
Nothing
>>> ipreview (lookedLT 4) $ Map.fromList ((3,'a') :| [(5,'b')])
Just (3,'a')

lookedLE :: Ord k => k -> Ixtraversal0' k (NEMap k a) a

O(log n). Indexed affine traversal into the value at the largest key smaller than or equal to the given one.

>>> ipreview (lookedLE 3) $ Map.fromList ((3,'a') :| [(5,'b')])
Just (3,'a')

lookedGE :: Ord k => k -> Ixtraversal0' k (NEMap k a) a

O(log n). Indexed affine traversal into the value at the largest key greater than or equal to the given one.

>>> ipreview (lookedGE 3) $ Map.fromList ((3,'a') :| [(5,'b')])
Just (3,'a')
>>> ipreview (lookedGE 4) $ Map.fromList ((3,'a') :| [(5,'b')])
Just (5,'b')
>>> ipreview (lookedGE 6) $ Map.fromList ((3,'a') :| [(5,'b')])
Nothing

lookedGT :: Ord k => k -> Ixtraversal0' k (NEMap k a) a

O(log n). Indexed affine traversal into the value at the largest key greater than the given one.

>>> ipreview (lookedGT 4) $ Map.fromList ((3,'a') :| [(5,'b')])
Just (5,'b')
>>> ipreview (lookedGT 5) $ Map.fromList ((3,'a') :| [(5,'b')])
Nothing

validated :: Ord k => Fold0 (NEMap k a) (NEMap k a)

O(n). Test if the internal map structure is valid.

>>> is validated $ Map.fromAscList ((3,'a') :| [(5,'b')])
True
>>> isnt validated $ Map.fromAscList ((5,'a') :| [(3,'b')])
True

See also valid.