Data.Map.Optic

Documentation

fromIxfold :: Ord k => Monoid k => AIxfold (Map k a) k s a -> s -> Map k a

O(1). Create a Map from an Ixfold.

at :: Ord k => k -> Traversal0' (Map k a) a

O(log n). Affine traversal into the value at a key of a Map.

>>> Map.fromList [(1,"world")] ^. at 0
""
>>> Map.fromList [(1,"world")] ^. at 1
"world"
>>> Map.fromList [(67,'C')] & at 99 ..~ toLower
fromList [(67,'C')]
>>> Map.fromList [(67,'C')] & at 67 ..~ toLower
fromList [(67,'c')]

iat :: Ord k => k -> Ixtraversal0' k (Map k a) a

O(log n). Indexed affine traversal into the value at a key of a Map.

>>> Map.fromList [(1,"world")] ^% iat 0
(Nothing,"")
>>> Map.fromList [(1,"world")] ^% iat 1
(Just 1,"world")
>>> Map.fromList [(67,'C')] & iat 99 %~ chr
fromList [(67,'C')]
>>> Map.fromList [(67,'C')] & iat 67 %%~ const toLower
fromList [(67,'c')]
>>> iat 2 %%~ (\i x -> if odd i then not x else x) $ Map.fromList [(1,True),(2,False)]
fromList [(1,True),(2,False)]

values :: Fold (Map k a) a

O(n). Fold over the values of a Map, in ascending key order.

Subject to list fusion.

>>> lists values $ Map.fromList [(1,3),(2,4)]
[3,4]

imapped :: Ixsetter k (Map k a) (Map k b) a b

O(n). Ixsetter over the values of a Map.

>>> iover imapped (+) $ Map.fromList [(1,3),(2,4)]
fromList [(1,4),(2,6)]

See also mapWithKey.

ifiltered :: Ixsetter k (Map k a) (Map k a) a Bool

O(n). Ixsetter filtering the values of a Map.

itraversed :: Ord k => Ixtraversal k (Map k a) (Map k b) a b

O(n). Ixtraversal over the values of a Map.

ifolded :: Ixfold k (Map k a) a

O(n). Ixfold over the values of a Map.

fromIxfold ifolded ≡ id

>>> ilists ifolded $ Map.fromList [(1,3),(2,4)]
[(1,3),(2,4)]
>>> ilists ifolded $ Map.empty
[]

altered :: Ord k => k -> Setter' (Map k a) (Maybe a)

O(log n). Alter the value at a specific key.

altered can be used to insert, delete, or update a value in a Map.

>>> let f _ = Nothing
>>> over (altered 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a")]
>>> over (altered 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b")]

>>> let f _ = Just "c"
>>> over (altered 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a"),(7,"c")]
>>> over (altered 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"c")]

See also alter.

ialtered :: Ord k => k -> Ixsetter' k (Map k a) (Maybe a)

O(log n). Alter the value at a specific key.

ialtered can be used to insert, delete, or update a value in a Map.

>>> let f _ _ = Nothing
>>> iover (ialtered 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a")]
>>> iover (ialtered 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b")]

>>> let f i _ = if i == 7 then Just "c" else Nothing
>>> iover (ialtered 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a"),(7,"c")]
>>> iover (ialtered 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b")]

See also alter.

alteredF :: Ord k => k -> Lens' (Map k a) (Maybe a)

O(log n). Lens into Maybe of a value at a key of a Map.

ialteredF :: Ord k => k -> Ixlens' k (Map k a) (Maybe a)

O(log n). Alter the value at a specific key.

>>> let f _ _ = Just "c"
>>> iover (ialteredF 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a"),(7,"c")]
>>> iover (ialteredF 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"c")]
>>> Map.fromList [(5,"a"), (3,"b")] ^%% ialteredF 5
[(5,Just "a")]

See also alterF.

adjusted :: Ord k => k -> Ixsetter' k (Map k a) a

O(log n). Adjust a value at a specific key.

>>> let f key x = (show key) ++ ":new " ++ x
>>> iover (adjusted 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"5:new a")]
>>> iover (adjusted 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a")]
>>> iover (adjusted 7) f Map.empty
fromList []

See also adjustWithKey.

updated :: Ord k => k -> Ixsetter k (Map k a) (Map k a) a (Maybe a)

O(log n). Update a value at a specific key.

If (f k x) is Nothing, the element is deleted. If it is (Just y), the key k is bound to the new value y.

>>> let f k x = if x == "a" then Just ((show k) ++ ":new a") else Nothing
>>> iover (updated 5) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"5:new a")]
>>> iover (updated 7) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(3,"b"),(5,"a")]
>>> iover (updated 3) f $ Map.fromList [(5,"a"), (3,"b")]
fromList [(5,"a")]

See also updateWithKey.

updateLooked :: Ord k => k -> Ixsetter k (Map k a) (Maybe a, Map k a) a (Maybe a)

O(log n). Lookup and update a value at a specific key.

Returns the changed value, if it is updated. Returns the original key value if the map entry is deleted.

>>> let f k x = if x == "a" then Just ((show k) ++ ":new a") else Nothing
>>> iover (updateLooked 5) f $ Map.fromList [(5,"a"), (3,"b")]
(Just "5:new a",fromList [(3,"b"),(5,"5:new a")])
>>> iover (updateLooked 7) f $ Map.fromList [(5,"a"), (3,"b")]
(Nothing,fromList [(3,"b"),(5,"a")])
>>> iover (updateLooked 3) f $ Map.fromList [(5,"a"), (3,"b")]
(Just "b",fromList [(5,"a")])

See also updateLookupWithKey.

lookedMin :: Ixfold0 k (Map k a) a

O(log n). Ixoption into the value at the minimal key of a Map.

lookedMax :: Ixfold0 k (Map k a) a

O(log n). Ixoption into the value at the maximal key of a Map.

lookedLT :: Ord k => k -> Ixtraversal0' k (Map k a) a

O(log n). Indexed affine traversal into the value at the largest key smaller than the given one.

>>> ipreview (lookedLT 3) (Map.fromList [(3,'a'), (5,'b')])
Nothing
>>> ipreview (lookedLT 4) (Map.fromList [(3,'a'), (5,'b')])
Just (3,'a')

lookedLE :: Ord k => k -> Ixtraversal0' k (Map k a) a

O(log n). Indexed affine traversal into the value at the largest key smaller than or equal to the given one.

>>> ipreview (lookedLE 3) $ Map.fromList [(3,'a'), (5,'b')]
Just (3,'a')

lookedGE :: Ord k => k -> Ixtraversal0' k (Map k a) a

O(log n). Indexed affine traversal into the value at the largest key greater than or equal to the given one.

>>> ipreview (lookedGE 3) $ Map.fromList [(3,'a'), (5,'b')]
Just (3,'a')
>>> ipreview (lookedGE 4) $ Map.fromList [(3,'a'), (5,'b')]
Just (5,'b')
>>> ipreview (lookedGE 6) $ Map.fromList [(3,'a'), (5,'b')]
Nothing

lookedGT :: Ord k => k -> Ixtraversal0' k (Map k a) a

O(log n). Indexed affine traversal into the value at the largest key greater than the given one.

>>> ipreview (lookedGT 4) $ Map.fromList [(3,'a'), (5,'b')]
Just (5,'b')
>>> ipreview (lookedGT 5) $ Map.fromList [(3,'a'), (5,'b')]
Nothing

validated :: Ord k => Fold0 (Map k a) (Map k a)

O(n). Test if the internal map structure is valid.

>>> is validated (Map.fromAscList [(3,"b"), (5,"a")])
True
>>> isnt validated (Map.fromAscList [(5,"a"), (3,"b")])
True

See also valid.

Structural (via pattern functor)

depth :: Map k v -> Int

Compute the depth of a Map using its tree structure.

This operates on the internal BST representation via MapF, something not possible with the standard Map API.

>>> depth (Map.fromList [(1,'a'),(2,'b'),(3,'c'),(4,'d'),(5,'e')])
3

sizes :: Map k v -> [Int]

Collect the sizes stored at each internal node.

Useful for inspecting or validating the balance invariant.

>>> sizes (Map.fromList [(1,'a'),(2,'b'),(3,'c')])
[3,1,1]

rebalanced :: Map k v -> Map k v

Rebuild a Map from its pattern functor representation.

rebalanced = fromMapF . toMapF

Since Bin preserves the original sizes, this is the identity. Useful as a building block for transformations that modify the tree structure then reconstruct.