Data.Profunctor.Optic.Combinator

Constructors

parr :: Traversing p => (a -> b) -> p a b

coarr :: Cotraversing p => (a -> b) -> p a b

star :: Applicative f => Star f a a

costar :: Coapplicative f => Costar f a a

unstar :: Coapplicative f => Star f a b -> a -> b

uncostar :: Applicative f => Costar f a b -> a -> b

Indexed constructors

reix :: Profunctor p => (k1 -> k2) -> (k2 -> k1) -> Ixoptic p k1 s t a b -> Ixoptic p k2 s t a b

Map over the indices of an indexed optic.

See also reixed.

Since: 0.0.3

recx :: Profunctor p => (k1 -> k2) -> (k2 -> k1) -> Cxoptic p k1 s t a b -> Cxoptic p k2 s t a b

Map over the indices of a coindexed optic.

See also recxed.

Since: 0.0.3

ixsum :: Profunctor p => Ixoptic p k s t a b -> Ixoptic p (Sum k) s t a b

Lift a numeric index into a sum monoid.

Since: 0.0.3

cxsum :: Profunctor p => Cxoptic p k s t a b -> Cxoptic p (Sum k) s t a b

Lift a numeric co-index into a sum monoid.

Since: 0.0.3

ixany :: Profunctor p => Ixoptic p Bool s t a b -> Ixoptic p Any s t a b

ixhead :: Profunctor p => Ixoptic p i s t a b -> Ixoptic p (First i) s t a b

TODO: Document

Since: 0.0.3

ixlast :: Profunctor p => Ixoptic p i s t a b -> Ixoptic p (Last i) s t a b

TODO: Document

Since: 0.0.3

Miscellaneous optics

ixmap :: Profunctor p => (s -> a) -> (b -> t) -> Ixoptic p k s t a b

TODO: Document

Since: 0.0.3

cxmap :: Profunctor p => (s -> a) -> (b -> t) -> Cxoptic p k s t a b

TODO: Document

Since: 0.0.3

constl :: Profunctor p => b -> Optic p a c b c

constr :: Profunctor p => c -> Optic p a c a b

shiftl :: Profunctor p => Optic p b (c + d) (a + b) c

shiftr :: Profunctor p => Optic p (a, b) c b (c, d)

coercel :: Profunctor p => CoercingL p => Optic p c b a b

coercer :: Profunctor p => CoercingR p => Optic p a c a b

represent :: Representable p => ((a -> Rep p b) -> s -> Rep p t) -> Optic p s t a b

TODO: Document

ixrepresent :: Representable p => ((i -> a -> Rep p b) -> s -> Rep p t) -> Ixoptic p i s t a b

TODO: Document

Since: 0.0.3

corepresent :: Corepresentable p => ((Corep p a -> b) -> Corep p s -> t) -> Optic p s t a b

TODO: Document

cxrepresent :: Corepresentable p => ((i -> Corep p a -> b) -> Corep p s -> t) -> Cxoptic p i s t a b

TODO: Document

Since: 0.0.3

Operations on representable profunctors

(.) :: (b -> c) -> (a -> b) -> a -> c infixr 9

Function composition.

(.~) :: Optic (->) s t a b -> b -> s -> t infixr 4

Set the focus of an Optic.

(..~) :: Optic (->) s t a b -> (a -> b) -> s -> t infixr 4

Map over an Optic.

>>> (10,20) & first ..~ show
("10",20)

over :: Optic (->) s t a b -> (a -> b) -> s -> t

Map over an Optic.

over o id ≡ id 
over o f . over o g ≡ over o (f . g)
over . setter ≡ id
over . resetter ≡ id

>>> over fmapped (+1) (Just 1)
Just 2
>>> over fmapped (*10) [1,2,3]
[10,20,30]
>>> over first (+1) (1,2)
(2,2)
>>> over first show (10,20)
("10",20)

(%) :: Monoid i => Representable p => Ixoptic p i c1 c2 b1 b2 -> Ixoptic p i b1 b2 a1 a2 -> Ixoptic p i c1 c2 a1 a2 infixr 8

Monoidally combine indices between subsequent levels of optic.

Its precedence is one lower than that of function composition, which allows . to be nested in %.

If you only need the final index then use ..

>>> listsWithKey (ix "*" traversed . ix "+" traversed) ["foo", "bar"]
[("",'f'),("+",'o'),("++",'o'),("",'b'),("+",'a'),("++",'r')]
>>> listsWithKey (ix "*" traversed % ix "+" traversed) ["foo", "bar"]
[("",'f'),("+",'o'),("++",'o'),("*",'b'),("*+",'a'),("*++",'r')]

Since: 0.0.3

(%~) :: Monoid i => Ixoptic (->) i s t a b -> (i -> b) -> s -> t infixr 4

Set the focus of an indexed optic.

Note: This function is different from the equivalent in the lens package. The profunctor-optics equivalent of %~ from lens is ..~.

Since: 0.0.3

(%%~) :: Monoid i => Ixoptic (->) i s t a b -> (i -> a -> b) -> s -> t infixr 4

Map over an indexed optic.

Since: 0.0.3

overWithKey :: Monoid i => Ixoptic (->) i s t a b -> (i -> a -> b) -> s -> t

TODO: Document

Since: 0.0.3

(#) :: Monoid i => Corepresentable p => Cxoptic p i c1 c2 b1 b2 -> Cxoptic p i b1 b2 a1 a2 -> Cxoptic p i c1 c2 a1 a2 infixr 8

Compose two coindexed traversals, combining indices.

Its precedence is one lower than that of function composition, which allows . to be nested in #.

If you only need the final index then use ..

>>> cofoldsWithKey (rxfrom Map.mapWithKey # rxfrom Map.mapWithKey) (\k r a -> Map.singleton k (a + r)) 1.0 $ Map.fromList [("k",Map.fromList [("l",2.0)])]
fromList [("k",fromList [("l",fromList [("kl",3.0)])])]

Since: 0.0.3

(#~) :: Monoid i => Cxoptic (->) i s t a b -> (i -> b) -> s -> t infixr 4

Set the focus of a coindexed optic.

Since: 0.0.3

(##~) :: Monoid i => Cxoptic (->) i s t a b -> (i -> a -> b) -> s -> t infixr 4

Map over a coindexed optic.

Since: 0.0.3

reoverWithKey :: Monoid i => Cxoptic (->) i s t a b -> (i -> a -> b) -> s -> t

TODO: Document

Since: 0.0.3

(*~) :: Optic (Star f) s t a b -> f b -> s -> f t infixr 4

Set the focus of a representable optic.

Since: 0.0.3

(**~) :: Optic (Star f) s t a b -> (a -> f b) -> s -> f t infixr 4

Map over a representable optic.

>>> [66,97,116,109,97,110] & traversed **~ \a -> ("na", chr a)
("nananananana","Batman")

Since: 0.0.3

reps :: Representable p => Optic p s t a b -> (a -> Rep p b) -> s -> Rep p t

TODO: Document

repsWithKey :: Representable p => Monoid i => Ixoptic p i s t a b -> (i -> a -> Rep p b) -> s -> Rep p t

TODO: Document

Since: 0.0.3

(/~) :: Optic (Costar f) s t a b -> b -> f s -> t infixr 4

Set the focus of a co-representable optic.

Since: 0.0.3

(//~) :: Optic (Costar f) s t a b -> (f a -> b) -> f s -> t infixr 4

Map over a co-representable optic.

Since: 0.0.3

coreps :: Corepresentable p => Optic p s t a b -> (Corep p a -> b) -> Corep p s -> t

TODO: Document

corepsWithKey :: Corepresentable p => Monoid i => Cxoptic p i s t a b -> (i -> Corep p a -> b) -> Corep p s -> t

TODO: Document

Since: 0.0.3

Arrow-style combinators

(<<*>>) :: Traversing1 p => p a (b -> c) -> p a b -> p a c infixl 4

Profunctor variant of <*>.

(****) :: Traversing1 p => p a1 b1 -> p a2 b2 -> p (a1, a2) (b1, b2) infixr 3

Profunctor variant of ***.

(++++) :: Cotraversing1 p => p a1 b1 -> p a2 b2 -> p (a1 + a2) (b1 + b2) infixr 2

Profunctor variant of +++.

(&&&&) :: Traversing1 p => p a b1 -> p a b2 -> p a (b1, b2) infixr 3

Profunctor variant of &&&.

(||||) :: Cotraversing1 p => p a1 b -> p a2 b -> p (a1 + a2) b infixr 2

Profunctor variant of |||.

liftR2 :: Traversing1 p => (b -> c -> d) -> p a b -> p a c -> p a d

Divisible-style combinators

divide :: Traversing1 p => (a -> (a1, a2)) -> p a1 b -> p a2 b -> p a b

Profunctor variant of divide.

divide' :: Traversing1 p => p a1 b -> p a2 b -> p (a1, a2) b

codivide :: Cotraversing1 p => ((b1 + b2) -> b) -> p a b1 -> p a b2 -> p a b

codivide' :: Cotraversing1 p => p a b1 -> p a b2 -> p a (b1 + b2)

choose :: Cotraversing1 p => (a -> a1 + a2) -> p a1 b -> p a2 b -> p a b

Profunctor variant of choose.

choose' :: Cotraversing1 p => p a1 b -> p a2 b -> p (a1 + a2) b

cochoose :: Traversing1 p => ((b1, b2) -> b) -> p a b1 -> p a b2 -> p a b

cochoose' :: Traversing1 p => p a b1 -> p a b2 -> p a (b1, b2)

pappend :: Traversing1 p => p a b -> p a b -> p a b