Data.Profunctor.Optic.Fold

Fold0

type Fold0 s a = forall p. (Affine p, CoercingR p) => Optic' p s a

fold0 :: (s -> Maybe a) -> Fold0 s a

Obtain a Fold0 directly.

fold0 . preview ≡ id
fold0 (view o) ≡ o . just

>>> preview (fold0 listToMaybe) "foo"
Just 'f'

ixfold0 :: (s -> Maybe (k, a)) -> Ixfold0 k s a

Obtain an Ixfold0 directly.

Since: 0.0.3

afold0 :: ((a -> Maybe r) -> s -> Maybe r) -> AFold0 r s a

TODO: Document

failing :: AFold0 a s a -> AFold0 a s a -> Fold0 s a infix 3

If the first Fold0 has no focus then try the second one.

toFold0 :: View s (Maybe a) -> Fold0 s a

Obtain a Fold0 from a View.

toFold0 o ≡ o . just
toFold0 o ≡ fold0 (view o)

fromFold0 :: AFold0 a s a -> View s (Maybe a)

Obtain a View from a Fold0

'fromFold0' = 'to' . 'preview'

Fold

type Fold s a = forall p. (Affine p, Traversing p, CoercingR p) => Optic' p s a

type Cofold t b = forall p. (Affine p, Cotraversing p, CoercingL p) => Optic' p t b

fold_ :: Foldable f => (s -> f a) -> Fold s a

Obtain a Fold directly.

fold_ (lists o) ≡ o
fold_ f ≡ to f . foldVl traverse_
fold_ f ≡ coercer . lmap f . lift traverse_

See Property.

This can be useful to lift operations from Data.List and elsewhere into a Fold.

>>> [1,2,3,4] ^.. fold_ tail
[2,3,4]

afold :: ((a -> r) -> s -> r) -> ATraversal (Const r) s t a b

TODO: Document

afold :: ((a -> r) -> s -> r) -> AFold r s a

aixfold :: ((k -> a -> r) -> s -> r) -> AIxtraversal (Const r) k s t a b

TODO: Document

Since: 0.0.3

folding :: Traversable f => (s -> a) -> Fold (f s) a

Obtain a Fold from a Traversable functor.

folding f ≡ traversed . to f
folding f ≡ foldVl traverse . to f

foldVl :: (forall f. Applicative f => (a -> f b) -> s -> f t) -> Fold s a

Obtain a Fold from a Van Laarhoven Fold.

ixfoldVl :: (forall f. Applicative f => (k -> a -> f b) -> s -> f t) -> Ixfold k s a

Obtain a Ixfold from a Van Laarhoven Fold.

Since: 0.0.3

acofold :: ((r -> b) -> r -> t) -> ACofold r t b

TODO: Document

cofolding :: Distributive g => (b -> t) -> Cofold (g t) b

TODO: Document

'cofoldVl' 'cotraverse' . 'from' f

cofoldVl :: (forall f. Coapplicative f => (f a -> b) -> f s -> t) -> Cofold t b

Obtain a Cofold from a Van Laarhoven Cofold.

Fold1

type Fold1 s a = forall p. (Strong p, Traversing1 p, CoercingR p) => Optic' p s a

type Cofold1 t b = forall p. (Choice p, Cotraversing1 p, CoercingL p) => Optic' p t b

fold1_ :: Foldable1 f => (s -> f a) -> Fold1 s a

Obtain a Fold1 directly.

fold1_ (toNonEmpty o) ≡ o
fold1_ f ≡ to f . foldVl1 traverse1_
fold1_ f ≡ coercer . lmap f . lift traverse1_

See Property.

This can be useful to represent operations from Data.List.NonEmpty and elsewhere into a Fold1.

folding1 :: Traversable1 f => (s -> a) -> Fold1 (f s) a

Obtain a Fold1 from a Traversable1 functor.

folding1 f ≡ traversed1 . to f
folding1 f ≡ foldVl1 traverse1 . to f

foldVl1 :: (forall f. Apply f => (a -> f b) -> s -> f t) -> Fold1 s a

Obtain a Fold1 from a Van Laarhoven Fold1.

See Property.

ixfoldVl1 :: (forall f. Apply f => (k -> a -> f b) -> s -> f t) -> Ixfold1 k s a

Obtain a Ixfold from a Van Laarhoven Fold.

Since: 0.0.3

cofolding1 :: Distributive1 g => (b -> t) -> Cofold1 (g t) b

TODO: Document

'cofolding1' f = 'cofoldVl1' 'cotraverse1' . 'from' f

cofoldVl1 :: (forall f. Coapply f => (f a -> b) -> f s -> t) -> Cofold1 t b

Obtain a Cofold1 from a Van Laarhoven Cofold1.

Optics

folded0 :: Fold0 (Maybe a) a

The canonical Fold0.

>>> [Just 1, Nothing] ^.. folded . folded0
[1]

filtered :: (a -> Bool) -> Fold0 a a

Filter another optic.

>>> [1..10] ^.. folded . filtered even
[2,4,6,8,10]

folded :: Traversable f => Fold (f a) a

Obtain a Fold from a Traversable functor.

cofolded :: Distributive g => Cofold (g b) b

TODO: Document

'cofoldVl' 'cotraverse' . 'from' f

folded_ :: Foldable f => Fold (f a) a

The canonical Fold.

foldMap ≡ withForget folded_'

ixfoldedRep :: Representable f => Traversable f => Ixfold (Rep f) (f a) a

Obtain an Ixfold from a Representable functor.

Since: 0.0.3

folded1 :: Traversable1 f => Fold1 (f a) a

Obtain a Fold1 from a Traversable1 functor.

folded1_ :: Foldable1 f => Fold1 (f a) a

The canonical Fold1.

foldMap1 ≡ withForget folded1_'

afolded :: Foldable f => Monoid r => AFold r (f a) a

TODO: Document

afolded1 :: Foldable1 f => Semigroup r => AFold r (f a) a

TODO: Document

acolist :: ACofold a [b] (Maybe (b, a))

Right list unfold over an optic.

acolist1 :: ACofold a (NonEmpty b) (b, Maybe a)

Right non-empty list unfold over an optic.

Operators

folds0 :: AFold0 r s a -> (a -> Maybe r) -> s -> Maybe r

TODO: Document

(^?) :: s -> AFold0 a s a -> Maybe a infix 8

An infk alias for preview'.

(^?) ≡ flip preview'

Perform a safe head of a Fold or Traversal or retrieve Just the result from a View or Lens.

When using a Traversal as a partial Lens, or a Fold as a partial View this can be a convenient way to extract the optional value.

>>> Left 4 ^? left'
Just 4
>>> Right 4 ^? left'
Nothing

preview :: MonadReader s m => AFold0 a s a -> m (Maybe a)

TODO: Document

previews :: MonadReader s m => AFold0 r s a -> (a -> r) -> m (Maybe r)

TODO: Document

preuse :: MonadState s m => AFold0 a s a -> m (Maybe a)

TODO: Document

preuses :: MonadState s m => AFold0 r s a -> (a -> r) -> m (Maybe r)

TODO: Document

folds :: ATraversal (Const r) s t a b -> (a -> r) -> s -> r

Map an optic to a monoid and combine the results.

foldMap = folds folded_

>>> folds bitraversed (\x -> [x, x + 1]) (1,3)
[1,2,3,4]
>>> folds bitraversed id (["foo"], ["bar", "baz"])
["foo","bar","baz"]

cofolds :: ACotraversal (Const r) s t a b -> (r -> b) -> r -> t

TODO: Document

>>> cofolds (from succ) (*2) 3
7

Compare reviews.

foldsa :: Applicative f => AFold (f a) s a -> s -> f a

TODO: Document

cofoldsa :: Coapplicative f => ACofold (f b) t b -> f b -> t

TODO: Document

(^..) :: s -> AFold (Endo [a]) s a -> [a] infix 8

Infix alias of lists.

toList xs ≡ xs ^.. folding
(^..) ≡ flip lists

>>> [[1,2], [3 :: Int64]] ^.. id
[[[1,2],[3]]]
>>> [[1,2], [3 :: Int64]] ^.. traversed
[[1,2],[3]]
>>> [[1,2], [3 :: Int64]] ^.. traversed . traversed
[1,2,3]

>>> (1,2) ^.. bitraversed
[1,2]

lists :: AFold (Endo [a]) s a -> s -> [a]

Collect the fock of an optic into a list.

lists folded_ = toList

foldsr :: AFold (Endo r) s a -> (a -> r -> r) -> r -> s -> r

Extract a NonEmpty of the fock of an optic. | Right fold over an optic.

>>> foldsr folded (+) 0 [1..5::Int64]
15

foldsl :: AFold ((Endo - Dual) r) s a -> (r -> a -> r) -> r -> s -> r

Left fold over an optic.

foldsr' :: AFold ((Endo - Dual) (Endo r)) s a -> (a -> r -> r) -> r -> s -> r

Strict right fold over an optic.

foldsl' :: AFold (Endo (Endo r)) s a -> (r -> a -> r) -> r -> s -> r

Strict left fold over an optic.

foldl' ≡ foldsl' folding

foldsrM :: Monad m => AFold ((Endo - Dual) (EndoM m r)) s a -> (a -> r -> m r) -> r -> s -> m r

Monadic right fold over an optic.

>>> foldsrM folded_ (\x y -> Identity (x++y)) "" ["foo","bar","baz"]
Identity "foobarbaz"

foldslM :: Monad m => AFold (Endo (EndoM m r)) s a -> (r -> a -> m r) -> r -> s -> m r

Monadic left fold over an optic.

traverses_ :: Applicative f => AFold (Endo (f ())) s a -> (a -> f r) -> s -> f ()

Applicative fold over an optic.

traverse_ ≡ traverses_ folded

>>> traverses_ bitraversed putStrLn ("hello","world")
hello
world

Indexed operators

folds0WithKey :: Monoid k => AIxfold0 r k s a -> (k -> a -> Maybe r) -> s -> Maybe r

TODO: Document

Since: 0.0.3

previewWithKey :: Monoid k => AIxfold0 (k, a) k s a -> s -> Maybe (k, a)

TODO: Document

Since: 0.0.3

previewsWithKey :: Monoid k => AIxfold0 r k s a -> (k -> a -> r) -> s -> Maybe r

TODO: Document

Since: 0.0.3

foldsWithKey :: Monoid k => AIxfold r k s a -> (k -> a -> r) -> s -> r

Map an indexed optic to a monoid and combine the results.

Since: 0.0.3

cofoldsWithKey :: Monoid k => ACxfold r k t b -> (k -> r -> b) -> r -> t

TODO: Document

Since: 0.0.3

(^%%) :: Monoid k => s -> AIxfold (Endo [(k, a)]) k s a -> [(k, a)] infix 8

Infix version of listsWithKey.

Since: 0.0.3

listsWithKey :: Monoid k => AIxfold (Endo [(k, a)]) k s a -> s -> [(k, a)]

Collect the fock of an indexed optic into a list of index-value pairs.

lists l ≡ map snd . listsWithKey l

>>> listsWithKey (ix (Sum 1) traversed) ["foo","bar"]
[(Sum {getSum = 0},"foo"),(Sum {getSum = 1},"bar")]

Since: 0.0.3

foldsrWithKey :: Monoid k => AIxfold (Endo r) k s a -> (k -> a -> r -> r) -> r -> s -> r

Indexed right fold over an indexed optic.

foldsr o ≡ foldsrWithKey o . const
foldrWithKey f ≡ foldsrWithKey ixfolded f

Since: 0.0.3

foldslWithKey :: Monoid k => AIxfold ((Endo - Dual) r) k s a -> (k -> r -> a -> r) -> r -> s -> r

Left fold over an indexed optic.

foldsl o ≡ foldslWithKey o . const
foldlWithKey f ≡ foldslWithKey ixfolded f

Since: 0.0.3

foldsrWithKey' :: Monoid k => AIxfold ((Endo - Dual) (Endo r)) k s a -> (k -> a -> r -> r) -> r -> s -> r

Strict right fold over an indexed optic.

foldsr' o ≡ foldsrWithKey' o . const
foldrWithKey' f ≡ foldsrWithKey' ixfolded f

Since: 0.0.3

foldslWithKey' :: Monoid k => AIxfold (Endo (Endo r)) k s a -> (k -> r -> a -> r) -> r -> s -> r

Strict left fold over an indexed optic.

foldsl' o ≡ foldslWithKey' o . const
foldlWithKey' f ≡ foldslWithKey' ixfolded f

Since: 0.0.3

foldsrMWithKey :: Monoid k => Monad m => AIxfold ((Endo - Dual) (EndoM m r)) k s a -> (k -> a -> r -> m r) -> r -> s -> m r

Monadic right fold over an indexed optic.

foldsrM ≡ ixfoldrM . const

Since: 0.0.3

foldslMWithKey :: Monoid k => Monad m => AIxfold (Endo (EndoM m r)) k s a -> (k -> r -> a -> m r) -> r -> s -> m r

Monadic left fold over an indexed optic.

foldslM ≡ foldslMWithKey . const

Since: 0.0.3

traversesWithKey_ :: Monoid k => Applicative f => AIxfold (Endo (f ())) k s a -> (k -> a -> f r) -> s -> f ()

Applicative fold over an indexed optic.

traversesWithKey_ ixfolded ≡ traverseWithKey_

Since: 0.0.3

EndoM

newtype EndoM m a

Constructors

EndoM

Fields

appEndoM :: a -> m a

Classes

class Profunctor p => Strong (p :: Type -> Type -> Type) where

Generalizing Star of a strong Functor

Note: Every Functor in Haskell is strong with respect to (,).

This describes profunctor strength with respect to the product structure of Hask.

http://www.riec.tohoku.ac.jp/~asada/papers/arrStrMnd.pdf

Minimal complete definition

first' | second'

Methods

first' :: p a b -> p (a, c) (b, c)

Laws:

first' ≡ dimap swap swap . second'
lmap fst ≡ rmap fst . first'
lmap (second' f) . first' ≡ rmap (second' f) . first'
first' . first' ≡ dimap assoc unassoc . first' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

second' :: p a b -> p (c, a) (c, b)

Laws:

second' ≡ dimap swap swap . first'
lmap snd ≡ rmap snd . second'
lmap (first' f) . second' ≡ rmap (first' f) . second'
second' . second' ≡ dimap unassoc assoc . second' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

class Profunctor p => Choice (p :: Type -> Type -> Type) where

The generalization of Costar of Functor that is strong with respect to Either.

Note: This is also a notion of strength, except with regards to another monoidal structure that we can choose to equip Hask with: the cocartesian coproduct.

Minimal complete definition

left' | right'

Methods

left' :: p a b -> p (Either a c) (Either b c)

Laws:

left' ≡ dimap swapE swapE . right' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Left ≡ lmap Left . left'
lmap (right f) . left' ≡ rmap (right f) . left'
left' . left' ≡ dimap assocE unassocE . left' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

right' :: p a b -> p (Either c a) (Either c b)

Laws:

right' ≡ dimap swapE swapE . left' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Right ≡ lmap Right . right'
lmap (left f) . right' ≡ rmap (left f) . right'
right' . right' ≡ dimap unassocE assocE . right' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

class Profunctor p => Closed (p :: Type -> Type -> Type) where

A strong profunctor allows the monoidal structure to pass through.

A closed profunctor allows the closed structure to pass through.

Methods

closed :: p a b -> p (x -> a) (x -> b)

Laws:

lmap (. f) . closed ≡ rmap (. f) . closed
closed . closed ≡ dimap uncurry curry . closed
dimap const ($()) . closed ≡ id

class (Sieve p (Rep p), Strong p) => Representable (p :: Type -> Type -> Type) where

A Profunctor p is Representable if there exists a Functor f such that p d c is isomorphic to d -> f c.

Associated Types

type Rep (p :: Type -> Type -> Type) :: Type -> Type

Methods

tabulate :: (d -> Rep p c) -> p d c

Laws:

tabulate . sieve ≡ id
sieve . tabulate ≡ id

class (Cosieve p (Corep p), Costrong p) => Corepresentable (p :: Type -> Type -> Type) where

A Profunctor p is Corepresentable if there exists a Functor f such that p d c is isomorphic to f d -> c.

Associated Types

type Corep (p :: Type -> Type -> Type) :: Type -> Type

Methods

cotabulate :: (Corep p d -> c) -> p d c

Laws:

cotabulate . cosieve ≡ id
cosieve . cotabulate ≡ id