Data.Profunctor.Optic.Machine

Types

type Moore s t a b = forall p. (Closed p, Cotraversing1 p, Foldable (Corep p)) => Optic p s t a b

A Moore machine.

type Mealy s t a b = forall p. (Coaffine p, Cotraversing p, Foldable1 (Corep p)) => Optic p s t a b

A Mealy machine.

type Moore' t b = Moore t t b b

type Mealy' t b = Mealy t t b b

type AFoldl s t a b = Optic Foldl s t a b

Reified Moore optic.

type AFoldl1 s t a b = Optic Foldl1 s t a b

Reified Mealy optic.

Moore

moore :: (s -> a) -> (forall f. Foldable' f => f s -> b -> t) -> Moore s t a b

Obtain a Moore directly.

listing :: (s -> a) -> ([s] -> b -> t) -> Moore s t a b

A list lens.

foldingr :: (s -> a) -> (s -> r -> r) -> r -> (r -> b -> t) -> Moore s t a b

Right fold over a Moore.

foldingr' :: (s -> a) -> (s -> r -> r) -> r -> (r -> b -> t) -> Moore s t a b

Strict right fold over a Moore.

foldingl :: (s -> a) -> (r -> s -> r) -> r -> (r -> b -> t) -> Moore s t a b

Left fold over a Moore.

foldingl' :: (s -> a) -> (r -> s -> r) -> r -> (r -> b -> t) -> Moore s t a b

Strict left fold over a Moore.

foldingrM :: Monad m => (s -> a) -> (s -> r -> m r) -> r -> (m r -> b -> t) -> Moore s t a b

Monadic right fold over a Moore.

foldinglM :: Monad m => (s -> a) -> (r -> s -> m r) -> r -> (m r -> b -> t) -> Moore s t a b

Monadic left fold over a Moore.

traversing_ :: Applicative f => (s -> a) -> (s -> f r) -> (f () -> b -> t) -> Moore s t a b

Traverse for effects over a Moore.

mappingM_ :: Monad m => (s -> a) -> (s -> m r) -> (m () -> b -> t) -> Moore s t a b

Map monadically for effects over a Moore.

foldMapping :: Monoid r => (s -> a) -> (s -> r) -> (r -> b -> t) -> Moore s t a b

Fold-map over a Moore.

Mealy

mealy :: (s -> a) -> (forall f. Foldable1' f => f s -> b -> t) -> Mealy s t a b

Obtain a Mealy directly.

listing1 :: (s -> a) -> (NonEmpty s -> b -> t) -> Mealy s t a b

A non-empty list lens.

foldingr1 :: (s -> a) -> (s -> s -> s) -> (s -> b -> t) -> Mealy s t a b

Non-empty right fold over a Mealy.

foldingl1 :: (s -> a) -> (s -> s -> s) -> (s -> b -> t) -> Mealy s t a b

Non-empty left fold over a Mealy.

foldingrM1 :: Monad m => (s -> a) -> (s -> s -> m s) -> (m s -> b -> t) -> Mealy s t a b

Non-empty monadic right fold over a Mealy.

foldinglM1 :: Monad m => (s -> a) -> (s -> s -> m s) -> (m s -> b -> t) -> Mealy s t a b

Non-empty monadic left fold over a Mealy.

traversing1_ :: Apply f => (s -> a) -> (s -> f r) -> (f () -> b -> t) -> Mealy s t a b

Traverse for effects over a Mealy (non-empty).

foldMapping1 :: Semigroup r => (s -> a) -> (s -> r) -> (r -> b -> t) -> Mealy s t a b

Fold-map over a Mealy (non-empty, semigroup).

Optics

head1 :: Lens s t a b -> Mealy s t a b

Retain the first out-of-focus part of a lens.

last1 :: Lens s t a b -> Mealy s t a b

Retain the last out-of-focus part of a lens.

projected :: (s -> a) -> Moore s b a b

Project away a structure.

minimized :: Ord s => Lens s t a b -> Mealy s t a b

Minimize over a lens.

maximized :: Ord s => Lens s t a b -> Mealy s t a b

Maximize over a lens.

minimizedDef :: Ord s => s -> Lens s t a b -> Moore s t a b

Minimize over a Lens using a default.

maximizedDef :: Ord s => s -> Lens s t a b -> Moore s t a b

Maximize over a Lens using a default.

minimizedBy :: (s -> s -> Ordering) -> Lens s t a b -> Mealy s t a b

Minimize over a Lens using a comparator.

maximizedBy :: (s -> s -> Ordering) -> Lens s t a b -> Mealy s t a b

Maximize over a Lens using a comparator.

minimizedByDef :: (s -> s -> Ordering) -> s -> Lens s t a b -> Moore s t a b

Minimize over a Lens using a comparator and a default.

maximizedByDef :: (s -> s -> Ordering) -> s -> Lens s t a b -> Moore s t a b

Maximize over a Lens using a comparator and a default.

foundDef :: (s -> b -> Bool) -> s -> Lens s t a b -> Moore s t a b

Search over a Lens using a predicate and a default.

Operators

listl :: Foldable f => AFoldl s t a [a] -> f s -> t

Fold a structure into a list through an optic.

listl o = buildl o list

listl1 :: Foldable1 f => AFoldl1 s t a (NonEmpty a) -> f s -> t

Fold a non-empty structure into a non-empty list through an optic.

steps :: Foldable1 f => AFoldl1 s t a b -> (x -> a -> x) -> x -> (x -> b) -> f s -> t

Fold a non-empty structure through an optic using a step function.

buildl :: Foldable f => AFoldl s t a b -> Foldl a b -> f s -> t

Fold a structure through an optic using a Foldl.

buildsl :: Foldable f => AFoldl s t a b -> (x -> a -> x) -> x -> (x -> b) -> f s -> t

Fold a structure through an optic using explicit stepbegindone.

buildl1 :: Foldable1 f => AFoldl1 s t a b -> Foldl1 a b -> f s -> t

Fold a non-empty structure through an optic using a Foldl1.

buildsl1 :: Foldable1 f => AFoldl1 s t a b -> (x -> a -> x) -> (a -> x) -> (x -> b) -> f s -> t

Fold a non-empty structure through an optic using explicit stepbegindone.

postscanl :: Traversable f => AFoldl s t a b -> Foldl a b -> f s -> f t

Post-scan a traversable structure through an optic.

postscansl :: Traversable f => AFoldl s t a b -> (x -> a -> x) -> x -> (x -> b) -> f s -> f t

Post-scan a traversable structure through an optic using explicit stepbegindone.

mconcats :: Foldable f => Monoid m => AFoldl s t a b -> (a -> m) -> (m -> b) -> f s -> t

Monoidal concatenation through an optic.

sconcats :: Foldable1 f => Semigroup m => AFoldl1 s t a b -> (a -> m) -> (m -> b) -> f s -> t

Semigroup concatenation through an optic.

minimizes :: Foldable1 f => Ord a => AFoldl1 s t a a -> f s -> t

Minimize through a non-empty optic.

maximizes :: Foldable1 f => Ord a => AFoldl1 s t a a -> f s -> t

Maximize through a non-empty optic.

minimizesDef :: Foldable f => Ord a => AFoldl s t a a -> a -> f s -> t

Minimize with a default through an optic.

maximizesDef :: Foldable f => Ord a => AFoldl s t a a -> a -> f s -> t

Maximize with a default through an optic.

minimizesBy :: Foldable1 f => AFoldl1 s t a a -> (a -> a -> Ordering) -> f s -> t

Minimize using a comparator through a non-empty optic.

maximizesBy :: Foldable1 f => AFoldl1 s t a a -> (a -> a -> Ordering) -> f s -> t

Maximize using a comparator through a non-empty optic.

minimizesByDef :: Foldable f => AFoldl s t a a -> (a -> a -> Ordering) -> a -> f s -> t

Minimize using a comparator and a default through an optic.

maximizesByDef :: Foldable f => AFoldl s t a a -> (a -> a -> Ordering) -> a -> f s -> t

Maximize using a comparator and a default through an optic.