Data.Profunctor.Optic.Newtype

Newtype isos

wrapped :: Newtype s => Iso' s (O s)

Obtain an Iso from a newtype.

view wrapped f . f ≡ id
f . view wrapped f ≡ id

>>> view wrapped $ Identity 'x'
'x'

>>> view wrapped (Const "hello")
"hello"

rewrapped :: Newtype s => Newtype t => Iso s t (O s) (O t)

Work between newtype wrappers.

>>> Const "hello" & rewrapped ..~ length & getConst
5

rewrapped' :: Newtype s => Newtype t => (O s -> s) -> Iso s t (O s) (O t)

Variant of rewrapped that ignores its argument.

Ala

ala :: Newtype s => Newtype t => Functor f => (O s -> s) -> ((O t -> t) -> f s) -> f (O s)

Based on ala from Conor McBride's work on Epigram.

>>> ala Sum foldMap [1,2,3,4]
10

>>> ala Any foldMap [True,False]
True

>>> ala Product foldMap [1,2,3,4]
24

ala :: Newtype s => Newtype t => (O s -> s) -> ((O t -> t) -> e -> s) -> e -> O s

Re-exports

class Newtype n where

As long as the type n is an instance of Generic, you can create an instance with just instance Newtype n

Minimal complete definition

Nothing

Associated Types

type O n

type O n = GO (Rep n)

Methods

pack :: O n -> n

unpack :: n -> O n

type family O n