Data.Profunctor.Optic.Prism

Prism

type Prism s t a b = forall p. Choice p => Optic p s t a b

\( \mathsf{Prism}\;S\;A = \exists D, S \cong D + A \)

type Prism' s a = Prism s s a a

prism :: (s -> t + a) -> (b -> t) -> Prism s t a b

Obtain a Prism from a constructor and a matcher function.

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions satisfy the following properties:

sta (bt b) ≡ Right b

(id ||| bt) (sta s) ≡ s

left sta (sta s) ≡ left Left (sta s)

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

prism' :: (s -> Maybe a) -> (b -> s) -> Prism s s a b

Obtain a Prism' from a reviewer and a matcher function that produces a Maybe.

Note: The arguments are reversed from the equivalent in the lens package. This is unfortunate but done to maintain cosistency with traversal0 etc.

handling :: (s -> c + a) -> ((c + b) -> t) -> Prism s t a b

Obtain a Prism from its free tensor representation.

Useful for constructing prisms from try and handle functions.

clonePrism :: APrism s t a b -> Prism s t a b

TODO: Document

Optics

left :: Prism (a + c) (b + c) a b

Focus on the Left constructor of Either.

right :: Prism (c + a) (c + b) a b

Focus on the Right constructor of Either.

just :: Prism (Maybe a) (Maybe b) a b

Focus on the Just constructor of Maybe.

>>> Just 1 :| [Just 2, Just 3] & cotraverses just sum
Just 6
>>> Nothing :| [Just 2, Just 3] & cotraverses just sum
Nothing

nothing :: Prism (Maybe a) (Maybe b) () ()

Focus on the Nothing constructor of Maybe.

prefixed :: Eq a => [a] -> Prism' [a] [a]

Focus on the remainder of a list with a given prefix.

only :: Eq a => a -> Prism' a ()

Focus not just on a case, but a specific value of that case.

nearly :: a -> (a -> Bool) -> Prism' a ()

Create a Prism from a value and a predicate.

>>> review (nearly [] null) ()
[]
>>> [1,2,3,4] ^? nearly [] null
Nothing

nearly [] null :: Prism' [a] ()

Caution: In order for the generated optic to be well-defined, you must ensure that f x holds iff x ≡ a.

nthbit :: Bits s => Int -> Prism' s ()

Focus on the truth value of the nth bit in a bit array.

Operators

aside :: APrism s t a b -> Prism (e, s) (e, t) (e, a) (e, b)

Use a Prism to lift part of a structure.

without :: APrism s t a b -> APrism u v c d -> Prism (s + u) (t + v) (a + c) (b + d)

Given a pair of prisms, project sums.

below :: Traversable f => APrism' s a -> Prism' (f s) (f a)

Lift a Prism through a Traversable functor.

Returns a Prism that matches only if each element matches the original Prism.

>>> [Left 1, Right "foo", Left 4, Right "woot"] ^.. below right'
[]
>>> [Right "hail hydra!", Right "foo", Right "blah", Right "woot"] ^.. below right'
[["hail hydra!","foo","blah","woot"]]

toPastroSum :: APrism s t a b -> p a b -> PastroSum p s t

Use a Prism to construct a PastroSum.

toTambaraSum :: Choice p => APrism s t a b -> p a b -> TambaraSum p s t

Use a Prism to construct a TambaraSum.

withPrism :: APrism s t a b -> ((s -> t + a) -> (b -> t) -> r) -> r

Extract the two functions that characterize a Prism.

Classes

class Profunctor p => Choice (p :: Type -> Type -> Type) where

The generalization of Costar of Functor that is strong with respect to Either.

Note: This is also a notion of strength, except with regards to another monoidal structure that we can choose to equip Hask with: the cocartesian coproduct.

Minimal complete definition

left' | right'

Methods

left' :: p a b -> p (Either a c) (Either b c)

Laws:

left' ≡ dimap swapE swapE . right' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Left ≡ lmap Left . left'
lmap (right f) . left' ≡ rmap (right f) . left'
left' . left' ≡ dimap assocE unassocE . left' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

right' :: p a b -> p (Either c a) (Either c b)

Laws:

right' ≡ dimap swapE swapE . left' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Right ≡ lmap Right . right'
lmap (left f) . right' ≡ rmap (left f) . right'
right' . right' ≡ dimap unassocE assocE . right' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c