Data.Profunctor.Optic.Sequence

Isos

strict :: LazySequence l s => Iso' l s

Iso between a lazy sequence and its strict counterpart.

chunked :: LazySequence l s => Iso' l [s]

Iso between a lazy sequence and a list of strict chunks.

unpacked :: IsSequence s => Iso' s [Element s]

Iso between a sequence and a list of its elements.

reversed :: IsSequence s => Iso' s s

Iso that reverses a sequence.

Packing

packing :: IsSequence s => (Element s -> a) -> (s -> b -> t) -> Moore (Element s) t a b

Pack elements into an IsSequence through a Moore.

packing1 :: IsSequence s => (Element s -> a) -> (NonNull s -> b -> t) -> Mealy (Element s) t a b

Pack elements into a NonNull sequence through a Mealy.

chunking :: LazySequence l s => (s -> a) -> (l -> b -> t) -> Moore s t a b

Chunk strict sequences into a lazy sequence through a Moore.

Sequential optics

packed :: IsSequence s => Lens s t a b -> Moore (Element s) t a b

Pack elements of a lens into a sequence.

packed' :: LazySequence l s => Lens l t a b -> Moore s t a b

Pack chunks of a lens into a lazy sequence.

taken :: IsSequence s => (b -> Index s) -> Moore (Element s) s (Element s) b

Take a number of elements from a sequence.

taken' :: LazySequence l s => (b -> Index l) -> Moore s l (Index s) b

Take a number of chunks from a lazy sequence.

padded :: IsSequence s => Element s -> Moore s [s] (Index s) (Index s)

Pad sequences to a uniform length.

padded' :: LazySequence l s => Element s -> Moore s l (Index s) (Index s)

Pad chunks to a uniform length in a lazy sequence.

takenWhile :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) s a b

Take elements while a predicate holds.

takenWhile' :: LazySequence l s => (s -> a) -> (Element l -> b -> Bool) -> Moore s l a b

Take chunks while a predicate holds on a lazy sequence.

droppedWhile :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) s a b

Drop elements while a predicate holds.

droppedWhile' :: LazySequence l s => (s -> a) -> (Element l -> b -> Bool) -> Moore s l a b

Drop chunks while a predicate holds on a lazy sequence.

filteredBy :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) s a b

Filter elements of a sequence.

filteredBy' :: LazySequence l s => (s -> a) -> (Element l -> b -> Bool) -> Moore s l a b

Filter chunks of a lazy sequence.

filteredBy1 :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Mealy (Element s) s a b

Filter elements of a non-empty sequence.

broken :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) (s, s) a b

Break a sequence at the first element matching a predicate.

spanned :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) (s, s) a b

Span a sequence while a predicate holds.

partitioned :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) (s, s) a b

Partition a sequence by a predicate.

partitioned' :: LazySequence l s => (s -> a) -> (Element l -> b -> Bool) -> Moore s (l, l) a b

Partition chunks of a lazy sequence.

groupedAllOn :: IsSequence s => Eq r => (Element s -> a) -> (Element s -> b -> r) -> Moore (Element s) [s] a b

Group elements by equality on a projection.

splitWhen :: IsSequence s => (Element s -> a) -> (Element s -> b -> Bool) -> Moore (Element s) [s] a b

Split a sequence at elements matching a predicate.

splitFirst :: IsSequence s => (Element s -> a) -> ((Element s, s) -> b -> t) -> Mealy (Element s) t a b

Split a non-empty sequence at the first element.

Operators

obuildl :: MonoFoldable s => AFoldl (Element s) t a b -> Foldl a b -> s -> t

Fold a MonoFoldable through an optic using a Foldl.

obuildsl :: MonoFoldable s => AFoldl (Element s) t a b -> (x -> a -> x) -> x -> (x -> b) -> s -> t

Fold a MonoFoldable through an optic using explicit stepbegindone.

obuildl1 :: IsSequence s => AFoldl1 (Element s) t a b -> Foldl1 a b -> NonNull s -> t

Fold a non-empty IsSequence through an optic using a Foldl1.

obuildsl1 :: IsSequence s => AFoldl1 (Element s) t a b -> (x -> a -> x) -> (a -> x) -> (x -> b) -> NonNull s -> t

Fold a non-empty IsSequence through an optic using explicit stepbegindone.

Re-exports

module Data.Profunctor.Optic.Machine