Data.Profunctor.Optic.Types

Optic

type Optic p s t a b = p a b -> p s t

type Optic' p s a = Optic p s s a a

type Ix p k a b = p (k, a) b

type Ix' p a b = Ix p a b b

type Cx p k a b = p a (k -> b)

type Cx' p a b = Cx p a a b

type Ixoptic p k s t a b = Ix p k a b -> Ix p k s t

type Ixoptic' p k s a = Ixoptic p k s s a a

type Cxoptic p k s t a b = Cx p k a b -> Cx p k s t

type Cxoptic' p k t b = Cxoptic p k t t b b

Constraints

type Affine p = (Strong p, Choice p)

type Coaffine p = (Closed p, Choice p)

type Traversing p = (Representable p, Applicative' (Rep p))

type Cotraversing p = (Closed p, Corepresentable p, Coapplicative (Corep p))

type Traversing1 p = (Representable p, Apply (Rep p))

type Cotraversing1 p = (Closed p, Corepresentable p, Coapply (Corep p))

type Mapping p = (Representable p, Distributive (Rep p))

type Remapping p = (Corepresentable p, Traversable (Corep p))

type Mapping1 p = (Representable p, Distributive1 (Rep p))

type Remapping1 p = (Corepresentable p, Traversable1 (Corep p))

type CoercingL p = Bifunctor p

type CoercingR p = forall x. Contravariant (p x)

type Foldable' f = (Functor f, Foldable f)

type Foldable1' f = (Functor f, Foldable1 f)

Equality

type Equality s t a b = forall p. Optic p s t a b

\( \mathsf{Equality}\;A = A \cong A \)

type Equality' s a = Equality s s a a

Iso

type Iso s t a b = forall p. Profunctor p => Optic p s t a b

\( \mathsf{Iso}\;S\;A = S \cong A \)

type Iso' s a = Iso s s a a

Prism

type Prism s t a b = forall p. Choice p => Optic p s t a b

\( \mathsf{Prism}\;S\;A = \exists D, S \cong D + A \)

type Prism' s a = Prism s s a a

Lens

type Lens s t a b = forall p. Strong p => Optic p s t a b

\( \mathsf{Lens}\;S\;A = \exists C, S \cong C \times A \)

type Colens s t a b = forall p. Closed p => Optic p s t a b

\( \mathsf{Colens}\;S\;A = \exists I, S \cong I \to A \)

type Ixlens k s t a b = forall p. Strong p => Ixoptic p k s t a b

type Cxlens k s t a b = forall p. Closed p => Cxoptic p k s t a b

type Lens' s a = Lens s s a a

type Colens' s a = Colens s s a a

type Ixlens' k s a = Ixlens k s s a a

type Cxlens' k t b = Cxlens k t t b b

Traversal

type Traversal0 s t a b = forall p. Affine p => Optic p s t a b

\( \mathsf{Traversal0}\;S\;A = \exists C, D, S \cong D + C \times A \)

type Cotraversal0 s t a b = forall p. Coaffine p => Optic p s t a b

\( \mathsf{Cotraversal0}\;S\;A = \exists D, I, S \cong I \to D + A \)

type Traversal s t a b = forall p. (Affine p, Traversing p) => Optic p s t a b

\( \mathsf{Traversal}\;S\;A = \exists F : \mathsf{Traversable}, S \equiv F\,A \)

type Cotraversal s t a b = forall p. (Coaffine p, Cotraversing p) => Optic p s t a b

\( \mathsf{Cotraversal}\;S\;A = \exists F : \mathsf{Distributive}, S \equiv F\,A \)

type Traversal1 s t a b = forall p. (Strong p, Traversing1 p) => Optic p s t a b

\( \mathsf{Traversal1}\;S\;A = \exists F : \mathsf{Traversable1}, S \equiv F\,A \)

type Cotraversal1 s t a b = forall p. (Closed p, Cotraversing1 p) => Optic p s t a b

\( \mathsf{Cotraversal1}\;S\;A = \exists F : \mathsf{Distributive1}, S \equiv F\,A \)

type Ixtraversal0 k s t a b = forall p. Affine p => Ixoptic p k s t a b

type Cxtraversal0 k s t a b = forall p. Coaffine p => Cxoptic p k s t a b

type Ixtraversal k s t a b = forall p. (Affine p, Traversing p) => Ixoptic p k s t a b

type Cxtraversal k s t a b = forall p. (Coaffine p, Cotraversing p) => Cxoptic p k s t a b

type Ixtraversal1 k s t a b = forall p. (Strong p, Traversing1 p) => Ixoptic p k s t a b

type Cxtraversal1 k s t a b = forall p. (Closed p, Cotraversing1 p) => Cxoptic p k s t a b

type Traversal0' s a = Traversal0 s s a a

type Cotraversal0' t b = Cotraversal0 t t b b

type Traversal' s a = Traversal s s a a

type Cotraversal' t b = Cotraversal t t b b

type Traversal1' s a = Traversal1 s s a a

type Cotraversal1' t b = Cotraversal1 t t b b

type Ixtraversal0' k s a = Ixtraversal0 k s s a a

type Cxtraversal0' k t b = Cxtraversal0 k t t b b

type Ixtraversal' k s a = Ixtraversal k s s a a

type Cxtraversal' k t b = Cxtraversal k t t b b

type Ixtraversal1' k s a = Ixtraversal1 k s s a a

type Cxtraversal1' k t b = Cxtraversal1 k t t b b

Fold

type Fold0 s a = forall p. (Affine p, CoercingR p) => Optic' p s a

type Cofold0 t b = forall p. (Coaffine p, CoercingL p) => Optic' p t b

type Fold s a = forall p. (Affine p, Traversing p, CoercingR p) => Optic' p s a

type Cofold t b = forall p. (Affine p, Cotraversing p, CoercingL p) => Optic' p t b

type Fold1 s a = forall p. (Strong p, Traversing1 p, CoercingR p) => Optic' p s a

type Cofold1 t b = forall p. (Choice p, Cotraversing1 p, CoercingL p) => Optic' p t b

type Ixfold0 k s a = forall p. (Affine p, CoercingR p) => Ixoptic' p k s a

type Cxfold0 k t b = forall p. (Coaffine p, CoercingL p) => Cxoptic' p k t b

type Ixfold k s a = forall p. (Affine p, Traversing p, CoercingR p) => Ixoptic' p k s a

type Cxfold k t b = forall p. (Affine p, Cotraversing p, CoercingL p) => Cxoptic' p k t b

type Ixfold1 k s a = forall p. (Strong p, Traversing1 p, CoercingR p) => Ixoptic' p k s a

type Cxfold1 k t b = forall p. (Choice p, Cotraversing1 p, CoercingL p) => Cxoptic' p k t b

Setter

type Setter s t a b = forall p. (Affine p, Traversing p, Mapping p) => Optic p s t a b

\( \mathsf{Functor}\;S\;A = \exists F : \mathsf{Functor}, S \equiv F\,A \)

type Resetter s t a b = forall p. (Coaffine p, Cotraversing p, Remapping p) => Optic p s t a b

\( \quad \mathsf{Resetter}\;S\;A = \exists n : \mathbb{N}, S \cong \mathsf{Fin}\,n \to A \)

See also section 3 on Kaleidoscopes here.

type Setter1 s t a b = forall p. (Strong p, Traversing1 p, Mapping1 p) => Optic p s t a b

type Resetter1 s t a b = forall p. (Closed p, Cotraversing1 p, Remapping1 p) => Optic p s t a b

type Ixsetter k s t a b = forall p. (Affine p, Traversing p, Mapping p) => Ixoptic p k s t a b

type Rxsetter k s t a b = forall p. (Coaffine p, Cotraversing p, Remapping p) => Cxoptic p k s t a b

type Ixsetter1 k s t a b = forall p. (Strong p, Traversing1 p, Mapping1 p) => Ixoptic p k s t a b

type Rxsetter1 k s t a b = forall p. (Closed p, Cotraversing1 p, Remapping1 p) => Cxoptic p k s t a b

type Setter' s a = Setter s s a a

type Resetter' s a = Resetter s s a a

type Setter1' s a = Setter1 s s a a

type Resetter1' s a = Resetter1 s s a a

type Ixsetter' k s a = Ixsetter k s s a a

type Rxsetter' k t b = Rxsetter k t t b b

type Ixsetter1' k s a = Ixsetter1 k s s a a

type Rxsetter1' k t b = Rxsetter1 k t t b b

View

type View s a = forall p. (Strong p, CoercingR p) => Optic' p s a

type Review t b = forall p. (Closed p, CoercingL p) => Optic' p t b

type Ixview k s a = forall p. (Strong p, CoercingR p) => Ixoptic' p k s a

type Rxview k t b = forall p. (Closed p, CoercingL p) => Cxoptic' p k t b

Re

newtype Re p s t a b

The Re type and its instances witness the symmetry between the parameters of a Profunctor.

Constructors

Re

Fields

runRe :: p b a -> p t s

re :: Optic (Re p a b) s t a b -> Optic p b a t s

Reverse an optic to obtain its dual.

re . re  ≡ id

re :: Iso s t a b   -> Iso b a t s
re :: Lens s t a b  -> Colens b a t s
re :: Prism s t a b -> Coprism b a t s
re :: Traversal s t a b  -> Cotraversal b a t s
re :: View s t a b  -> Review b a t s

>>> 5 ^. re left'
Left 5

between :: (c -> d) -> (a -> b) -> (b -> c) -> a -> d

Can be used to rewrite

\g -> f . g . h

to

between f h

module Data.Profunctor.Types

Orphan instances

Contravariant f => Bifunctor (Costar f)

Instance details

Methods

bimap :: (a -> b) -> (c -> d) -> Costar f a c -> Costar f b d

first :: (a -> b) -> Costar f a c -> Costar f b c

second :: (b -> c) -> Costar f a b -> Costar f a c

Coapplicative f => Choice (Costar f)

Instance details

Methods

left' :: Costar f a b -> Costar f (Either a c) (Either b c)

right' :: Costar f a b -> Costar f (Either c a) (Either c b)

Apply (Costar f a)

Instance details

Methods

(<.>) :: Costar f a (a0 -> b) -> Costar f a a0 -> Costar f a b

(.>) :: Costar f a a0 -> Costar f a b -> Costar f a b

(<.) :: Costar f a a0 -> Costar f a b -> Costar f a a0

liftF2 :: (a0 -> b -> c) -> Costar f a a0 -> Costar f a b -> Costar f a c

Apply f => Apply (Star f a)

Instance details

Methods

(<.>) :: Star f a (a0 -> b) -> Star f a a0 -> Star f a b

(.>) :: Star f a a0 -> Star f a b -> Star f a b

(<.) :: Star f a a0 -> Star f a b -> Star f a a0

liftF2 :: (a0 -> b -> c) -> Star f a a0 -> Star f a b -> Star f a c