Data.Profunctor.Optic.View

View

type View s a = forall p. (Strong p, CoercingR p) => Optic' p s a

type Review t b = forall p. (Closed p, CoercingL p) => Optic' p t b

to :: (s -> a) -> View s a

Obtain a View from an arbitrary function.

to f . to g ≡ to (g . f)
a ^. to f ≡ f a

>>> ("hello","world") ^. to snd
"world"

>>> 5 ^. to succ
6

>>> (0, -5) ^. second' . to abs
5

to :: (s -> a) -> View s a

ixto :: (s -> (k, a)) -> Ixview k s a

TODO: Document

Since: 0.0.3

from :: (b -> t) -> Review t b

Obtain a Review from an arbitrary function.

from ≡ re . to

>>> review (from Prelude.length) [1,2,3]
3

from :: (b -> t) -> Review t b

rxfrom :: ((k -> b) -> t) -> Rxview k t b

TODO: Document

>>> cofoldsWithKey (rxfrom Map.mapWithKey # rxfrom Map.mapWithKey) (\k r a -> Map.singleton k (a + r)) 1.0 $ Map.fromList [("k",Map.fromList [("l",2.0)])]
fromList [("k",fromList [("l",fromList [("kl",3.0)])])]

Since: 0.0.3

cloneView :: AView a s a -> View s a

TODO: Document

cloneView :: Monoid a => AView a s a -> Fold s a

cloneReview :: AReview t b -> Review t b

TODO: Document

Optics

like :: a -> View s a

Obtain a constant-valued (index-preserving) View from an arbitrary value.

This can be useful as a second case failing a Fold e.g. foo failing like 0

like a . like b ≡ like b
a ^. like b ≡ b
a ^. like b ≡ a ^. to (const b)

like :: a -> View s a

ixlike :: k -> a -> Ixview k s a

TODO: Document

Since: 0.0.3

relike :: t -> Review t b

Obtain a constant-valued (index-preserving) Review from an arbitrary value.

relike a . relike b ≡ relike a
relike a .^ b ≡ a
relike a .^ b ≡ from (const a) # b

toProduct :: AView a1 s a1 -> AView a2 s a2 -> View s (a1, a2)

Combine two Views into a View to a product.

toProduct :: View s a1 -> View s a2 -> View s (a1 , a2)

fromSum :: AReview t b1 -> AReview t b2 -> Review t (b1 + b2)

Combine two Reviews into a Review from a sum.

fromSum :: Review t b1 -> Review t b2 -> Review t (b1 + b2)

Operators

(^.) :: s -> AView a s a -> a infix 8

An infix alias for view.

Fiity and semantics are such that subsequent field accesses can be performed with (.).

>>> ("hello","world") ^. second'
"world"

>>> 5 ^. to succ
6

>>> import Data.Complex
>>> ((0, 1 :+ 2), 3) ^. first' . second' . to magnitude
2.23606797749979

view :: MonadReader s m => AView a s a -> m a

View the focus of an optic.

view . to ≡ id

>>> view second' (1, "hello")
"hello"

>>> view (to succ) 5
6

>>> view (second' . first') ("hello",("world","!!!"))
"world"

views :: MonadReader s m => AView r s a -> (a -> r) -> m r

Map each part of a structure viewed to a semantic editor combinator.

'views o f ≡ withForget o f'
foldMap = views folding'

>>> views bitraversed id (["foo"], ["bar", "baz"])
["foo","bar","baz"]

(^%) :: Monoid k => s -> AIxview k s a -> (Maybe k, a) infix 8

View the focus of an indexed optic along with its index.

Note: if the optic focuses on more than one element, then the returned index will be a monoidal sum of all indices visited.

>>> [("foo",41), ("bar",42), ("baz",43)] ^% ix "yo" traversed . ixfirst
(Just "yoyoyo","foobarbaz")

Since: 0.0.3

viewWithKey :: MonadReader s m => Monoid k => AIxview k s a -> m (Maybe k, a)

A prefix alias for ^%.

>>> viewWithKey ixfirst ("foo", 42) :: (Maybe (Sum Int), String)
(Just (Sum {getSum = 0}),"foo")

Since: 0.0.3

viewsWithKey :: MonadReader s m => Monoid k => Ixoptic' (Star (Const r)) k s a -> (k -> a -> r) -> m r

Bring a function of the index and value of an indexed optic into the current environment.

Use viewWithKey if there is a need to disambiguate between mempty as a miss vs. as a return value.

Since: 0.0.3

(.^) :: AReview t b -> b -> t infix 8

An infix alias of review.

>>> from succ .^ 5
6

review :: AReview t b -> b -> t

Review the focus of an optic.

review ≡ view . re
review . from ≡ id

>>> review left' 4
Left 4

reviews :: AReview t b -> (t -> r) -> b -> r

Turn an optic around and look through the other end, applying a function.

reviews ≡ views . re
reviews (from f) g ≡ g . f

>>> reviews left isRight "mustard"
False
>>> reviews (from succ) (*2) 3
8

reviewWithKey :: ARxview k t b -> b -> k -> t

Bring a function of the index of a co-indexed optic into the current environment.

Since: 0.0.3

reviewsWithKey :: ARxview k t b -> ((k -> t) -> r) -> b -> r

Bring a continuation of the index of a co-indexed optic into the current environment.

reviewsWithKey :: ARxview k t b -> ((k -> t) -> r) -> b -> r

Since: 0.0.3

MonadState

use :: MonadState s m => AView a s a -> m a

TODO: Document

uses :: MonadState s m => AFold r s a -> (a -> r) -> m r

Use the target of an optic in the current state.

>>> evalState (uses first length) ("hello","world!")
5

reuse :: MonadState b m => AReview t b -> m t

Turn an optic around and use a value (or the current environment) through it the other way.

reuse ≡ use . re
reuse . from ≡ gets

>>> evalState (reuse left) 5
Left 5
>>> evalState (reuse (from succ)) 5
6

reuses :: MonadState b m => AReview t b -> (t -> r) -> m r

Turn an optic around and use the current state through it the other way, applying a function.

reuses ≡ uses . re
reuses (from f) g ≡ gets (g . f)

>>> evalState (reuses left isLeft) (5 :: Int)
True