Data.Profunctor.Rep.Foldl

Documentation

type L r a b = forall x. (x -> a -> x) -> x -> (x -> b) -> r

Foldl

data Foldl a b

Constructors

forall x. Foldl (x -> a -> x) x (x -> b)

run :: Foldl a b -> a -> b

foldl :: Foldable f => Foldl a b -> f a -> b

withFoldl :: Foldl a b -> L r a b -> r

TODO: Document

prefix :: a -> Foldl a b -> Foldl a b

prescan :: Traversable f => Foldl a b -> f a -> f b

Convert a Foldl into a prescan for any Traversable type

"Prescan" means that the lastDef element of the scan is not included

postscan :: Traversable f => Foldl a b -> f a -> f b

Convert a Foldl into a postscan for any Traversable type

"Postscan" means that the first element of the scan is not included

Folds

list :: Foldl a [a]

Foldl all values into a list

revList :: Foldl a [a]

Foldl all values into a list, in reverse order

mconcat :: Monoid m => (a -> m) -> (m -> b) -> Foldl a b

Convert a "mconcats" to a Fold

stepMay :: (a -> a -> a) -> Foldl a (Maybe a)

Return the result of a step function.

Results in a Nothing value for empty containers.

stepDef :: a -> (a -> a -> a) -> Foldl a a

headDef :: a -> Foldl a a

Return the first element of a collection.

Returns a default if the container is empty.

lastDef :: a -> Foldl a a

Return the last element of a collection.

Returns a default if the container is empty.

maximumDef :: Ord a => a -> Foldl a a

Return the maximumDef element of a collection.

Returns a default if the container is empty.

maximumByDef :: (a -> a -> Ordering) -> a -> Foldl a a

Return the maximumDef element of a collection wrt a comparator.

Returns a default if the container is empty.

minimumDef :: Ord a => a -> Foldl a a

Return the minimumDef element of a collection.

Returns a default if the container is empty.

minimumByDef :: (a -> a -> Ordering) -> a -> Foldl a a

Return the minimumDef element of a collection wrt a comparator.

Returns a default if the container is empty.