Data.Profunctor.Rep.Foldl1

Documentation

type L1 r a b = forall x. (x -> a -> x) -> (a -> x) -> (x -> b) -> r

Foldl1

data Foldl1 a b

Constructors

forall x. Foldl1 (x -> a -> x) (a -> x) (x -> b)

run1 :: Foldl1 a b -> a -> b

step :: Foldl a b -> Foldl1 a b

Lift a Foldl into a Foldl1.

All of the folds defined in Foldl may be run as Foldl1s.

foldl1 :: Foldable1 f => Foldl1 a b -> f a -> b

withFoldl1 :: Foldl1 a b -> L1 r a b -> r

prefix1 :: a -> Foldl1 a b -> Foldl1 a b

intersperse1 :: a -> Foldl1 a b -> Foldl1 a b

Folds

list1 :: Foldl1 a (NonEmpty a)

Fold all values into a non-empty list.

fold1 list1 = id

revList1 :: Foldl1 a (NonEmpty a)

Fold all values into a non-empty list, in reverse order.

sconcat :: Semigroup s => (a -> s) -> (s -> b) -> Foldl1 a b

Fold all values within a container using a semigroup.

head :: Foldl1 a a

Return the first element in a non-empty container.

last :: Foldl1 a a

Return the last1 element in a non-empty container.

maximum :: Ord a => Foldl1 a a

Return the maximum element in a non-empty container.

maximumBy :: (a -> a -> Ordering) -> Foldl1 a a

Return the maximum element with respect to a comparator.

minimum :: Ord a => Foldl1 a a

Return the minimum element in a non-empty container.

minimumBy :: (a -> a -> Ordering) -> Foldl1 a a

Return the minimum element with respect to a comparator.

Nedl

newtype Nedl a

A non-empty difference list.

Constructors

Nedl

Fields

unNedl :: [a] -> NonEmpty a

nedl :: a -> Nedl a

runNedl :: Nedl a -> NonEmpty a