Data.Sequence.Optic

Description

Left and right carriers are characterized by their representation functors (Rep and Corep). Iso sits at the top where both are trivial (Identity). Adjoint sits at the bottom where they form a full adjunction (Corep p ⊣ Rep p). A left-indexed (Ix) optic threads an index through the left adjoint functor (,) k, while a right-indexed (Cx) optic threads a coindex through the right adjoint functor (->) k.

Types

data Seq a

General-purpose finite sequences.

Traversal0

at :: Int -> Traversal0' (Seq a) a

O(log(min(i, n-i))). Affine traversal into the element at an index of a Seq.

ixat :: Int -> Ixtraversal0' Int (Seq a) a

O(log(min(i, n-i))). Indexed affine traversal into the element at an index of a Seq.

Traversal

traversed :: Traversable f => Traversal (f a) (f b) a b

TODO: Document

ixtraversed :: Ixtraversal Int (Seq a) (Seq b) a b

O(n). Indexed traversal over the elements of a Seq.

slicedTo :: Int -> Traversal' (Seq a) a

Traverse the first n elements of a Seq

>>> fromList [1,2,3,4,5] ^.. slicedTo 2
[1,2]

>>> fromList [1,2,3,4,5] & slicedTo 2 %~ (*10)
fromList [10,20,3,4,5]

>>> fromList [1,2,4,5,6] & slicedTo 10 .~ 0
fromList [0,0,0,0,0]

slicedFrom :: Int -> Traversal' (Seq a) a

Traverse all but the first n elements of a Seq

>>> fromList [1,2,3,4,5] ^.. slicedFrom 2
[3,4,5]

>>> fromList [1,2,3,4,5] & slicedFrom 2 %~ (*10)
fromList [1,2,30,40,50]

>>> fromList [1,2,3,4,5] & slicedFrom 10 .~ 0
fromList [1,2,3,4,5]

sliced :: Int -> Int -> Traversal' (Seq a) a

Traverse all the elements numbered from i to j of a Seq

>>> fromList [1,2,3,4,5] & sliced 1 3 %~ (*10)
fromList [1,20,30,4,5]

>>> fromList [1,2,3,4,5] ^.. sliced 1 3
[2,3]

>>> fromList [1,2,3,4,5] & sliced 1 3 .~ 0
fromList [1,0,0,4,5]

Fold

folded :: Traversable f => Fold (f a) a

Obtain a Fold from a Traversable functor.

ixfolded :: Ixfold Int (Seq a) a

O(n). Indexed fold over the elements of a Seq.

Setter

ixmapped :: Ixsetter Int (Seq a) (Seq b) a b

O(n). Indexed setter over the elements of a Seq.

Dual Optics

Colens

zipped :: Int -> Colens (Seq a) (Seq b) (Int -> a) (Int -> b)

Grate viewing a Seq as a function from Int indices. Requires known length to be representable.

Cotraversal

zippedTraverse :: Int -> Cotraversal (Seq a) (Seq b) a b

Pointwise Cotraversal over the elements of a Seq at a fixed length. Extends zipped from Colens to Cotraversal.

Requires known length because Seq is not Distributive (it has variable size).

Cxsetter

cxmapped :: Cxsetter Int (Seq a) (Seq b) a b

O(n). Cxsetter over the elements of a Seq.

Cx dual of ixmapped. Threads the Int index as coindex.

cxsets cxmapped ≡ mapWithIndex

Cxtraversal

cxtraversed :: Cxtraversal Int (Seq a) (Seq b) a b

O(n). Cxtraversal over the elements of a Seq.

Cx dual of ixtraversed. Threads the Int index as coindex.

cxtraverseOf cxtraversed ≡ traverseWithIndex

Cxfold

cxfolded :: Cxfold Int (Seq a) a

O(n). Cxfold over the elements of a Seq.

Cx dual of ixfolded. Threads the Int index as coindex.

Iso

viewedl :: Iso (Seq a) (Seq b) (ViewL a) (ViewL b)

A Seq is isomorphic to a ViewL

viewl m ≡ m ^. viewedl

>>> Seq.fromList [1,2,3] ^. viewedl
1 :< fromList [2,3]

>>> Seq.empty ^. viewedl
EmptyL

>>> EmptyL ^. re viewedl
fromList []

>>> review viewedl $ 1 Seq.:< fromList [2,3]
fromList [1,2,3]

viewedr :: Iso (Seq a) (Seq b) (ViewR a) (ViewR b)

A Seq is isomorphic to a ViewR

viewr m ≡ m ^. viewedr

>>> Seq.fromList [1,2,3] ^. viewedr
fromList [1,2] :> 3

>>> Seq.empty ^. viewedr
EmptyR

>>> EmptyR ^. re viewedr
fromList []

>>> review viewedr $ fromList [1,2] Seq.:> 3
fromList [1,2,3]