Data.Tree.Optic

Lens

root :: Lens' (Tree a) a

A Lens that focuses on the root of a Tree.

>>> view root $ Node 42 []
42

branches :: Lens' (Tree a) [Tree a]

A Lens returning the direct descendants of the root of a Tree

view branches ≡ subForest

Dual Optics

flattened :: Traversal (Tree a) (Tree b) a b

Pre-order traversal of all values in a Tree.

>>> over flattened (+1) (Node 1 [Node 2 [], Node 3 [Node 4 []]])
Node 2 [Node 3 [],Node 4 [Node 5 []]]

ixflattened :: Ixtraversal Int (Tree a) (Tree b) a b

Depth-indexed traversal of all values in a Tree. The index is the depth from the root (root = 0).

Fold

folded :: Fold (Tree a) a

Fold over all values in a Tree (pre-order).

toListOf folded ≡ flatten