Data.List.Optic

Documentation

lists :: AFold (Endo [a]) s a -> s -> [a]

Collect the fock of an optic into a list.

lists folded_ = toList

listsWithKey :: Monoid k => AIxfold (Endo [(k, a)]) k s a -> s -> [(k, a)]

Collect the fock of an indexed optic into a list of index-value pairs.

lists l ≡ map snd . listsWithKey l

>>> listsWithKey (ix (Sum 1) traversed) ["foo","bar"]
[(Sum {getSum = 0},"foo"),(Sum {getSum = 1},"bar")]

Since: profunctor-optics-0.0.3

at :: Int -> Traversal0' [a] a

O(log n). Affine traversal into the value at a key of a list.

iat :: Int -> Ixtraversal0' Int [a] a

O(log n). Indexed affine traversal into the value at a key of a list.

>>> iover (iat 1) (<>) [1,2,3]
[1,3,3]
>>> iover (iat 5) (<>) [1,2,3]
[1,2,3]

imapped :: Ixsetter Int [a] [b] a b

O(n). Ixsetter over the values of a list.

>>> iover imapped (<>) $ [0,3,4]
[0,4,6]

ifiltered :: Ixsetter Int [a] [a] a Bool

O(n). Ixsetter filtering the values of a list.

itraversed :: Ixtraversal Int [a] [b] a b

O(n). Ixtraversal over the values of a list.

ifolded :: Ixfold Int [a] a

O(n). Ixfold over the values of a list.

>>> ilists ifolded [0,3,4]
[(0,0),(1,3),(2,4)]
>>> ilists ifolded []
[]