Data.Profunctor.Optic.Combinator

Constructors

Main

star :: Applicative f => Star f a a

TODO: Document

unstar :: Coapplicative f => Star f a b -> a -> b

TODO: Document

ixmap :: Profunctor p => (s -> a) -> (b -> t) -> Ixoptic p k s t a b

dimap for indexed optics. Maps over the outer types of an Ixoptic without affecting the index.

Since: 0.0.3

representing :: Representable p => ((a -> Rep p b) -> s -> Rep p t) -> Optic p s t a b

TODO: Document

ixrepresenting :: Representable p => ((i -> a -> Rep p b) -> s -> Rep p t) -> Ixoptic p i s t a b

TODO: Document

Since: 0.0.3

Dual

costar :: Coapplicative f => Costar f a a

TODO: Document

uncostar :: Applicative f => Costar f a b -> a -> b

TODO: Document

cxmap :: Profunctor p => (s -> a) -> (b -> t) -> Cxoptic p k s t a b

dimap for coindexed optics. Maps over the outer types of a Cxoptic without affecting the coindex.

Since: 0.0.3

corepresenting :: Corepresentable p => ((Corep p a -> b) -> Corep p s -> t) -> Optic p s t a b

TODO: Document

cxrepresenting :: Corepresentable p => ((i -> Corep p a -> b) -> Corep p s -> t) -> Cxoptic p i s t a b

TODO: Document

Since: 0.0.3

Cx algebra

cxjoin :: Strong p => Cx p a a b -> p a b

Collapse a coindexed profunctor where the coindex matches the focus.

cxjoin ≡ dimap fork apply . first'

cxreturn :: Profunctor p => p a b -> Cx p k a b

Lift a profunctor into a coindexed profunctor by ignoring the coindex.

cxunit :: Strong p => Cx' p a b -> p a b

Extract a profunctor from a self-coindexed profunctor.

cxjoin . cxunit ≡ id

cxstrength :: Profunctor p => Cx' p a b -> Cx' p (a, c) (b, c)

Cx' is freely Strong.

See https://r6research.livejournal.com/27858.html.

Transforms

Main

(%) :: Monoid i => Representable p => Ixoptic p i c1 c2 b1 b2 -> Ixoptic p i b1 b2 a1 a2 -> Ixoptic p i c1 c2 a1 a2 infixr 8

Monoidally combine indices between subsequent levels of optic.

Its precedence is one lower than that of function composition, which allows . to be nested in %.

If you only need the final index then use ..

>>> ixtoListOf (ix "*" traversed . ix "+" traversed) ["foo", "bar"]
[("",'f'),("+",'o'),("++",'o'),("",'b'),("+",'a'),("++",'r')]
>>> ixtoListOf (ix "*" traversed % ix "+" traversed) ["foo", "bar"]
[("",'f'),("+",'o'),("++",'o'),("*",'b'),("*+",'a'),("*++",'r')]

Since: 0.0.3

reix :: Profunctor p => (k1 -> k2) -> (k2 -> k1) -> Ixoptic p k1 s t a b -> Ixoptic p k2 s t a b

Map over the indices of an indexed optic.

See also reixed.

Since: 0.0.3

ixsum :: Profunctor p => Ixoptic p k s t a b -> Ixoptic p (Sum k) s t a b

Lift a numeric index into a sum monoid.

Since: 0.0.3

ixany :: Profunctor p => Ixoptic p Bool s t a b -> Ixoptic p Any s t a b

TODO: Document

ixhead :: Profunctor p => Ixoptic p i s t a b -> Ixoptic p (First i) s t a b

TODO: Document

Since: 0.0.3

ixlast :: Profunctor p => Ixoptic p i s t a b -> Ixoptic p (Last i) s t a b

TODO: Document

Since: 0.0.3

arr :: Traversing p => (a -> b) -> p a b

TODO: Document

(***) :: Traversing1 p => p a1 b1 -> p a2 b2 -> p (a1, a2) (b1, b2) infixr 3

Profunctor variant of ***.

(&&&) :: Traversing1 p => p a b1 -> p a b2 -> p a (b1, b2) infixr 3

Profunctor variant of &&&.

(<<*>>) :: Traversing1 p => p a (b -> c) -> p a b -> p a c infixl 4

Profunctor variant of <*>.

liftR2 :: Traversing1 p => (b -> c -> d) -> p a b -> p a c -> p a d

pappend :: Traversing1 p => p a b -> p a b -> p a b

TODO: Document

divide :: Traversing1 p => p a1 b -> p a2 b -> p (a1, a2) b

TODO: Document

divideWith :: Traversing1 p => (a -> (a1, a2)) -> p a1 b -> p a2 b -> p a b

Profunctor variant of divideWith.

cochoose :: Traversing1 p => p a b1 -> p a b2 -> p a (b1, b2)

TODO: Document

cochooseWith :: Traversing1 p => ((b1, b2) -> b) -> p a b1 -> p a b2 -> p a b

TODO: Document

Dual

(#) :: Monoid i => Corepresentable p => Cxoptic p i c1 c2 b1 b2 -> Cxoptic p i b1 b2 a1 a2 -> Cxoptic p i c1 c2 a1 a2 infixr 8

Compose two coindexed traversals, combining indices.

Its precedence is one lower than that of function composition, which allows . to be nested in #.

If you only need the final index then use ..

>>> cxfoldMapOf (cxfrom Map.mapWithKey # cxfrom Map.mapWithKey) (\k r a -> Map.singleton k (a + r)) 1.0 $ Map.fromList [("k",Map.fromList [("l",2.0)])]
fromList [("k",fromList [("l",fromList [("kl",3.0)])])]

Since: 0.0.3

recx :: Profunctor p => (k1 -> k2) -> (k2 -> k1) -> Cxoptic p k1 s t a b -> Cxoptic p k2 s t a b

Map over the indices of a coindexed optic.

See also recxed.

Since: 0.0.3

cxsum :: Profunctor p => Cxoptic p k s t a b -> Cxoptic p (Sum k) s t a b

Lift a numeric co-index into a sum monoid.

Since: 0.0.3

coarr :: Cotraversing p => (a -> b) -> p a b

TODO: Document

(+++) :: Cotraversing1 p => p a1 b1 -> p a2 b2 -> p (a1 + a2) (b1 + b2) infixr 2

Profunctor variant of +++.

(|||) :: Cotraversing1 p => p a1 b -> p a2 b -> p (a1 + a2) b infixr 2

Profunctor variant of |||.

choose :: Cotraversing1 p => p a1 b -> p a2 b -> p (a1 + a2) b

TODO: Document

chooseWith :: Cotraversing1 p => (a -> a1 + a2) -> p a1 b -> p a2 b -> p a b

Profunctor variant of chooseWith.

codivide :: Cotraversing1 p => p a b1 -> p a b2 -> p a (b1 + b2)

TODO: Document

codivideWith :: Cotraversing1 p => ((b1 + b2) -> b) -> p a b1 -> p a b2 -> p a b

TODO: Document

Optics

constL :: Profunctor p => b -> Optic p a c b c

constR :: Profunctor p => c -> Optic p a c a b

shiftedL :: Profunctor p => Optic p b (c + d) (a + b) c

shiftedR :: Profunctor p => Optic p (a, b) c b (c, d)

coercedL :: Profunctor p => CoercingL p => Optic p c b a b

coercedR :: Profunctor p => CoercingR p => Optic p a c a b

Operators

over :: Optic (->) s t a b -> (a -> b) -> s -> t

Map over an Optic.

over o id ≡ id
over o f . over o g ≡ over o (f . g)
over . setter ≡ id
over . resetter ≡ id

>>> over fmapped (+1) (Just 1)
Just 2
>>> over fmapped (*10) [1,2,3]
[10,20,30]
>>> over first (+1) (1,2)
(2,2)
>>> over first show (10,20)
("10",20)

Benchmark: 1.00x vs direct (Lens), 0.89x vs fmap (Traversal). See Data.Profunctor.Optic.Bench.

Main

ixover :: Monoid i => Ixoptic (->) i s t a b -> (i -> a -> b) -> s -> t

Indexed over: apply a key-dependent function through an indexed optic.

Routes through Conjoin wrapping internally.

Benchmark: 1.08x vs direct mapWithKey (Conjoin overhead negligible). See Data.Profunctor.Optic.Bench.

Since: 0.0.3

reps :: Representable p => Optic p s t a b -> (a -> Rep p b) -> s -> Rep p t

TODO: Document

ixreps :: Representable p => Monoid i => Ixoptic p i s t a b -> (i -> a -> Rep p b) -> s -> Rep p t

TODO: Document

Since: 0.0.3

Dual

cxover :: Monoid i => Cxoptic (->) i s t a b -> (i -> a -> b) -> s -> t

Coindexed over: apply a coindex-dependent function through a coindexed optic.

Routes through Conjoin wrapping internally (dual of overWithKey).

Benchmark: ~1.08x overhead (same Conjoin path as overWithKey). See Data.Profunctor.Optic.Bench.

Since: 0.0.3

coreps :: Corepresentable p => Optic p s t a b -> (Corep p a -> b) -> Corep p s -> t

TODO: Document

cxreps :: Corepresentable p => Monoid i => Cxoptic p i s t a b -> (i -> Corep p a -> b) -> Corep p s -> t

TODO: Document

Since: 0.0.3