Data.Profunctor.Optic.Fold

Constructors

Fold, Ixfold

type Fold s a = forall p. (Affine p, Traversing p, CoercingR p) => Optic' p s a

type Ixfold k s a = forall p. (Affine p, Traversing p, CoercingR p) => Ixoptic' p k s a

fold_ :: Foldable f => (s -> f a) -> Fold s a

Obtain a Fold directly.

fold_ (toListOf o) ≡ o
fold_ f ≡ to f . foldVl traverse_
fold_ f ≡ coercedR . lmap f . lift traverse_

See Property.

This can be useful to lift operations from Data.List and elsewhere into a Fold.

>>> [1,2,3,4] ^.. fold_ (Prelude.drop 1)
[2,3,4]

ixfold_ :: Foldable f => (s -> f (k, a)) -> Ixfold k s a

Obtain an Ixfold directly from an indexed extraction function.

Since: 0.0.3

afold :: ((a -> r) -> s -> r) -> ATraversal (Const r) s t a b

TODO: Document

afold :: ((a -> r) -> s -> r) -> AFold r s a

aixfold :: ((k -> a -> r) -> s -> r) -> AIxtraversal (Const r) k s t a b

TODO: Document

Since: 0.0.3

foldVl :: (forall f. Applicative f => (a -> f b) -> s -> f t) -> Fold s a

Obtain a Fold from a Van Laarhoven Fold.

ixfoldVl :: (forall f. Applicative f => (k -> a -> f b) -> s -> f t) -> Ixfold k s a

Obtain a Ixfold from a Van Laarhoven Fold.

Since: 0.0.3

folding :: Traversable f => (s -> a) -> Fold (f s) a

Obtain a Fold from a Traversable functor.

folding f ≡ traversed . to f
folding f ≡ foldVl traverse . to f

ixfolding :: Traversable f => (s -> (k, a)) -> Ixfold k (f s) a

Obtain an Ixfold from a Traversable functor with an indexed getter.

Since: 0.0.3

Fold0, Ixfold0

type Fold0 s a = forall p. (Affine p, CoercingR p) => Optic' p s a

type Ixfold0 k s a = forall p. (Affine p, CoercingR p) => Ixoptic' p k s a

afold0 :: ((a -> Maybe r) -> s -> Maybe r) -> AFold0 r s a

TODO: Document

aixfold0 :: ((k -> a -> Maybe r) -> s -> Maybe r) -> AIxfold0 r k s a

TODO: Document

Since: 0.0.3

fold0 :: (s -> Maybe a) -> Fold0 s a

Obtain a Fold0 directly.

fold0 . preview ≡ id
fold0 (view o) ≡ o . just

>>> preview (fold0 listToMaybe) "foo"
Just 'f'

fold0' :: View s (Maybe a) -> Fold0 s a

Obtain a Fold0 from a View.

fold0' o ≡ o . just
fold0' o ≡ fold0 (view o)

ixfold0 :: (s -> Maybe (k, a)) -> Ixfold0 k s a

Obtain an Ixfold0 directly.

Since: 0.0.3

failing :: AFold0 a s a -> AFold0 a s a -> Fold0 s a infix 3

If the first Fold0 has no focus then try the second one.

Fold1, Ixfold1

type Fold1 s a = forall p. (Strong p, Traversing1 p, CoercingR p) => Optic' p s a

type Ixfold1 k s a = forall p. (Strong p, Traversing1 p, CoercingR p) => Ixoptic' p k s a

fold1_ :: Foldable1 f => (s -> f a) -> Fold1 s a

Obtain a Fold1 directly.

fold1_ (toNonEmpty o) ≡ o
fold1_ f ≡ to f . foldVl1 traverse1_
fold1_ f ≡ coercedR . lmap f . lift traverse1_

See Property.

This can be useful to representing operations from Data.List.NonEmpty and elsewhere into a Fold1.

ixfold1_ :: Foldable1 f => (s -> f (k, a)) -> Ixfold1 k s a

Obtain an Ixfold1 directly from an indexed extraction function.

Since: 0.0.3

afold1 :: ((a -> r) -> s -> r) -> ATraversal1 (Const r) s t a b

TODO: Document

Since: 0.0.3

aixfold1 :: ((k -> a -> r) -> s -> r) -> AIxtraversal1 (Const r) k s t a b

TODO: Document

Since: 0.0.3

foldVl1 :: (forall f. Apply f => (a -> f b) -> s -> f t) -> Fold1 s a

Obtain a Fold1 from a Van Laarhoven Fold1.

See Property.

ixfoldVl1 :: (forall f. Apply f => (k -> a -> f b) -> s -> f t) -> Ixfold1 k s a

Obtain a Ixfold from a Van Laarhoven Fold.

Since: 0.0.3

folding1 :: Traversable1 f => (s -> a) -> Fold1 (f s) a

Obtain a Fold1 from a Traversable1 functor.

folding1 f ≡ traversed1 . to f
folding1 f ≡ foldVl1 traverse1 . to f

ixfolding1 :: Traversable1 f => (s -> (k, a)) -> Ixfold1 k (f s) a

Obtain an Ixfold1 from a Traversable1 functor with an indexed getter.

Since: 0.0.3

Dual Constructors

Cofold, Cxfold

type Cofold t b = forall p. (Coaffine p, Cotraversing p, CoercingL p) => Optic' p t b

type Cxfold k t b = forall p. (Coaffine p, Cotraversing p, CoercingL p) => Cxoptic' p k t b

acofold :: ((r -> b) -> r -> t) -> ACofold r t b

TODO: Document

cofoldVl :: (forall f. Coapplicative f => (f a -> b) -> f s -> t) -> Cofold t b

Obtain a Cofold from a Van Laarhoven Cofold.

cxfoldVl :: (forall f. Coapplicative f => (f a -> k -> b) -> f s -> t) -> Cxfold k t b

Obtain a Cxfold from a coindexed Van Laarhoven Cofold.

Since: 0.0.3

cofolding :: Distributive g => (b -> t) -> Cofold (g t) b

TODO: Document

'cofoldVl' 'cotraverse' . 'from' f

Cofold1, Cxfold1

type Cofold1 t b = forall p. (Closed p, Cotraversing1 p, CoercingL p) => Optic' p t b

type Cxfold1 k t b = forall p. (Closed p, Cotraversing1 p, CoercingL p) => Cxoptic' p k t b

acofold1 :: ((r -> b) -> r -> t) -> ACofold1 r t b

TODO: Document

Since: 0.0.3

cofoldVl1 :: (forall f. Coapply f => (f a -> b) -> f s -> t) -> Cofold1 t b

Obtain a Cofold1 from a Van Laarhoven Cofold1.

cxfoldVl1 :: (forall f. Coapply f => (f a -> k -> b) -> f s -> t) -> Cxfold1 k t b

Obtain a Cxfold1 from a coindexed Van Laarhoven Cofold1.

Since: 0.0.3

cofolding1 :: Distributive1 g => (b -> t) -> Cofold1 (g t) b

TODO: Document

'cofolding1' f = 'cofoldVl1' 'cotraverse1' . 'from' f

Optics

Fold, Ixfold

afolded :: Foldable f => Monoid r => AFold r (f a) a

TODO: Document

folded :: Traversable f => Fold (f a) a

Obtain a Fold from a Traversable functor.

folded_ :: Foldable f => Fold (f a) a

The canonical Fold.

foldMap ≡ withForget folded_'

ixfoldedRep :: Representable f => Traversable f => Ixfold (Rep f) (f a) a

Obtain an Ixfold from a Representable functor.

Since: 0.0.3

Fold0, Ixfold0

folded0 :: Fold0 (Maybe a) a

The canonical Fold0.

>>> [Just 1, Nothing] ^.. folded . folded0
[1]

filtered :: (a -> Bool) -> Fold0 a a

Filter another optic.

>>> [1..10] ^.. folded . filtered even
[2,4,6,8,10]

Fold1, Ixfold1

afolded1 :: Foldable1 f => Semigroup r => AFold r (f a) a

TODO: Document

folded1 :: Traversable1 f => Fold1 (f a) a

Obtain a Fold1 from a Traversable1 functor.

folded1_ :: Foldable1 f => Fold1 (f a) a

The canonical Fold1.

foldMap1 ≡ withForget folded1_'

Dual Optics

Cofold, Cxfold

acolist :: ACofold a [b] (Maybe (b, a))

Right list unfold over an optic.

cofolded :: Distributive g => Cofold (g b) b

TODO: Document

'cofoldVl' 'cotraverse' . 'from' f

Cofold1, Cxfold1

acolist1 :: ACofold a (NonEmpty b) (b, Maybe a)

Right non-empty list unfold over an optic.

cofolded1 :: Distributive1 g => Cofold1 (g b) b

Obtain a Cofold1 from a Distributive1 functor.

Since: 0.0.3

Operators

Fold, Ixfold

foldMapOf :: ATraversal (Const r) s t a b -> (a -> r) -> s -> r

Map an optic to a monoid and combine the results.

foldMap = foldMapOf folded_

>>> foldMapOf bitraversed (\x -> [x, x + 1]) (1,3)
[1,2,3,4]
>>> foldMapOf bitraversed id (["foo"], ["bar", "baz"])
["foo","bar","baz"]

ixfoldMapOf :: Monoid k => AIxfold r k s a -> (k -> a -> r) -> s -> r

Map an indexed optic to a monoid and combine the results.

Benchmark: indexed fold via Conjoin, ~1.08x overhead. See Data.Profunctor.Optic.Bench.

Since: 0.0.3

foldOfA :: Applicative f => AFold (f a) s a -> s -> f a

TODO: Document

toListOf :: AFold ((Endo - List) a) s a -> s -> [a]

Infix alias of toListOf.

toList xs ≡ xs ^.. folding
(^..) ≡ flip toListOf

>>> [[1,2], [3 :: Int64]] ^.. id
[[[1,2],[3]]]
>>> [[1,2], [3 :: Int64]] ^.. traversed
[[1,2],[3]]
>>> [[1,2], [3 :: Int64]] ^.. traversed . traversed
[1,2,3]

| Collect the foci of an optic into a list.

toListOf folded_ = toList

Benchmark: 1.73x vs Map.elems on Map (Endo closure chain). Same overhead as lens's toListOf — both use the Endo path. Note: ~8x vs foldr (:) [] on toListOf is misleading (foldr (:) [] on [a] is id). See Data.Profunctor.Optic.Bench.

ixtoListOf :: Monoid k => AIxfold ((Endo - List) (k, a)) k s a -> s -> [(k, a)]

Collect the foci of an indexed optic into a list of index-value pairs.

toListOf l ≡ map snd . ixtoListOf l

>>> ixtoListOf (ix (Sum 1) traversed) ["foo","bar"]
[(Sum {getSum = 0},"foo"),(Sum {getSum = 1},"bar")]

Since: 0.0.3

foldrOf :: AFold (Endo r) s a -> (a -> r -> r) -> r -> s -> r

Extract a NonEmpty of the foci of an optic. | Right fold over an optic.

>>> foldrOf folded (+) 0 [1..5::Int64]
15

ixfoldrOf :: Monoid k => AIxfold (Endo r) k s a -> (k -> a -> r -> r) -> r -> s -> r

Indexed right fold over an indexed optic.

foldrOf o ≡ ixfoldrOf o . const
foldrWithKey f ≡ ixfoldrOf ixfolded f

Since: 0.0.3

foldlOf :: AFold ((Dual - Endo) r) s a -> (r -> a -> r) -> r -> s -> r

Left fold over an optic.

ixfoldlOf :: Monoid k => AIxfold ((Dual - Endo) r) k s a -> (k -> r -> a -> r) -> r -> s -> r

Left fold over an indexed optic.

foldlOf o ≡ ixfoldlOf o . const
foldlWithKey f ≡ ixfoldlOf ixfolded f

Since: 0.0.3

foldrOf' :: AFold (((Dual - Endo) - Endo) r) s a -> (a -> r -> r) -> r -> s -> r

Strict right fold over an optic.

ixfoldrOf' :: Monoid k => AIxfold (((Dual - Endo) - Endo) r) k s a -> (k -> a -> r -> r) -> r -> s -> r

Strict right fold over an indexed optic.

foldrOf' o ≡ ixfoldrOf' o . const
foldrWithKey' f ≡ ixfoldrOf' ixfolded f

Since: 0.0.3

foldlOf' :: AFold ((Endo - Endo) r) s a -> (r -> a -> r) -> r -> s -> r

Strict left fold over an optic.

foldl' ≡ foldlOf' folding

ixfoldlOf' :: Monoid k => AIxfold ((Endo - Endo) r) k s a -> (k -> r -> a -> r) -> r -> s -> r

Strict left fold over an indexed optic.

foldlOf' o ≡ ixfoldlOf' o . const
foldlWithKey' f ≡ ixfoldlOf' ixfolded f

Since: 0.0.3

foldrMOf :: Monad m => AFold (((Dual - Endo) - EndoM m) r) s a -> (a -> r -> m r) -> r -> s -> m r

Monadic right fold over an optic.

>>> foldrMOf folded_ (\x y -> Identity (x++y)) "" ["foo","bar","baz"]
Identity "foobarbaz"

ixfoldrMOf :: Monoid k => Monad m => AIxfold (((Dual - Endo) - EndoM m) r) k s a -> (k -> a -> r -> m r) -> r -> s -> m r

Monadic right fold over an indexed optic.

foldrMOf ≡ ixfoldrM . const

Since: 0.0.3

foldlMOf :: Monad m => AFold ((Endo - EndoM m) r) s a -> (r -> a -> m r) -> r -> s -> m r

Monadic left fold over an optic.

ixfoldlMOf :: Monoid k => Monad m => AIxfold ((Endo - EndoM m) r) k s a -> (k -> r -> a -> m r) -> r -> s -> m r

Monadic left fold over an indexed optic.

foldlMOf ≡ ixfoldlMOf . const

Since: 0.0.3

traverseOf_ :: Applicative f => AFold ((Endo - f) ()) s a -> (a -> f r) -> s -> f ()

Applicative fold over an optic.

traverse_ ≡ traverseOf_ folded

>>> traverseOf_ bitraversed putStrLn ("hello","world")
hello
world

ixtraverseOf_ :: Monoid k => Applicative f => AIxfold ((Endo - f) ()) k s a -> (k -> a -> f r) -> s -> f ()

Applicative fold over an indexed optic.

ixtraverseOf_ ixfolded ≡ traverseWithKey_

Since: 0.0.3

has :: AFold0 a s a -> s -> Bool

Check whether an optic matches any focus.

>>> has just (Just 1)
True
>>> has just Nothing
False

hasn't :: AFold0 a s a -> s -> Bool

Check whether an optic has no focus.

anyOf :: AFold Any s a -> (a -> Bool) -> s -> Bool

Check whether any focus matchOf a predicate.

allOf :: AFold All s a -> (a -> Bool) -> s -> Bool

Check whether all foci match a predicate.

noneOf :: AFold Any s a -> (a -> Bool) -> s -> Bool

Check that no focus matchOf a predicate.

lengthOf :: AFold (Sum Int) s a -> s -> Int

Count the number of foci.

sumOf :: Num a => AFold (Sum a) s a -> s -> a

Sum all foci.

productOf :: Num a => AFold (Product a) s a -> s -> a

Multiply all foci.

maximumOf :: Ord a => AFold (((Endo - Endo) - Maybe) a) s a -> s -> Maybe a

Find the maximum focus, if any.

minimumOf :: Ord a => AFold (((Endo - Endo) - Maybe) a) s a -> s -> Maybe a

Find the minimum focus, if any.

maximumByOf :: AFold (((Endo - Endo) - Maybe) a) s a -> (a -> a -> Ordering) -> s -> Maybe a

Find the maximum focus by a comparison function, if any.

minimumByOf :: AFold (((Endo - Endo) - Maybe) a) s a -> (a -> a -> Ordering) -> s -> Maybe a

Find the minimum focus by a comparison function, if any.

findOf :: AFold ((Endo - Maybe) a) s a -> (a -> Bool) -> s -> Maybe a

Find the first focus matching a predicate, if any.

elemOf :: Eq a => AFold Any s a -> a -> s -> Bool

Check whether a value is a focus of an optic.

headOf :: AFold ((Endo - Maybe) a) s a -> s -> Maybe a

Retrieve the first focus, if any.

lastOf :: AFold (((Endo - Endo) - Maybe) a) s a -> s -> Maybe a

Retrieve the last focus, if any.

Fold0, Ixfold0

foldOf0 :: AFold0 r s a -> (a -> Maybe r) -> s -> Maybe r

TODO: Document

ixfoldOf0 :: Monoid k => AIxfold0 r k s a -> (k -> a -> Maybe r) -> s -> Maybe r

TODO: Document

Since: 0.0.3

preview :: MonadReader s m => AFold0 a s a -> m (Maybe a)

An infix alias for preview'.

(^?) ≡ flip preview'

Perform a safe head of a Fold or Traversal or retrieve Just the result from a View or Lens.

When using a Traversal as a partial Lens, or a Fold as a partial View this can be a convenient way to extract the optional value. | Preview the focus of a Fold0.

Benchmark: 1.01x vs direct (zero-cost). See Data.Profunctor.Optic.Bench.

ixpreview :: Monoid k => AIxfold0 (k, a) k s a -> s -> Maybe (k, a)

TODO: Document

Since: 0.0.3

previews :: MonadReader s m => AFold0 r s a -> (a -> r) -> m (Maybe r)

TODO: Document

ixpreviews :: Monoid k => AIxfold0 r k s a -> (k -> a -> r) -> s -> Maybe r

TODO: Document

Since: 0.0.3

preuse :: MonadState s m => AFold0 a s a -> m (Maybe a)

TODO: Document

preuses :: MonadState s m => AFold0 r s a -> (a -> r) -> m (Maybe r)

TODO: Document

Dual Operators

Cofold, Cxfold

cofoldMapOf :: ACotraversal (Const r) s t a b -> (r -> b) -> r -> t

TODO: Document

>>> cofoldMapOf (from succ) (*2) 3
7

Compare reviews.

cofoldOfA :: Coapplicative f => ACofold (f b) t b -> f b -> t

TODO: Document

cxfoldMapOf :: Monoid k => ACxfold r k t b -> (k -> r -> b) -> r -> t

TODO: Document

Since: 0.0.3

Auxiliary Types

newtype EndoM m a

Constructors

EndoM

Fields

appEndoM :: a -> m a

Reexports

class Profunctor p => Strong (p :: Type -> Type -> Type) where

Generalizing Star of a strong Functor

Note: Every Functor in Haskell is strong with respect to (,).

This describes profunctor strength with respect to the product structure of Hask.

http://www.riec.tohoku.ac.jp/~asada/papers/arrStrMnd.pdf

Minimal complete definition

first' | second'

Methods

first' :: p a b -> p (a, c) (b, c)

Laws:

first' ≡ dimap swap swap . second'
lmap fst ≡ rmap fst . first'
lmap (second' f) . first' ≡ rmap (second' f) . first'
first' . first' ≡ dimap assoc unassoc . first' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

second' :: p a b -> p (c, a) (c, b)

Laws:

second' ≡ dimap swap swap . first'
lmap snd ≡ rmap snd . second'
lmap (first' f) . second' ≡ rmap (first' f) . second'
second' . second' ≡ dimap unassoc assoc . second' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

class Profunctor p => Choice (p :: Type -> Type -> Type) where

The generalization of Costar of Functor that is strong with respect to Either.

Note: This is also a notion of strength, except with regards to another monoidal structure that we can choose to equip Hask with: the cocartesian coproduct.

Minimal complete definition

left' | right'

Methods

left' :: p a b -> p (Either a c) (Either b c)

Laws:

left' ≡ dimap swapE swapE . right' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Left ≡ lmap Left . left'
lmap (right f) . left' ≡ rmap (right f) . left'
left' . left' ≡ dimap assocE unassocE . left' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

right' :: p a b -> p (Either c a) (Either c b)

Laws:

right' ≡ dimap swapE swapE . left' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Right ≡ lmap Right . right'
lmap (left f) . right' ≡ rmap (left f) . right'
right' . right' ≡ dimap unassocE assocE . right' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

class Profunctor p => Closed (p :: Type -> Type -> Type) where

A strong profunctor allows the monoidal structure to pass through.

A closed profunctor allows the closed structure to pass through.

Methods

closed :: p a b -> p (x -> a) (x -> b)

Laws:

lmap (. f) . closed ≡ rmap (. f) . closed
closed . closed ≡ dimap uncurry curry . closed
dimap const ($()) . closed ≡ id

class (Sieve p (Rep p), Strong p) => Representable (p :: Type -> Type -> Type) where

A Profunctor p is Representable if there exists a Functor f such that p d c is isomorphic to d -> f c.

Associated Types

type Rep (p :: Type -> Type -> Type) :: Type -> Type

Methods

tabulate :: (d -> Rep p c) -> p d c

Laws:

tabulate . sieve ≡ id
sieve . tabulate ≡ id

class (Cosieve p (Corep p), Costrong p) => Corepresentable (p :: Type -> Type -> Type) where

A Profunctor p is Corepresentable if there exists a Functor f such that p d c is isomorphic to f d -> c.

Associated Types

type Corep (p :: Type -> Type -> Type) :: Type -> Type

Methods

cotabulate :: (Corep p d -> c) -> p d c

Laws:

cotabulate . cosieve ≡ id
cosieve . cotabulate ≡ id