Data.Profunctor.Optic.Iso

Constructors

type Equality s t a b = forall p. Optic p s t a b

type Equality' s a = Equality s s a a

type Iso s t a b = forall p. Profunctor p => Optic p s t a b

type Iso' s a = Iso s s a a

iso :: (s -> a) -> (b -> t) -> Iso s t a b

Obtain an Iso from two inverses.

o . re o ≡ id
re o . o ≡ id
view o (review o b) ≡ b
review o (view o s) ≡ s

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions satisfy the following properties:

sa . bt ≡ id

bt . sa ≡ id

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

isoVl :: (forall f g. Functor f => Functor g => (g a -> f b) -> g s -> f t) -> Iso s t a b

Transform a Van Laarhoven Iso into a profunctor Iso.

reixing :: Profunctor p => AIso' k1 k2 -> Ixoptic p k1 s t a b -> Ixoptic p k2 s t a b

Remap the indices of an indexed optic.

recxing :: Profunctor p => AIso' k1 k2 -> Cxoptic p k1 s t a b -> Cxoptic p k2 s t a b

Remap the indices of a coindexed optic.

fmapping :: Functor f => Functor g => AIso s t a b -> Iso (f s) (g t) (f a) (g b)

Lift an Iso into a pair of functors.

contramapping :: Contravariant f => Contravariant g => AIso s t a b -> Iso (f a) (g b) (f s) (g t)

Lift an Iso into a pair of Contravariant functors.

dimapping :: Profunctor p => Profunctor q => AIso s1 t1 a1 b1 -> AIso s2 t2 a2 b2 -> Iso (p a1 s2) (q b1 t2) (p s1 a2) (q t1 b2)

Lift a pair of Isos into a pair of profunctors.

yoneda :: Profunctor p => Iso s t a b -> p a b -> Yoneda p s t

Lift an Iso into a Yoneda.

coyoneda :: Iso s t a b -> p a b -> Coyoneda p s t

Lift an Iso into a Coyoneda.

inverting :: AIso s t a b -> Iso b a t s

Invert an isomorphism.

inverting (inverting o) ≡ o
inverting ≡ cloneIso . re

cloneIso :: AIso s t a b -> Iso s t a b

Convert from AIso back to any Iso.

Optics

equaled :: s ~ a => t ~ b => Equality s t a b

Obtain an Iso' directly from type equaled constraints.

>>> :t (^. equaled)
(^. equaled) :: a -> a

coerced :: Coercible s a => Coercible t b => Iso s t a b

Obtain an Iso from data types that are representationally equal.

>>> view coerced 'x' :: Identity Char
Identity 'x'

generic :: Generic a => Generic b => Iso a b (Rep a c) (Rep b c)

An Iso between Generic representations.

>>> view (generic . re generic) "hello" :: String
"hello"

generic1 :: Generic1 f => Generic1 g => Iso (f a) (g b) (Rep1 f a) (Rep1 g b)

An Iso between Generic1 representations.

adjuncted :: Adjunction f u => Iso (f a -> b) (f s -> t) (a -> u b) (s -> u t)

An Iso between a functor and its adjoint.

Useful for converting between lens-like optics and grate-like optics:

over adjuncted :: Adjunction f u => ((a -> u b) -> s -> u t) -> (f a -> b) -> f s -> t

tabulated :: Representable f => Representable g => Iso (f a) (g b) (Rep f -> a) (Rep g -> b)

An Iso between a functor and its Yoneda representation.

indexing :: ((a -> b) -> s -> t) -> Iso s t (Index s x x) (Index s a b)

TODO: Document

coindexing :: ((a -> b) -> s -> t) -> Iso s t (Coindex t b a) (Coindex t x x)

TODO: Document

unzipped :: Adjunction f u => Iso (u a, u b) (u c, u d) (u (a, b)) (u (c, d))

A right adjoint admits an intrinsic notion of zipping.

cozipped :: Adjunction f u => Iso (f a + f b) (f c + f d) (f (a + b)) (f (c + d))

A left adjoint must be inhabited by exactly one element.

swapped :: Iso (a, b) (c, d) (b, a) (d, c)

Swap sides of a product.

eswapped :: Iso (a + b) (c + d) (b + a) (d + c)

Swap sides of a sum.

associated :: Iso (a, (b, c)) (d, (e, f)) ((a, b), c) ((d, e), f)

An Iso defined by left-association of nested tuples.

eassociated :: Iso (a + (b + c)) (d + (e + f)) ((a + b) + c) ((d + e) + f)

An Iso defined by left-association of nested tuples.

excised :: Eq a => a -> Iso' (Maybe a) a

Excise a single value from a type.

>>> review (excised "foo") "foo"
Nothing
>>> review (excised "foo") "foobar"
Just "foobar"

flipped :: Iso (a -> b -> c) (d -> e -> f) (b -> a -> c) (e -> d -> f)

Flip two arguments of a function.

>>> (view flipped (,)) 1 2
(2,1)

involuted :: (s -> a) -> Iso s a a s

An Iso defined by a function that is its own inverse.

involuted ≡ join iso

>>> "live" ^. involuted reverse
"evil"

>>> "live" & involuted reverse ..~ ('d':)
"lived"

uncurried :: Iso (a -> b -> c) (d -> e -> f) ((a, b) -> c) ((d, e) -> f)

Uncurry a function.

>>> (fst ^. inverting uncurried) 3 4
3

Operators

au :: Functor f => AIso s t a b -> ((b -> t) -> f s) -> f a

Based on ala from Conor McBride's work on Epigram.

This version is generalized to accept any Iso, not just a newtype.

>>> au (iso getSum Sum) foldMap [1,2,3,4]
10

You may want to think of this combinator as having the following, simpler type:

au :: AIso s t a b -> ((b -> t) -> e -> s) -> e -> a

aup :: Profunctor p => Functor f => AIso s t a b -> (p c a -> f b) -> p c s -> f t

Variant of au for profunctors.

flip aup runStar :: Functor f => AIso s t a (f a) -> Star f c s -> c -> t

reover :: AIso s t a b -> (t -> s) -> b -> a

Given a conversion on one side of an Iso, recover the other.

reover ≡ over . re

Compare over.

withIso :: AIso s t a b -> ((s -> a) -> (b -> t) -> r) -> r

Extract the two functions that characterize an Iso.

re :: Optic (Re p a b) s t a b -> Optic p b a t s

Reverse an optic to obtain its Re-dual.

re . re  ≡ id

re swaps Strong \(\leftrightarrow\) Costrong and Choice \(\leftrightarrow\) Cochoice:

re :: Iso s t a b    -> Iso b a t s
re :: Lens s t a b   -> Relens b a t s
re :: Prism s t a b  -> Reprism b a t s
re :: View s a       -> Review a s
re :: Review t b     -> View b t

Note: this is not the same as the categorical co-dual (Colens, Cotraversal, etc.), which replaces Strong with Closed.

>>> 5 ^. re left'
Left 5

Auxiliary Types

newtype Re p s t a b

The Re type and its instances witness the symmetry between the parameters of a Profunctor.

Re can reverse optics built from classes that have a mirror image:

Strong \(\leftrightarrow\) Costrong

Choice \(\leftrightarrow\) Cochoice

Bifunctor \(\leftrightarrow\) Contravariant

Re cannot reverse Closed, Representable, or Corepresentable because these classes have no standard dual. A hypothetical Coclosed with unclosed :: p (x -> a) (x -> b) -> p a b would enable Closed p => Coclosed (Re p), but no such class exists in the profunctor hierarchy. Consequently, Colens, Cotraversal, and other Closed-based optics cannot be reversed with re.

Constructors

Re

Fields

runRe :: p b a -> p t s

Classes

class Profunctor (p :: Type -> Type -> Type) where

Formally, the class Profunctor represents a profunctor from Hask -> Hask.

Intuitively it is a bifunctor where the first argument is contravariant and the second argument is covariant.

You can define a Profunctor by either defining dimap or by defining both lmap and rmap.

If you supply dimap, you should ensure that:

dimap id id ≡ id

If you supply lmap and rmap, ensure:

lmap id ≡ id
rmap id ≡ id

If you supply both, you should also ensure:

dimap f g ≡ lmap f . rmap g

These ensure by parametricity:

dimap (f . g) (h . i) ≡ dimap g h . dimap f i
lmap (f . g) ≡ lmap g . lmap f
rmap (f . g) ≡ rmap f . rmap g

Minimal complete definition

dimap | lmap, rmap

Methods

dimap :: (a -> b) -> (c -> d) -> p b c -> p a d

Map over both arguments at the same time.

dimap f g ≡ lmap f . rmap g

lmap :: (a -> b) -> p b c -> p a c

Map the first argument contravariantly.

lmap f ≡ dimap f id

rmap :: (b -> c) -> p a b -> p a c

Map the second argument covariantly.

rmap ≡ dimap id

(#.) :: forall a b c q. Coercible c b => q b c -> p a b -> p a c infixr 9

Strictly map the second argument argument covariantly with a function that is assumed operationally to be a cast, such as a newtype constructor.

Note: This operation is explicitly unsafe since an implementation may choose to use unsafeCoerce to implement this combinator and it has no way to validate that your function meets the requirements.

If you implement this combinator with unsafeCoerce, then you are taking upon yourself the obligation that you don't use GADT-like tricks to distinguish values.

If you import Data.Profunctor.Unsafe you are taking upon yourself the obligation that you will only call this with a first argument that is operationally identity.

The semantics of this function with respect to bottoms should match the default definition:

(#.) ≡ \_ -> \p -> p `seq` rmap coerce p

(.#) :: forall a b c q. Coercible b a => p b c -> q a b -> p a c infixl 8

Strictly map the first argument argument contravariantly with a function that is assumed operationally to be a cast, such as a newtype constructor.

Note: This operation is explicitly unsafe since an implementation may choose to use unsafeCoerce to implement this combinator and it has no way to validate that your function meets the requirements.

If you implement this combinator with unsafeCoerce, then you are taking upon yourself the obligation that you don't use GADT-like tricks to distinguish values.

If you import Data.Profunctor.Unsafe you are taking upon yourself the obligation that you will only call this with a second argument that is operationally identity.

(.#) ≡ \p -> p `seq` \f -> lmap coerce p