Data.Profunctor.Optic.Lens

Constructors

Lens, Ixlens

type Lens s t a b = forall p. Strong p => Optic p s t a b

type Lens' s a = Lens s s a a

type Ixlens k s t a b = forall p. Strong p => Ixoptic p k s t a b

type Ixlens' k s a = Ixlens k s s a a

lens :: (s -> a) -> (s -> b -> t) -> Lens s t a b

Obtain a Lens from a getter and setter.

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions satisfy the following properties:

sa (sbt s a) ≡ a

sbt s (sa s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

See Property.

ixlens :: (s -> (k, a)) -> (s -> b -> t) -> Ixlens k s t a b

Obtain an indexed Lens from an indexed getter and a setter.

Compare lens and itraversal.

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions constitute a legal indexed lens:

snd . sia (sbt s a) ≡ a

sbt s (snd $ sia s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

See Property.

Since: 0.0.3

lensVl :: (forall f. Functor f => (a -> f b) -> s -> f t) -> Lens s t a b

Transform a Van Laarhoven lens into a profunctor lens.

Compare grateVl and traversalVl.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

abst Identity ≡ Identity

fmap (abst f) . (abst g) ≡ getCompose . abst (Compose . fmap f . g)

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

ixlensVl :: (forall f. Functor f => (k -> a -> f b) -> s -> f t) -> Ixlens k s t a b

Transform an indexed Van Laarhoven lens into an indexed profunctor Lens.

An Ixlens is a valid Ixtraversal. Compare itraversalVl.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

iabst (const Identity) ≡ Identity

fmap (iabst $ const f) . (iabst $ const g) ≡ getCompose . iabst (const $ Compose . fmap f . g)

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

Since: 0.0.3

matching :: (s -> (c, a)) -> ((c, b) -> t) -> Lens s t a b

Obtain a Lens from its free tensor representation.

cloneLens :: ALens s t a b -> Lens s t a b

TODO: Document

cloneLensVl :: ALens s t a b -> forall f. Functor f => (a -> f b) -> s -> f t

Extract the higher order function that characterizes a Lens.

The lens laws can be stated in terms of withLens:

Identity:

cloneLensVl o Identity ≡ Identity

Composition:

Compose . fmap (cloneLensVl o f) . cloneLensVl o g ≡ cloneLensVl o (Compose . fmap f . g)

See Property.

cloneIxlens :: Monoid k => AIxlens k s t a b -> Ixlens k s t a b

Clone an Ixlens.

Since: 0.0.3

cloneIxlensVl :: Monoid k => AIxlens k s t a b -> forall f. Functor f => (k -> a -> f b) -> s -> f t

Extract the indexed Van Laarhoven form of an Ixlens.

Since: 0.0.3

Dual Constructors

Colens, Cxlens

type Colens s t a b = forall p. Closed p => Optic p s t a b

type Colens' s a = Colens s s a a

type Cxlens k s t a b = forall p. Closed p => Cxoptic p k s t a b

type Cxlens' k t b = Cxlens k t t b b

colens :: (b -> s -> a) -> (b -> t) -> Colens s t a b

Obtain a Colens from a getter and setter.

colens f g ≡ \f g -> re (lens f g)
colens bsia bt ≡ colensVl $ \ts b -> bsia b <$> (ts . bt $ b)
review $ colens f g ≡ f
set . re $ re (lens f g) ≡ g

Caution: Colenses are recursive, similar to ArrowLoop. In addition to the normal optic laws, the input functions must have the correct laziness annotations.

For example, this is a perfectly valid Colens:

ct21 :: Colens a b (a, c) (b, c)
ct21 = flip colens fst $  ~(_,c) b -> (b,c)

However removing the annotation will result in a faulty optic.

See Property.

cxlens :: (((s -> a) -> k -> b) -> t) -> Cxlens k s t a b

TODO: Document

Since: 0.0.3

colensVl :: (forall f. Functor f => (t -> f s) -> b -> f a) -> Colens s t a b

Transform a Van Laarhoven colens into a profunctor colens.

Compare grateVl.

Caution: In addition to the normal optic laws, the input functions must have the correct laziness annotations.

For example, this is a perfectly valid Colens:

ct21 :: Colens a b (a, c) (b, c)
ct21 = colensVl $ f ~(a,b) -> (,b) $ f a

However removing the annotation will result in a faulty optic.

cxlensVl :: (forall f. Functor f => (f a -> k -> b) -> f s -> t) -> Cxlens k s t a b

Transform a coindexed Van Laarhoven grate into a coindexed profunctor grate.

Since: 0.0.3

grate :: (((s -> a) -> b) -> t) -> Colens s t a b

Obtain a Colens from a nested continuation.

The resulting optic is the corepresentable counterpart to Lens, and sits between Iso and Setter.

A Colens lets you lift a profunctor through any representable functor (aka Naperian container). In the special case where the indexing type is finitary (e.g. Bool) then the tabulated type is isomorphic to a fixed length vector (e.g. 'V2 a').

The identity container is representable, and representable functors are closed under composition.

See https://www.cs.ox.ac.uk/jeremy.gibbons/publications/proyo.pdf section 4.6 for more background on Colenss, and compare to the lens-family version.

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

sabt ($ s) ≡ s

sabt (k -> f (k . sabt)) ≡ sabt (k -> f ($ k))

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

grate' :: (s -> a) -> (b -> t) -> Colens s t a b

Construct a Colens from a pair of inverses.

grateVl :: (forall f. Functor f => (f a -> b) -> f s -> t) -> Colens s t a b

Transform a Van Laarhoven grate into a profunctor grate.

Compare lensVl & cotraversalVl.

Caution: In order for the generated family to be well-defined, you must ensure that the traversal1 law holds for the input function:

abst runIdentity ≡ runIdentity

abst f . fmap (abst g) ≡ abst (f . fmap g . getCompose) . Compose

See Property.

comatching :: ((c, s) -> a) -> (b -> (c, t)) -> Colens s t a b

Obtain a Colens from its free tensor representation.

>>> fib = comatching (uncurry L.take . swap) (fanout id L.reverse) --fib :: Colens Int [Int] [Int] [Int]
>>> 10 & fib ..~ \xs -> 1 : 1 : Prelude.zipWith (+) xs (drop 1 xs)
[89,55,34,21,13,8,5,3,2,1,1]

inside :: Corepresentable p => ALens s t a b -> Lens (p e s) (p e t) (p e a) (p e b)

Lift a Lens so it can run under a function (or other corepresentable profunctor).

inside :: Lens s t a b -> Lens (e -> s) (e -> t) (e -> a) (e -> b)

>>> (\x -> (x-1,x+1)) ^. inside first $ 5
4

cloneColens :: AColens s t a b -> Colens s t a b

TODO: Document

cloneColensVl :: AColens s t a b -> forall f. Functor f => (f a -> b) -> f s -> t

Extract the higher order function that characterizes a Colens.

The grate laws can be stated in terms or withColens:

Identity:

cloneColensVl o runIdentity ≡ runIdentity

Composition:

cloneColensVl o f . fmap (cloneColensVl o g) ≡ cloneColensVl o (f . fmap g . getCompose) . Compose

cloneCxlens :: Monoid k => ACxlens k s t a b -> Cxlens k s t a b

Clone a Cxlens.

Since: 0.0.3

cloneCxlensVl :: Monoid k => ACxlens k s t a b -> forall f. Functor f => (f a -> k -> b) -> f s -> t

Extract the coindexed Van Laarhoven form of a Cxlens.

Since: 0.0.3

Relens, Rxlens

type Relens s t a b = forall p. Costrong p => Optic p s t a b

type Relens' s a = Relens s s a a

type Rxlens k s t a b = forall p. Costrong p => Ixoptic p k s t a b

type Rxlens' k s a = Rxlens k s s a a

relens :: (b -> s -> a) -> (b -> t) -> Relens s t a b

Obtain a Relens from a co-getter and co-setter.

relens bsa bt ≡ re (lens sa sbt) (with roles swapped)

A Relens is simultaneously a View and a Review:

review (relens bsa bt) ≡ bt

relensVl :: (forall f. Functor f => (t -> f s) -> b -> f a) -> Relens s t a b

Obtain a Relens from its van Laarhoven representation.

rematching :: (c -> s -> a) -> (b -> (c, t)) -> Relens s t a b

Obtain a Relens from its free tensor representation.

rematching' :: (t -> s -> a) -> Relens s t a t

Obtain a Relens from a single combining function.

rematching' f ≡ rematching f (\a -> (a, a))

cloneRelens :: ARelens s t a b -> Relens s t a b

Clone a Relens.

cloneRelensVl :: ARelens s t a b -> forall f. Functor f => (t -> f s) -> b -> f a

Extract the Van Laarhoven form of a Relens.

Since: 0.0.3

rxlens :: (b -> s -> (r, a)) -> (b -> t) -> Rxlens r s t a b

Obtain an indexed Relens from an indexed getter and a setter.

Since: 0.0.3

rxlensVl :: (forall f. Functor f => (t -> f s) -> b -> f (r, a)) -> Rxlens r s t a b

Transform an indexed Van Laarhoven relens into an indexed profunctor Relens.

Compare ixlensVl and relensVl.

Since: 0.0.3

Optics

Lens, Ixlens

first :: Lens (a, c) (b, c) a b

TODO: Document

second :: Lens (c, a) (c, b) a b

TODO: Document

ixfirst :: Ixlens k (a, c) (b, c) a b

TODO: Document

>>> B.first getSum <$> ixtoListOf (noix traversed . ixfirst . ix (Sum 1) traversed) [("foo",1), ("bar",2)]
[(0,'f'),(1,'o'),(2,'o'),(0,'b'),(1,'a'),(2,'r')]
>>> B.first getSum <$> ixtoListOf (ix (Sum 3) traversed % ixfirst % ix (Sum 1) traversed) [("foo",1), ("bar",2)]
[(0,'f'),(1,'o'),(2,'o'),(3,'b'),(4,'a'),(5,'r')]

Since: 0.0.3

ixsecond :: Ixlens k (c, a) (c, b) a b

TODO: Document

Since: 0.0.3

united :: Lens' a ()

There is a () in everything.

>>> "hello" ^. united
()
>>> "hello" & united .~ ()
"hello"

voided :: Lens' Void a

There is everything in a Void.

>>> Nothing & fmapped . voided ..~ abs
Nothing

Colens, Cxlens

cofirst :: Colens a b (a, c) (b, c)

TODO: Document

cosecond :: Colens a b (c, a) (c, b)

TODO: Document

cxfirst :: Cxlens k a b (a, c) (b, c)

TODO: Document

Since: 0.0.3

cxsecond :: Cxlens k a b (c, a) (c, b)

TODO: Document

Since: 0.0.3

closed :: Closed p => p a b -> p (x -> a) (x -> b)

Laws:

lmap (. f) . closed ≡ rmap (. f) . closed
closed . closed ≡ dimap uncurry curry . closed
dimap const ($()) . closed ≡ id

cxclosed :: Cxlens k (c -> a) (c -> b) a b

TODO: Document

>>> cxover cxclosed (,) (*2) 5
((),10)

Since: 0.0.3

represented :: Representable f => Colens (f a) (f b) a b

Obtain a Colens from a Representable functor.

distributed :: Distributive f => Colens (f a) (f b) a b

Obtain a Colens from a distributive functor.

endomorphed :: Colens' (Endo a) a

Obtain a Colens from an endomorphism.

>>> flip appEndo 2 $ zipsWith endomorphed (+) (Endo (*3)) (Endo (*4))
14

continued :: Colens c (Cont a c) a a

Obtain a Colens from a continuation.

zipsWith continued :: (a -> a -> a) -> c -> c -> Cont a c

continuedT :: Colens c (ContT a m c) (m a) (m a)

Obtain a Colens from a continuation.

zipsWith continued :: (m a -> m a -> m a) -> c -> c -> ContT a m c

calledCC :: MonadCont m => Colens a (m a) (m b) (m a)

Lift the current continuation into the calling context.

zipsWith calledCC :: MonadCont m => (m b -> m b -> m s) -> s -> s -> m s

Relens, Rxlens

refirst :: Relens a b (a, c) (b, c)

Relens into the first component of a pair.

refirst ≡ re first

resecond :: Relens a b (c, a) (c, b)

Relens into the second component of a pair.

resecond ≡ re second

Operators

Lens, Ixlens

pastro :: ALens s t a b -> p a b -> Pastro p s t

Use a Lens to construct a Pastro.

tambara :: Strong p => ALens s t a b -> p a b -> Tambara p s t

Use a Lens to construct a Tambara.

withLens :: ALens s t a b -> ((s -> a) -> (s -> b -> t) -> r) -> r

Extract the two functions that characterize a Lens.

withIxlens :: Monoid k => AIxlens k s t a b -> ((s -> (k, a)) -> (s -> b -> t) -> r) -> r

Extract the two functions that characterize an Ixlens.

Since: 0.0.3

Dual Operators

Colens, Cxlens

coview :: AColens s t a b -> b -> t

Set all fields to the given value.

Compare review.

cxzips :: Monoid k => ACxlens k s t a b -> (k -> a -> a -> b) -> s -> s -> t

TODO: Document

Since: 0.0.3

zipsWith :: AColens s t a b -> (a -> a -> b) -> s -> s -> t

Zip over a Colens.

\f -> zipsWith closed (zipsWith closed f) ≡ zipsWith (closed . closed)

zipsWith3 :: AColens s t a b -> (a -> a -> a -> b) -> s -> s -> s -> t

Zip over a Colens with 3 arguments.

zipsWith4 :: AColens s t a b -> (a -> a -> a -> a -> b) -> s -> s -> s -> s -> t

Zip over a Colens with 4 arguments.

zipsWithF :: Functor f => AColens s t a b -> (f a -> b) -> f s -> t

Extract the higher order function that characterizes a Colens.

The grate laws can be stated in terms or withColens:

Identity:

zipsWithF o runIdentity ≡ runIdentity

Composition:

zipsWithF o f . fmap (zipsWithF o g) ≡ zipsWithF o (f . fmap g . getCompose) . Compose

closure :: Closed p => AColens s t a b -> p a b -> Closure p s t

Use a Colens to construct a Closure.

environment :: Closed p => AColens s t a b -> p a b -> Environment p s t

Use a Colens to construct an Environment.

withColens :: AColens s t a b -> ((((s -> a) -> b) -> t) -> r) -> r

Extract the function that characterizes a Colens.

withCxlens :: Monoid k => ACxlens k s t a b -> ((((s -> a) -> k -> b) -> t) -> r) -> r

Extract the function that characterizes a Cxlens.

Since: 0.0.3

Relens, Rxlens

reover :: ARelens s t a b -> (t -> s) -> b -> a

The Re-dual of over: apply a function through a Relens, going in the reverse direction.

over   :: ALens   s t a b -> (a -> b) -> s -> t
reover :: ARelens s t a b -> (t -> s) -> b -> a

reover accepts any optic with a RelensRep carrier (ARelens), including Relens, Iso, and any optic that is both Strong and Costrong.

Example: collapse a pair-producing Sort to its first component:

reover refirst :: (t -> s) -> (a, c) -> a

withRelens :: ARelens s t a b -> ((b -> s -> a) -> (b -> t) -> r) -> r

Extract the two functions that characterize a Relens.

Reexports

class Profunctor p => Strong (p :: Type -> Type -> Type) where

Generalizing Star of a strong Functor

Note: Every Functor in Haskell is strong with respect to (,).

This describes profunctor strength with respect to the product structure of Hask.

http://www.riec.tohoku.ac.jp/~asada/papers/arrStrMnd.pdf

Minimal complete definition

first' | second'

Methods

first' :: p a b -> p (a, c) (b, c)

Laws:

first' ≡ dimap swap swap . second'
lmap fst ≡ rmap fst . first'
lmap (second' f) . first' ≡ rmap (second' f) . first'
first' . first' ≡ dimap assoc unassoc . first' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

second' :: p a b -> p (c, a) (c, b)

Laws:

second' ≡ dimap swap swap . first'
lmap snd ≡ rmap snd . second'
lmap (first' f) . second' ≡ rmap (first' f) . second'
second' . second' ≡ dimap unassoc assoc . second' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

class Profunctor p => Closed (p :: Type -> Type -> Type) where

A strong profunctor allows the monoidal structure to pass through.

A closed profunctor allows the closed structure to pass through.

Methods

closed :: p a b -> p (x -> a) (x -> b)

Laws:

lmap (. f) . closed ≡ rmap (. f) . closed
closed . closed ≡ dimap uncurry curry . closed
dimap const ($()) . closed ≡ id

class Profunctor p => Costrong (p :: Type -> Type -> Type) where

Analogous to ArrowLoop, loop = unfirst

Minimal complete definition

unfirst | unsecond

Methods

unfirst :: p (a, d) (b, d) -> p a b

Laws:

unfirst ≡ unsecond . dimap swap swap
lmap (,()) ≡ unfirst . rmap (,())
unfirst . lmap (second f) ≡ unfirst . rmap (second f)
unfirst . unfirst = unfirst . dimap assoc unassoc where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

unsecond :: p (d, a) (d, b) -> p a b

Laws:

unsecond ≡ unfirst . dimap swap swap
lmap ((),) ≡ unsecond . rmap ((),)
unsecond . lmap (first f) ≡ unsecond . rmap (first f)
unsecond . unsecond = unsecond . dimap unassoc assoc where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)