Data.Profunctor.Optic.Prism

Description

Left and right carriers are characterized by their representation functors (Rep and Corep). Iso sits at the top where both are trivial (Identity). Adjoint sits at the bottom where they form a full adjunction (Corep p ⊣ Rep p). A left-indexed (Ix) optic threads an index through the left adjoint functor (,) k, while a right-indexed (Cx) optic threads a coindex through the right adjoint functor (->) k.

Constructors

Prism, Ixprism

type Prism s t a b = forall p. Choice p => Optic p s t a b

type Prism' s a = Prism s s a a

type Ixprism k s t a b = forall p. Choice p => Ixoptic p k s t a b

type Ixprism' k s a = Ixprism k s s a a

prism :: (s -> t + a) -> (b -> t) -> Prism s t a b

Obtain a Prism from a constructor and a matcher function.

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions satisfy the following properties:

sta (bt b) ≡ Right b

(either id bt) (sta s) ≡ s

left sta (sta s) ≡ left Left (sta s)

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

prism' :: (s -> Maybe a) -> (b -> s) -> Prism s s a b

Obtain a Prism' from a reviewer and a matcher function that produces a Maybe.

Note: The arguments are reversed from the equivalent in the lens package. This is unfortunate but done to maintain consistency with traversal0 etc.

ixprism :: (s -> t + (k, a)) -> (b -> t) -> Ixprism k s t a b

Obtain an Ixprism from an indexed matcher and a constructor.

Since: 0.0.3

ixprism' :: (s -> Maybe (k, a)) -> (b -> s) -> Ixprism k s s a b

Obtain an Ixprism' from an indexed matcher and a constructor.

Since: 0.0.3

handling :: (s -> c + a) -> ((c + b) -> t) -> Prism s t a b

Obtain a Prism from its free tensor representation.

Useful for constructing prisms from try and handle functions.

ixhandling :: (s -> c + (k, a)) -> ((c + b) -> t) -> Ixprism k s t a b

Indexed handling: obtain an Ixprism from an indexed matcher and a constructor on the complement.

Since: 0.0.3

clonePrism :: APrism s t a b -> Prism s t a b

TODO: Document

cloneIxprism :: Monoid k => AIxprism k s t a b -> Ixprism k s t a b

Clone an Ixprism.

Since: 0.0.3

Reprism, Rxprism

type Reprism s t a b = forall p. Cochoice p => Optic p s t a b

type Reprism' s a = Reprism s s a a

type Rxprism k s t a b = forall p. Cochoice p => Ixoptic p k s t a b

type Rxprism' k s a = Rxprism k s s a a

reprism :: (s -> a) -> (b -> Either a t) -> Reprism s t a b

Obtain a Reprism from a viewer and a matcher.

reprism sa bat ≡ re (prism sta bt) (with roles swapped)

A Reprism is simultaneously a View and a Review:

view (reprism sa bat) ≡ sa

reprism' :: (s -> a) -> (a -> Maybe s) -> Reprism' s a

Obtain a simple Reprism from a viewer and a Maybe matcher.

rxprism :: (s -> (k, a)) -> (b -> (k, a) + t) -> Rxprism k s t a b

Indexed Reprism: coindexed by k.

The viewer returns (k, a) and the matcher returns (k, a) + t.

Since: 0.0.3

rxprism' :: Monoid k => (s -> (k, a)) -> ((k, a) -> Maybe s) -> Rxprism' k s a

Simple indexed Reprism from a viewer and a Maybe matcher.

Since: 0.0.3

rehandling :: (Either c s -> a) -> (b -> Either c t) -> Reprism s t a b

Obtain a Reprism from its free tensor representation.

rehandling' :: (b -> Either s t) -> Reprism s t s b

Obtain a Reprism from a single discriminating function.

rehandling' f ≡ rehandling (either id id) f

cloneReprism :: AReprism s t a b -> Reprism s t a b

Clone a Reprism.

Optics

Prism, Ixprism

left :: Prism (a + c) (b + c) a b

Focus on the Left constructor of Either.

right :: Prism (c + a) (c + b) a b

Focus on the Right constructor of Either.

just :: Prism (Maybe a) (Maybe b) a b

Focus on the Just constructor of Maybe.

>>> Just 1 :| [Just 2, Just 3] & cotraverseOf just sum
Just 6
>>> Nothing :| [Just 2, Just 3] & cotraverseOf just sum
Nothing

nothing :: Prism (Maybe a) (Maybe b) () ()

Focus on the Nothing constructor of Maybe.

ixleft :: Monoid k => Ixprism k (a + c) (b + c) a b

Indexed left. The index is mempty.

Since: 0.0.3

ixright :: Monoid k => Ixprism k (c + a) (c + b) a b

Indexed right. The index is mempty.

Since: 0.0.3

ixjust :: Monoid k => Ixprism k (Maybe a) (Maybe b) a b

Indexed just. The index is mempty.

Since: 0.0.3

prefixed :: Eq a => [a] -> Prism' [a] [a]

Focus on the remainder of a list with a given prefix.

only :: Eq a => a -> Prism' a ()

Focus not just on a case, but a specific value of that case.

nearly :: a -> (a -> Bool) -> Prism' a ()

Create a Prism from a value and a predicate.

>>> review (nearly [] null) ()
[]
>>> [1,2,3,4] ^? nearly [] null
Nothing

nearly [] null :: Prism' [a] ()

Caution: In order for the generated optic to be well-defined, you must ensure that f x holds iff x ≡ a.

nthbit :: Bits s => Int -> Prism' s ()

Focus on the truth value of the nth bit in a bit array.

Reprism, Rxprism

releft :: Reprism a b (Either a c) (Either b c)

Reprism out of the Left constructor.

releft ≡ re left

reright :: Reprism a b (Either c a) (Either c b)

Reprism out of the Right constructor.

reright ≡ re right

Operators

aside :: APrism s t a b -> Prism (e, s) (e, t) (e, a) (e, b)

Use a Prism to lift part of a structure.

without :: APrism s t a b -> APrism u v c d -> Prism (s + u) (t + v) (a + c) (b + d)

Given a pair of prisms, project sums.

below :: Traversable f => APrism' s a -> Prism' (f s) (f a)

Lift a Prism through a Traversable functor.

Returns a Prism that matchOf only if each element matchOf the original Prism.

>>> [Left 1, Right "foo", Left 4, Right "woot"] ^.. below right'
[]
>>> [Right "hail hydra!", Right "foo", Right "blah", Right "woot"] ^.. below right'
[["hail hydra!","foo","blah","woot"]]

withPrism :: APrism s t a b -> ((s -> t + a) -> (b -> t) -> r) -> r

Extract the two functions that characterize a Prism.

withIxprism :: Monoid k => AIxprism k s t a b -> ((s -> t + (k, a)) -> (b -> t) -> r) -> r

Extract the two functions that characterize an Ixprism.

Since: 0.0.3

pastroSum :: APrism s t a b -> p a b -> PastroSum p s t

Use a Prism to construct a PastroSum.

tambaraSum :: Choice p => APrism s t a b -> p a b -> TambaraSum p s t

Use a Prism to construct a TambaraSum.

Dual Operators

rematches :: AReprism s t a b -> b -> a + t

The Re-dual of matchOf: match through a Reprism, going in the reverse direction.

matchOf    :: APrism   s t a b -> s -> t + a
rematches :: AReprism s t a b -> b -> a + t

rematches accepts any optic with a ReprismRep carrier (AReprism), including Reprism, Iso, and any optic that is both Choice and Cochoice.

Example: filter an Either-producing Sort to the Left branch:

rematches releft :: b -> (Either a c) + a

withReprism :: AReprism s t a b -> ((s -> a) -> (b -> a + t) -> r) -> r

Extract the two functions that characterize a Reprism.

Reexports

class Profunctor p => Choice (p :: Type -> Type -> Type) where

The generalization of Costar of Functor that is strong with respect to Either.

Note: This is also a notion of strength, except with regards to another monoidal structure that we can choose to equip Hask with: the cocartesian coproduct.

Minimal complete definition

left' | right'

Methods

left' :: p a b -> p (Either a c) (Either b c)

Laws:

left' ≡ dimap swapE swapE . right' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Left ≡ lmap Left . left'
lmap (right f) . left' ≡ rmap (right f) . left'
left' . left' ≡ dimap assocE unassocE . left' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

right' :: p a b -> p (Either c a) (Either c b)

Laws:

right' ≡ dimap swapE swapE . left' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Right ≡ lmap Right . right'
lmap (left f) . right' ≡ rmap (left f) . right'
right' . right' ≡ dimap unassocE assocE . right' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

class Profunctor p => Cochoice (p :: Type -> Type -> Type) where

Minimal complete definition

unleft | unright

Methods

unleft :: p (Either a d) (Either b d) -> p a b

Laws:

unleft ≡ unright . dimap swapE swapE where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap (either id absurd) ≡ unleft . lmap (either id absurd)
unfirst . rmap (second f) ≡ unfirst . lmap (second f)
unleft . unleft ≡ unleft . dimap assocE unassocE where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

unright :: p (Either d a) (Either d b) -> p a b

Laws:

unright ≡ unleft . dimap swapE swapE where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap (either absurd id) ≡ unright . lmap (either absurd id)
unsecond . rmap (first f) ≡ unsecond . lmap (first f)
unright . unright ≡ unright . dimap unassocE assocE where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c