Data.Profunctor.Optic.Property

Iso

type Iso s t a b = forall p. Profunctor p => Optic p s t a b

fromto_iso :: Eq s => Iso' s a -> s -> Bool

Going back and forth doesn't change anything.

tofrom_iso :: Eq a => Iso' s a -> a -> Bool

Going back and forth doesn't change anything.

Prism

type Prism s t a b = forall p. Choice p => Optic p s t a b

tofrom_prism :: Eq s => Prism' s a -> s -> Bool

If we are able to view an existing focus, then building it will return the original structure.

(either id bt) (sta s) ≡ s

fromto_prism :: Eq s => Eq a => Prism' s a -> a -> Bool

If we build a whole from a focus, that whole must contain the focus.

sta (bt b) ≡ Right b

idempotent_prism :: Eq s => Eq a => Prism' s a -> s -> Bool

left sta (sta s) ≡ left Left (sta s)

Reprism

type Reprism s t a b = forall p. Cochoice p => Optic p s t a b

tofrom_reprism :: Eq a => Reprism' s a -> a -> Bool

If we build and then match, we get back the original.

either id sa (bat a) ≡ a

fromto_reprism :: Eq s => Eq a => Reprism' s a -> s -> Bool

If we match a built value, we always get Right.

bat (sa s) ≡ Right s

idempotent_reprism :: Eq s => Eq a => Reprism' s a -> a -> Bool

Matching the result of a match is the same as wrapping in Left.

left' bat (bat a) ≡ left' Left (bat a)

Lens

type Lens s t a b = forall p. Strong p => Optic p s t a b

id_lens :: Eq s => Lens' s a -> s -> Bool

tofrom_lens :: Eq s => Lens' s a -> s -> Bool

You get back what you put in.

view o (set o b a) ≡ b

fromto_lens :: Eq a => Lens' s a -> s -> a -> Bool

Putting back what you got doesn't change anything.

set o (view o a) a  ≡ a

idempotent_lens :: Eq s => Lens' s a -> s -> a -> a -> Bool

Setting twice is the same as setting once.

set o c (set o b a) ≡ set o c a

Relens

type Relens s t a b = forall p. Costrong p => Optic p s t a b

const_relens :: Eq a => Relens' s a -> a -> Bool

Co-get-set: setting to what we got gives back the structure.

bsa a (bt a) ≡ a

tofrom_relens :: Eq s => Relens' s a -> a -> s -> Bool

Co-set-get: getting from what we set gives back the focus.

bt (bsa a s) ≡ s

idempotent_relens :: Eq a => Relens' s a -> a -> s -> s -> Bool

Co-set-set: setting twice is the same as setting once.

bsa (bsa a s1) s2 ≡ bsa a s2

Ixlens

type Ixlens k s t a b = forall p. Strong p => Ixoptic p k s t a b

tofrom_ixlens :: (Eq s, Monoid k) => Ixlens' k s a -> s -> Bool

You get back what you put in (indexed).

sbt s (snd $ ska s) ≡ s

fromto_ixlens :: (Eq a, Monoid k) => Ixlens' k s a -> s -> a -> Bool

Putting back what you got doesn't change anything (indexed).

snd (ska (sbt s a)) ≡ a

idempotent_ixlens :: (Eq s, Monoid k) => Ixlens' k s a -> s -> a -> a -> Bool

Setting twice is the same as setting once (indexed).

sbt (sbt s a1) a2 ≡ sbt s a2

Colens

type Colens s t a b = forall p. Closed p => Optic p s t a b

id_grate :: Eq s => Colens' s a -> s -> Bool

const_grate :: Eq s => Colens' s a -> s -> Bool

sabt ($ s) ≡ s

compose_grate :: Eq s => Functor f => Functor g => Colens' s a -> (f a -> a) -> (g a -> a) -> f (g s) -> Bool

Traversal0

type Traversal0 s t a b = forall p. Affine p => Optic p s t a b

tofrom_traversal0 :: Eq a => Eq s => Traversal0' s a -> s -> a -> Bool

You get back what you put in.

sta (sbt s a) ≡ either Left (const (Right a)) (sta s)

fromto_traversal0 :: Eq s => Traversal0' s a -> s -> Bool

Putting back what you got doesn't change anything.

either id (sbt s) (sta s) ≡ s

idempotent_traversal0 :: Eq s => Traversal0' s a -> s -> a -> a -> Bool

Setting twice is the same as setting once.

sbt (sbt s a1) a2 ≡ sbt s a2

Traversal

type Traversal s t a b = forall p. (Affine p, Traversing p) => Optic p s t a b

id_traversal :: Eq s => Traversal' s a -> s -> Bool

id_traversal1 :: Eq s => Traversal1' s a -> s -> Bool

pure_traversal :: Eq (f s) => Applicative f => ATraversal' f s a -> s -> Bool

compose_traversal :: Eq (f (g s)) => Applicative' f => Applicative' g => Traversal' s a -> (a -> g a) -> (a -> f a) -> s -> Bool

compose_traversal1 :: Eq (f (g s)) => Apply f => Apply g => Traversal1' s a -> (a -> g a) -> (a -> f a) -> s -> Bool

Ixtraversal

type Ixtraversal k s t a b = forall p. (Affine p, Traversing p) => Ixoptic p k s t a b

id_ixtraversal :: (Eq s, Monoid k) => Ixtraversal' k s a -> s -> Bool

Identity law for indexed traversals.

runIdentity . ixtraverseOf o (const Identity) ≡ id

Cotraversal

type Cotraversal s t a b = forall p. (Coaffine p, Cotraversing p) => Optic p s t a b

compose_cotraversal :: Eq s => Coapplicative f => Coapplicative g => Cotraversal' s a -> (f a -> a) -> (g a -> a) -> f (g s) -> Bool

A Cotraversal is a valid Cosetter with the following additional law:

abst f . fmap (abst g) ≡ abst (f . fmap g . getCompose) . Compose

The cotraversal laws can be restated in terms of cotraverseOf:

cotraverseOf o (f . copure) ≡ fmap f . copure

cotraverseOf o f . fmap (cotraverseOf o g) ≡ cotraverseOf o (f . fmap g . getCompose) . Compose

See also https://www.cs.ox.ac.uk/jeremy.gibbons/publications/iterator.pdf

Cofold

type Cofold t b = forall p. (Coaffine p, Cotraversing p, CoercingL p) => Optic' p t b

id_cofold :: Eq t => ACofold t t t -> t -> Bool

cofoldMapOf o id ≡ id

compose_cofold :: Eq t => ACofold t t t -> (t -> t) -> (t -> t) -> t -> Bool

cofoldMapOf o f . cofoldMapOf o g ≡ cofoldMapOf o (f . g)

Setter

type Setter s t a b = forall p. (Affine p, Traversing p, Mapping p) => Optic p s t a b

id_setter :: Eq s => Setter' s a -> s -> Bool

over o id ≡ id

compose_setter :: Eq s => Setter' s a -> (a -> a) -> (a -> a) -> s -> Bool

over o f . over o g ≡ over o (f . g)

idempotent_setter :: Eq s => Setter' s a -> s -> a -> a -> Bool

set o y (set o x a) ≡ set o y a

Cosetter

type Cosetter s t a b = forall p. (Coaffine p, Cotraversing p, Comapping p) => Optic p s t a b

id_cosetter :: Eq s => ACosetter s s a a -> s -> Bool

cosets o id ≡ id

compose_cosetter :: Eq s => ACosetter s s a a -> (a -> a) -> (a -> a) -> s -> Bool

cosets o f . cosets o g ≡ cosets o (f . g)

Sort

data Sort i k a b

An indexed continuation profunctor for discrimination.

Sort i k a b = (i -> (k, a)) -> b

The indexed generalization of Fmt from stringfmt: Fmt m a b = Sort m () a b

Sort is Costar (Compose ((->) i) ((,) k)). Most instances derive via this representation. Choice is hand-rolled (needs Coapplicative on the Corep functor, requiring Monoid i).

               Profunctor  Closed  Costrong  Cochoice  Choice    Cosieve  Corepresentable  Category
Sort i k          ✓          ✓       ✓         ✓      ✓(Mon i)    ✓           ✓          ✓(Mon i)

id_sort :: Eq b => Sort i k a b -> (i -> (k, a)) -> Bool

dimap id id ≡ id

compose_sort :: Eq c => Sort i k a c -> (a -> a) -> (c -> c) -> (a -> a) -> (c -> c) -> (i -> (k, a)) -> Bool

dimap f g . dimap h k ≡ dimap (h . f) (g . k)

id_category_sort :: (Monoid i, Eq b) => Sort i k a b -> (i -> (k, a)) -> Bool

id . f ≡ f and f . id ≡ f

assoc_category_sort :: (Monoid i, Eq d) => Sort i k a b -> Sort i k b c -> Sort i k c d -> (i -> (k, a)) -> Bool

(f . g) . h ≡ f . (g . h)

Adjoint

type Adjoint s t a b = forall p. Adjoining p => Optic p s t a b

id_adjoint :: Eq s => Adjoint' s a -> s -> Bool

An Adjoint is a Setter, so the Setter laws hold.

over o id ≡ id

compose_adjoint :: Eq s => Adjoint' s a -> (a -> a) -> (a -> a) -> s -> Bool

over o f . over o g ≡ over o (f . g)

idempotent_adjoint :: Eq s => Adjoint' s a -> s -> a -> a -> Bool

over o (const y) (over o (const x) a) ≡ over o (const y) a

id_adjoined :: Eq b => (a -> b) -> a -> Bool

adjoined is the identity for Adjoint optics:

over adjoined f ≡ f

Adjoint Ix/Cx duality

roundtrip_ixadjoining :: Eq t => Cxoptic (Conjoin ()) k s t a b -> (a -> k -> b) -> s -> k -> Bool

ixadjoining . cxadjoining ≡ id on a Cxoptic at Conjoin.

Round-tripping a coindexed optic through Ix and back via the adjunction should be the identity.

Tested at Conjoin j (the canonical Adjoining profunctor). Cx (Conjoin j) k a b = j -> a -> k -> b.

roundtrip_cxadjoining :: Eq t => Ixoptic (Conjoin ()) k s t a b -> ((k, a) -> b) -> (k, s) -> Bool

cxadjoining . ixadjoining ≡ id on an Ixoptic at Conjoin.

Round-tripping an indexed optic through Cx and back via the adjunction should be the identity.

Ix (Conjoin j) k a b = j -> (k, a) -> b.

Sort-Conjoin bridge

retract_embedSort :: Eq b => Conjoin k a b -> k -> a -> Bool

absorbSort . embedSort ≡ id

Embedding Conjoin into Sort and absorbing back is the identity.

Re/Co duality

involutive_re :: (Eq s, Eq a) => Iso' s a -> s -> a -> Bool

re is involutive: re . re ≡ id.

Test on an Iso', since re . re on an Iso gives back an Iso. The predicate checks that view (re (re o)) s ≡ view o s and review (re (re o)) b ≡ review o b.

cxreturn_cxjoin :: Eq b => (a -> b) -> a -> Bool

cxjoin . cxreturn ≡ id

Lifting a profunctor into Cx and immediately collapsing should be the identity.

cxreturn_cxunit :: Eq b => (a -> b) -> a -> Bool

cxunit . cxreturn ≡ id

Same law via cxunit.

roundtrip_ixcx :: Eq t => Cxoptic (->) k s t a b -> (a -> k -> b) -> s -> k -> Bool

ixcx . cxix ≡ id on an Iso.

Round-tripping a Cxoptic through Ix and back should be the identity (at (->), where both Strong and Closed hold).

Cx (->) k a b = a -> (k -> b), so the optic is (a -> (k -> b)) -> (s -> (k -> t)).

roundtrip_cxix :: Eq t => Ixoptic (->) k s t a b -> ((k, a) -> b) -> (k, s) -> Bool

cxix . ixcx ≡ id on an Iso.

Round-tripping an Ixoptic through Cx and back should be the identity (at (->), where both Strong and Closed hold).

Ix (->) k a b = (k, a) -> b, so the optic is ((k, a) -> b) -> ((k, s) -> t).

Ixprism

tofrom_ixprism :: (Eq s, Monoid k) => Ixprism' k s a -> s -> Bool

If we are able to view an existing focus, then building it will return the original structure.

fromto_ixprism :: (Eq s, Eq a, Monoid k) => Ixprism' k s a -> a -> Bool

If we build a whole from a focus, that whole must contain the focus.

idempotent_ixprism :: (Eq s, Eq a, Monoid k) => Ixprism' k s a -> s -> Bool

Matching the result of a match is the same as wrapping in Left.