Data.Profunctor.Optic.Setter

Constructors

Setter, Ixsetter

type Setter s t a b = forall p. (Affine p, Traversing p, Mapping p) => Optic p s t a b

type Setter' s a = Setter s s a a

type Ixsetter k s t a b = forall p. (Affine p, Traversing p, Mapping p) => Ixoptic p k s t a b

type Ixsetter' k s a = Ixsetter k s s a a

setter :: ((a -> b) -> s -> t) -> Setter s t a b

Obtain a Setter from a SEC.

To demote an optic to a semantic edit combinator, use the section (l ..~) or over l.

>>> [("The",0),("quick",1),("brown",1),("fox",2)] & setter fmap . first' ..~ Prelude.length
[(3,0),(5,1),(5,1),(3,2)]

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

abst id ≡ id

abst f . abst g ≡ abst (f . g)

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

ixsetter :: ((i -> a -> b) -> s -> t) -> Ixsetter i s t a b

Build an Ixsetter from an indexed function.

ixsetter . ixsets ≡ id
ixsets . ixsetter ≡ id

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

iabst (const id) ≡ id

fmap (iabst $ const f) . (iabst $ const g) ≡ iabst (const $ f . g)

See Property.

Since: 0.0.3

closing :: (((s -> a) -> b) -> t) -> Setter s t a b

Every valid Colens is a Setter.

cloneSetter :: ASetter s t a b -> Setter s t a b

Clone a Setter.

cloneIxsetter :: Monoid k => AIxsetter k s t a b -> Ixsetter k s t a b

Clone an Ixsetter.

Setter1, Ixsetter1

type Setter1 s t a b = forall p. (Strong p, Traversing1 p, Mapping1 p) => Optic p s t a b

type Setter1' s a = Setter1 s s a a

type Ixsetter1 k s t a b = forall p. (Strong p, Traversing1 p, Mapping1 p) => Ixoptic p k s t a b

type Ixsetter1' k s a = Ixsetter1 k s s a a

setter1 :: ((a -> b) -> a -> t) -> Setter1 a t a b

TODO: Document

Since: 0.0.3

ixsetter1 :: ((i -> a -> b) -> a -> t) -> Ixsetter1 i a t a b

Build an Ixsetter1 from an indexed function.

Since: 0.0.3

Adjoint, Ixadjoint, Cxadjoint

type Adjoint s t a b = forall p. Adjoining p => Optic p s t a b

type Adjoint' s a = Adjoint s s a a

type Ixadjoint k s t a b = forall p. Adjoining p => Ixoptic p k s t a b

type Ixadjoint' k s a = Ixadjoint k s s a a

type Cxadjoint k s t a b = forall p. Adjoining p => Cxoptic p k s t a b

type Cxadjoint' k t b = Cxadjoint k t t b b

adjoint :: ((a -> b) -> s -> t) -> Adjoint s t a b

Obtain an Adjoint from a SEC.

adjoint abst ≡ adjointl (rightAdjunct abst)
adjoint abst ≡ adjointr (leftAdjunct abst)

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

abst id ≡ id

abst f . abst g ≡ abst (f . g)

ixadjoint :: ((i -> a -> b) -> s -> t) -> Ixadjoint i s t a b

Build an Ixadjoint from an indexed function.

ixadjoint . ixsets ≡ id
ixsets . ixadjoint ≡ id

adjointl :: (forall l u. Adjunction l u => (a -> u b) -> l s -> t) -> Adjoint s t a b

Construct an Adjoint from a Costar-side Van Laarhoven function.

adjointl f = corepresenting (f . leftAdjunct)

For each Adjoining profunctor p with Corep p = l and Rep p = u, the input is instantiated at the adjunction l ⊣ u.

adjointr :: (forall l u. Adjunction l u => (l a -> b) -> s -> u t) -> Adjoint s t a b

Construct an Adjoint from a Star-side Van Laarhoven function.

adjointr f = representing (f . rightAdjunct)

For each Adjoining profunctor p with Corep p = l and Rep p = u, the input is instantiated at the adjunction l ⊣ u.

ixadjoining :: Adjoining p => Ixoptic p k s t a b -> Cxoptic p k s t a b

Convert an indexed Adjoint optic to a coindexed one via the adjunction on the profunctor's representation functors.

Ix p k a b = p (k, a) b threads the index as a product on the left — the Corep (left adjoint) side. Cx p k a b = p a (k -> b) threads the index as a function on the right — the Rep (right adjoint) side. For an Adjoining profunctor, these are isomorphic because the adjunction Corep p ⊣ Rep p relates products and functions.

The conversion factors into three steps:

p (k, a) b
  ─── cosieve ──────> Corep p (k, a) -> b       (enter Corep)
  ─── leftAdjunct ──> (k, a) -> Rep p b          (cross via adjunction)
  ─── curry/distribute > a -> Rep p (k -> b)        (rearrange k)
  ─── tabulate ─────> p a (k -> b)               (exit Rep)

Compare ixcx, which achieves the same isomorphism using Strong + Closed via a mechanical shuffle through first' and closed. Here, the adjunction makes the structure explicit: Ix threads through the left adjoint, Cx through the right, and leftAdjunct/rightAdjunct is the bridge.

ixadjoining . cxadjoining ≡ id
cxadjoining . ixadjoining ≡ id

cxadjoining :: Adjoining p => Cxoptic p k s t a b -> Ixoptic p k s t a b

Convert a coindexed Adjoint optic to an indexed one via the adjunction. The inverse of ixadjoining.

The conversion factors into the reverse three steps:

p a (k -> b)
  ─── sieve ────────> a -> Rep p (k -> b)        (enter Rep)
  ─── undistribute/fmap > (k, a) -> Rep p b          (rearrange k)
  ─── rightAdjunct ──> Corep p (k, a) -> b        (cross via adjunction)
  ─── cotabulate ───> p (k, a) b                  (exit Corep)

cxadjoining . ixadjoining ≡ id
ixadjoining . cxadjoining ≡ id

Dual Constructors

Cosetter, Cxsetter

type Cosetter s t a b = forall p. (Coaffine p, Cotraversing p, Comapping p) => Optic p s t a b

type Cosetter' s a = Cosetter s s a a

type Cxsetter k s t a b = forall p. (Coaffine p, Cotraversing p, Comapping p) => Cxoptic p k s t a b

type Cxsetter' k t b = Cxsetter k t t b b

cosetter :: ((a -> b) -> s -> t) -> Cosetter s t a b

Obtain a Cosetter from a SEC.

The construction uses copure to extract the focal element and fmap (const a) to build a constant container, avoiding the Applicative (Coindex t b) constraint that would force b = t.

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

abst id ≡ id

abst f . abst g ≡ abst (f . g)

cxsetter :: ((i -> a -> b) -> s -> t) -> Cxsetter i s t a b

Obtain a Cxsetter from a coindexed SEC.

The coindex i threads through the right side of the profunctor via Cxoptic p i s t a b = p a (i -> b) -> p s (i -> t), which is equivalent to Cosetter s (i -> t) a (i -> b). The constructor delegates to cosetter on these function-wrapped types.

cxsets . cxsetter ≡ id
cxsetter . cxsets ≡ id

cloneCosetter :: ACosetter s t a b -> Cosetter s t a b

Clone a Cosetter.

cloneCxsetter :: Monoid i => ACxsetter i s t a b -> Cxsetter i s t a b

Clone a Cxsetter.

Cosetter1, Cxsetter1

type Cosetter1 s t a b = forall p. (Closed p, Cotraversing1 p, Comapping1 p) => Optic p s t a b

type Cosetter1' s a = Cosetter1 s s a a

type Cxsetter1 k s t a b = forall p. (Closed p, Cotraversing1 p, Comapping1 p) => Cxoptic p k s t a b

type Cxsetter1' k t b = Cxsetter1 k t t b b

Optics

Setter, Ixsetter

fmapped :: Functor f => Setter (f a) (f b) a b

Setter on each value of a functor.

contramapped :: Contravariant f => Setter (f b) (f a) a b

Setter on each value of a contravariant functor.

contramap ≡ over contramapped

>>> getPredicate (over contramapped (*2) (Predicate even)) 5
True
>>> getOp (over contramapped (*5) (Op show)) 100
"500"

imappedRep :: Representable f => Ixsetter (Rep f) (f a) (f b) a b

Ixsetter on each value of a representable functor.

>>> 1 :+ 2 & ixany imappedRep %~ bool 20 10 . getAny
20 :+ 10

domain :: Profunctor p => Setter (p b r) (p a r) a b

Map contravariantly over the input of a profunctor.

The most common profunctor to use this with is (->).

(domain ..~ f) g x ≡ g (f x)

>>> (domain ..~ show) length [1,2,3]
7

codomain :: Profunctor p => Setter (p r a) (p r b) a b

Map covariantly over the output of a profunctor.

The most common profunctor to use this with is (->).

(codomain ..~ f) g x ≡ f (g x)
codomain @(->) ≡ withColens closed closing

>>> (codomain ..~ show) length [1,2,3]
"3"

liftedM :: Monad m => Setter (m a) (m b) a b

Setter on each value of a monad.

liftedA :: Applicative f => Setter (f a) (f b) a b

Setter on each value of an applicative.

liftA ≡ setter liftedA

>>> setter liftedA Identity [1,2,3]
[Identity 1,Identity 2,Identity 3]
>>> set liftedA 2 (Just 1)
Just 2

forwarded :: Setter (ReaderT r2 m a) (ReaderT r1 m a) r1 r2

Setter on the local environment of a Reader.

Use to lift reader actions into a larger environment:

>>> runReader (ask & forwarded ..~ fst) (1,2)
1

censored :: MonadWriter w m => Setter' (m a) w

TODO: Document

zipped :: Setter (u -> v -> a) (u -> v -> b) a b

Setter on the codomain of a zipping function.

>>> ((,) & zipped ..~ swap) 1 2
(2,1)

modded :: (a -> Bool) -> Setter' (a -> b) b

TODO: Document

conditioned :: (a -> Bool) -> Setter' a a

Apply a function only when the given condition holds.

See also predicated & filtered.

Setter1, Ixsetter1

Adjoint, Ixadjoint

adjoined :: Adjoint a b a b

The identity for Adjoint optics.

adjoined = adjointl rightAdjunct
adjoined = adjointr leftAdjunct

mappedException :: (Exception e1, Exception e2) => Adjoint s s e1 e2

Map one exception into another as proposed in "A semantics for imprecise exceptions".

mappedException :: (Exception e1, Exception e2) => Adjoint s s e1 e2

Dual Optics

Cosetter, Cxsetter

Cosetter1, Cxsetter1

coliftedA :: Applicative f => Cosetter (f a) (f b) a b

TODO: Document

zipListed :: Cosetter [a] [b] a b

Variant of coliftedA specialized to zip-toListOf.

Useful because toListOf are not Coapplicative.

Since: 0.0.3

Operators

Setter, Ixsetter

set :: ASetter s t a b -> b -> s -> t

Set the focus of a Setter.

set o y (set o x a) ≡ set o y a
set o b = runIdentity . (o *~ Identity b)

ixset :: Monoid i => AIxsetter i s t a b -> (i -> b) -> s -> t

Set the focus of a Ixsetter.

Equivalent to ixsets with the current value ignored.

set o ≡ ixset o . const

Since: 0.0.3

sets :: ASetter s t a b -> (a -> b) -> s -> t

Set the focus of a Setter.

ixsets :: Monoid i => AIxsetter i s t a b -> (i -> a -> b) -> s -> t

Set the focus of a Ixsetter.

Since: 0.0.3

over :: ((a -> b) -> s -> t) -> (a -> b) -> s -> t

Map over an Optic.

over o id ≡ id
over o f . over o g ≡ over o (f . g)
over . setter ≡ id
over . resetter ≡ id

>>> over fmapped (+1) (Just 1)
Just 2
>>> over fmapped (*10) [1,2,3]
[10,20,30]
>>> over first (+1) (1,2)
(2,2)
>>> over first show (10,20)
("10",20)

Benchmark: 1.00x vs direct (Lens), 0.89x vs fmap (Traversal). See Data.Profunctor.Optic.Bench.

ixover :: Monoid i => Ixoptic (->) i s t a b -> (i -> a -> b) -> s -> t

Indexed over: apply a key-dependent function through an indexed optic.

Routes through Conjoin wrapping internally.

Benchmark: 1.08x vs direct mapWithKey (Conjoin overhead negligible). See Data.Profunctor.Optic.Bench.

Since: 0.0.3

Setter1, Ixsetter1

Adjoint

lifts :: Adjunction l u => ACotraversal l s t a b -> (l a -> b) -> s -> u t

Lift a Costar-side (co-Van Laarhoven) traversal through an adjunction to obtain a Star-side result.

lifts o f = leftAdjunct (cotraverseOf o f)

lowers :: Adjunction l u => ATraversal u s t a b -> (a -> u b) -> l s -> t

Lower a Star-side (Van Laarhoven) traversal through an adjunction to obtain a Costar-side result.

lowers o f = rightAdjunct (traverseOf o f)

alower :: Adjunction l u => ((a -> b) -> s -> t) -> (l a -> b) -> l s -> t

Convert a Star-side traversal to the Costar side through an adjunction, applied to an SEC.

alower abst = over adjuncted (aupper abst)

aupper :: Representable u => ((a -> b) -> s -> t) -> (a -> u b) -> s -> u t

Convert an SEC to a Star-side (Van Laarhoven) form.

For each index i of the representable functor u, the SEC abst is applied to the i-th component extracted from a -> u b.

Dual Operators

Cosetter, Cxsetter

coset :: ACosetter s t a b -> b -> s -> t

Set the focus of a Cosetter.

coset o b = (o /~ b) . Identity

cxset :: Monoid i => ACxsetter i s t a b -> (i -> b) -> s -> t

Set the focus of a Cxsetter.

Equivalent to cxsets with the current value ignored.

Since: 0.0.3

cosets :: ACosetter s t a b -> (a -> b) -> s -> t

Set the focus of a Cosetter.

cxsets :: Monoid i => ACxsetter i s t a b -> (i -> a -> b) -> s -> t

Set the focus of a Cxsetter.

Since: 0.0.3

cxover :: Monoid i => Cxoptic (->) i s t a b -> (i -> a -> b) -> s -> t

Coindexed over: apply a coindex-dependent function through a coindexed optic.

Routes through Conjoin wrapping internally (dual of overWithKey).

Benchmark: ~1.08x overhead (same Conjoin path as overWithKey). See Data.Profunctor.Optic.Bench.

Since: 0.0.3

Cosetter1, Cxsetter1

MTL

assigns :: MonadState s m => Optic (->) s s a b -> b -> m ()

Replace the target(s) of a settable in a monadic state.

ixassigns :: MonadState s m => Monoid i => Ixoptic (->) i s s a b -> (i -> b) -> m ()

Indexed assigns: replace the target(s) of an indexed settable in a monadic state.

Since: 0.0.3

modifies :: MonadState s m => Optic (->) s s a b -> (a -> b) -> m ()

Map over the target(s) of a Setter in a monadic state.

ixmodifies :: MonadState s m => Monoid i => Ixoptic (->) i s s a b -> (i -> a -> b) -> m ()

Indexed modifies: map over the target(s) of an indexed settable in a monadic state.

Since: 0.0.3

localizes :: MonadReader s m => Optic (->) s s a b -> (a -> b) -> m r -> m r

Modify the value of a Reader environment.

localizes l id a ≡ a
localizes l f . localizes l g ≡ localizes l (f . g)

>>> (1,1) & localizes first' (+1) (uncurry (+))
3
>>> "," & localizes (setter ($)) ("Hello" <>) (<> " world!")
"Hello, world!"

Compare forwarded.

tells :: MonadWriter w m => Monoid s => Optic (->) s w a b -> b -> m ()

Write to a fragment of a larger Writer format.

Re-exports