Data.Profunctor.Optic.Setter

Setter

type Setter s t a b = forall p. (Affine p, Traversing p, Mapping p) => Optic p s t a b

\( \mathsf{Functor}\;S\;A = \exists F : \mathsf{Functor}, S \equiv F\,A \)

type Setter' s a = Setter s s a a

type Resetter s t a b = forall p. (Coaffine p, Cotraversing p, Remapping p) => Optic p s t a b

\( \quad \mathsf{Resetter}\;S\;A = \exists n : \mathbb{N}, S \cong \mathsf{Fin}\,n \to A \)

See also section 3 on Kaleidoscopes here.

type Resetter' s a = Resetter s s a a

setter :: ((a -> b) -> s -> t) -> Setter s t a b

Obtain a Setter from a SEC.

To demote an optic to a semantic edit combinator, use the section (l ..~) or over l.

>>> [("The",0),("quick",1),("brown",1),("fox",2)] & setter fmap . first' ..~ Prelude.length
[(3,0),(5,1),(5,1),(3,2)]

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

abst id ≡ id

abst f . abst g ≡ abst (f . g)

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

ixsetter :: ((i -> a -> b) -> s -> t) -> Ixsetter i s t a b

Build an Ixsetter from an indexed function.

ixsetter . setsWithKey ≡ id
setsWithKey . ixsetter ≡ id

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

iabst (const id) ≡ id

fmap (iabst $ const f) . (iabst $ const g) ≡ iabst (const $ f . g)

See Property.

Since: 0.0.3

closing :: (((s -> a) -> b) -> t) -> Setter s t a b

Every valid Colens is a Setter.

resetter :: ((a -> t) -> s -> t) -> Resetter s t a t

Obtain a Resetter from a SEC.

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

abst id ≡ id

abst f . abst g ≡ abst (f . g)

Setter1

type Setter1 s t a b = forall p. (Strong p, Traversing1 p, Mapping1 p) => Optic p s t a b

type Setter1' s a = Setter1 s s a a

type Resetter1 s t a b = forall p. (Closed p, Cotraversing1 p, Remapping1 p) => Optic p s t a b

type Resetter1' s a = Resetter1 s s a a

setter1 :: ((a -> b) -> a -> t) -> Setter1 a t a b

TODO: Document

Since: 0.0.3

resetter1 :: ((a -> t) -> s -> t) -> Resetter1 s t a t

TODO: Document

Since: 0.0.3

Optics

cod :: Profunctor p => Setter (p r a) (p r b) a b

Map covariantly over the output of a profunctor.

The most common profunctor to use this with is (->).

(dom ..~ f) g x ≡ f (g x)
cod @(->) ≡ withColens closed closing

>>> (cod ..~ show) length [1,2,3]
"3"

dom :: Profunctor p => Setter (p b r) (p a r) a b

Map contravariantly over the input of a profunctor.

The most common profunctor to use this with is (->).

(dom ..~ f) g x ≡ g (f x)

>>> (dom ..~ show) length [1,2,3]
7

fmapped :: Functor f => Setter (f a) (f b) a b

Setter on each value of a functor.

imappedRep :: Representable f => Ixsetter (Rep f) (f a) (f b) a b

Ixsetter on each value of a representable functor.

>>> 1 :+ 2 & ixany imappedRep %~ bool 20 10 . getAny
20 :+ 10

contramapped :: Contravariant f => Setter (f b) (f a) a b

Setter on each value of a contravariant functor.

contramap ≡ over contramapped

>>> getPredicate (over contramapped (*2) (Predicate even)) 5
True
>>> getOp (over contramapped (*5) (Op show)) 100
"500"

liftedM :: Monad m => Setter (m a) (m b) a b

Setter on each value of a monad.

liftedA :: Applicative f => Setter (f a) (f b) a b

Setter on each value of an applicative.

liftA ≡ setter liftedA

>>> setter liftedA Identity [1,2,3]
[Identity 1,Identity 2,Identity 3]
>>> set liftedA 2 (Just 1)
Just 2

reliftedA :: Applicative f => Resetter (f a) (f b) a b

TODO: Document

reliftedF :: Apply f => Resetter1 (f a) (f b) a b

TODO: Document

Since: 0.0.3

zipListed :: Resetter [a] [b] a b

Variant of reliftedA specialized to zip-lists.

Useful because lists are not Coapplicative.

Since: 0.0.3

forwarded :: Setter (ReaderT r2 m a) (ReaderT r1 m a) r1 r2

Setter on the local environment of a Reader.

Use to lift reader actions into a larger environment:

>>> runReader (ask & forwarded ..~ fst) (1,2)
1

censored :: MonadWriter w m => Setter' (m a) w

TODO: Document

zipped :: Setter (u -> v -> a) (u -> v -> b) a b

Setter on the codomain of a zipping function.

>>> ((,) & zipped ..~ swap) 1 2
(2,1)

modded :: (a -> Bool) -> Setter' (a -> b) b

TODO: Document

cond :: (a -> Bool) -> Setter' a a

Apply a function only when the given condition holds.

See also predicated & filtered.

Operators

(.~) :: Optic (->) s t a b -> b -> s -> t infixr 4

Set the focus of an Optic.

(..~) :: Optic (->) s t a b -> (a -> b) -> s -> t infixr 4

Map over an Optic.

>>> (10,20) & first ..~ show
("10",20)

over :: Optic (->) s t a b -> (a -> b) -> s -> t

Map over an Optic.

over o id ≡ id 
over o f . over o g ≡ over o (f . g)
over . setter ≡ id
over . resetter ≡ id

>>> over fmapped (+1) (Just 1)
Just 2
>>> over fmapped (*10) [1,2,3]
[10,20,30]
>>> over first (+1) (1,2)
(2,2)
>>> over first show (10,20)
("10",20)

(%~) :: Monoid i => Ixoptic (->) i s t a b -> (i -> b) -> s -> t infixr 4

Set the focus of an indexed optic.

Note: This function is different from the equivalent in the lens package. The profunctor-optics equivalent of %~ from lens is ..~.

Since: 0.0.3

(%%~) :: Monoid i => Ixoptic (->) i s t a b -> (i -> a -> b) -> s -> t infixr 4

Map over an indexed optic.

Since: 0.0.3

overWithKey :: Monoid i => Ixoptic (->) i s t a b -> (i -> a -> b) -> s -> t

TODO: Document

Since: 0.0.3

(#~) :: Monoid i => Cxoptic (->) i s t a b -> (i -> b) -> s -> t infixr 4

Set the focus of a coindexed optic.

Since: 0.0.3

(##~) :: Monoid i => Cxoptic (->) i s t a b -> (i -> a -> b) -> s -> t infixr 4

Map over a coindexed optic.

Since: 0.0.3

reoverWithKey :: Monoid i => Cxoptic (->) i s t a b -> (i -> a -> b) -> s -> t

TODO: Document

Since: 0.0.3

set :: ASetter s t a b -> b -> s -> t

Set the focus of a Setter.

set o y (set o x a) ≡ set o y a
set o b = runIdentity . (o *~ Identity b)

sets :: ASetter s t a b -> (a -> b) -> s -> t

Set the focus of a Setter.

setWithKey :: Monoid i => AIxsetter i s t a b -> (i -> b) -> s -> t

Set the focus of a Ixsetter.

Equivalent to setsWithKey with the current value ignored.

set o ≡ setWithKey o . const

Since: 0.0.3

setsWithKey :: Monoid i => AIxsetter i s t a b -> (i -> a -> b) -> s -> t

Set the focus of a Ixsetter.

Since: 0.0.3

reset :: AResetter s t a b -> b -> s -> t

Set the focus of a Resetter.

reset o b = (o /~ b) . Identity

resets :: AResetter s t a b -> (a -> b) -> s -> t

Set the focus of a Resetter.

resetWithKey :: Monoid i => ARxsetter i s t a b -> (i -> b) -> s -> t

Set the focus of a Rxsetter.

Equivalent to resetsWithKey with the current value ignored.

Since: 0.0.3

resetsWithKey :: Monoid i => ARxsetter i s t a b -> (i -> a -> b) -> s -> t

Set the focus of a Rxsetter.

Since: 0.0.3

mtl

(.=) :: MonadState s m => Optic (->) s s a b -> b -> m () infix 4

Replace the target(s) of a settable in a monadic state.

This is an infiversion of assigns.

>>> execState (do first' .= 1; second' .= 2) (3,4)
(1,2)
>>> execState (bitraversed .= 3) (1,2)
(3,3)

(..=) :: MonadState s m => Optic (->) s s a b -> (a -> b) -> m () infix 4

Map over the target(s) of a Setter in a monadic state.

This is an infiversion of modifies.

>>> execState (do just ..= (+1) ) Nothing
Nothing
>>> execState (do first' ..= (+1) ;second' ..= (+2)) (1,2)
(2,4)
>>> execState (do bitraversed ..= (+1)) (1,2)
(2,3)

assigns :: MonadState s m => Optic (->) s s a b -> b -> m ()

A prefix alias for .=.

modifies :: MonadState s m => Optic (->) s s a b -> (a -> b) -> m ()

A prefix alias for ..=

locally :: MonadReader s m => Optic (->) s s a b -> (a -> b) -> m r -> m r

Modify the value of a Reader environment.

locally l id a ≡ a
locally l f . locally l g ≡ locally l (f . g)

>>> (1,1) & locally first' (+1) (uncurry (+))
3
>>> "," & locally (setter ($)) ("Hello" <>) (<> " world!")
"Hello, world!"

Compare forwarded.

scribe :: MonadWriter w m => Monoid s => Optic (->) s w a b -> b -> m ()

Write to a fragment of a larger Writer format.