Data.Profunctor.Optic.Traversal

Description

Left and right carriers are characterized by their representation functors (Rep and Corep). Iso sits at the top where both are trivial (Identity). Adjoint sits at the bottom where they form a full adjunction (Corep p ⊣ Rep p). A left-indexed (Ix) optic threads an index through the left adjoint functor (,) k, while a right-indexed (Cx) optic threads a coindex through the right adjoint functor (->) k.

Constructors

Traversal, Ixtraversal

type Traversal s t a b = forall p. (Affine p, Traversing p) => Optic p s t a b

type Traversal' s a = Traversal s s a a

type Ixtraversal k s t a b = forall p. (Affine p, Traversing p) => Ixoptic p k s t a b

type Ixtraversal' k s a = Ixtraversal k s s a a

atraversal :: ((a -> f b) -> s -> f t) -> ATraversal f s t a b

TODO: Document

traversing :: Traversable f => (s -> a) -> (s -> b -> t) -> Traversal (f s) (f t) a b

Obtain a Traversal by lifting a lens getter and setter into a Traversable functor.

 withLens o traversing ≡ traversed . o

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions constitute a legal lens:

sa (sbt s a) ≡ a

sbt s (sa s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

See Property.

The resulting optic can detect copies of the lens structure inside any Traversable container. For example:

>>> toListOf (traversing snd $ \(s,_) b -> (s,b)) [(0,'f'),(1,'o'),(2,'o'),(3,'b'),(4,'a'),(5,'r')]
"foobar"

ixtraversing :: Monoid k => Traversable f => (s -> (k, a)) -> (s -> b -> t) -> Ixtraversal k (f s) (f t) a b

Obtain a Ixtraversal by lifting an indexed lens getter and setter into a Traversable functor.

 withIxlens o ixtraversing ≡ ixtraversed . o

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions constitute a legal indexed lens:

snd . sia (sbt s a) ≡ a

sbt s (snd $ sia s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

See Property.

Since: 0.0.3

traversalVl :: (forall f. Applicative' f => (a -> f b) -> s -> f t) -> Traversal s t a b

Obtain a profunctor Traversal from a Van Laarhoven Traversal.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

abst pure ≡ pure

fmap (abst f) . abst g ≡ getCompose . abst (Compose . fmap f . g)

The traversal laws can be stated in terms of traverseOf:

traverseOf t (pure . f) ≡ pure (fmap f)

Compose . fmap (traverseOf t f) . traverseOf t g ≡ traverseOf t (Compose . fmap f . g)

See Property.

ixtraversalVl :: (forall f. Applicative f => (k -> a -> f b) -> s -> f t) -> Ixtraversal k s t a b

Lift an indexed VL traversal into an indexed profunctor traversal.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

kabst (const pure) ≡ pure

fmap (kabst $ const f) . (kabst $ const g) ≡ getCompose . kabst (const $ Compose . fmap f . g)

See Property.

Since: 0.0.3

beside :: Bitraversable r => Traversal s1 t1 a b -> Traversal s2 t2 a b -> Traversal (r s1 s2) (r t1 t2) a b

TODO: Document

reversing :: ATraversal (Backwards f) s t a b -> ATraversal f s t a b

TODO: Document

Since: 0.0.3

ix :: Monoid k => k -> Traversal s t a b -> Ixtraversal k s t a b

Iteratively index a traversal with an incrementing value.

>>> B.first getSum <$> ixtoListOf (ix (Sum 1) traversed) "foobar"
[(0,'f'),(1,'o'),(2,'o'),(3,'b'),(4,'a'),(5,'r')]
>>> ixtoListOf (noix traversed . ix "o" traversed) ["foo", "bar"]
[("",'f'),("o",'o'),("oo",'o'),("",'b'),("o",'a'),("oo",'r')]
>>> ixtoListOf (ix "x" traversed % ix "o" traversed) ["foo", "bar"]
[("",'f'),("o",'o'),("oo",'o'),("x",'b'),("xo",'a'),("xoo",'r')]
>>> B.first getSum <$> ixtoListOf (ix (Sum 3) traversed % ix (Sum 1) traversed) ["foo", "bar"]
[(0,'f'),(1,'o'),(2,'o'),(3,'b'),(4,'a'),(5,'r')]

Since: 0.0.3

noix :: Monoid k => Traversal s t a b -> Ixtraversal k s t a b

Lift a VL traversal into an indexed profunctor traversal that ignores its input.

Useful as the first optic in a chain when no indexed equivalent is at hand.

>>> B.first getSum <$> ixtoListOf (noix traversed . ix (Sum 1) traversed) ["foo", "bar"]
[(0,'f'),(1,'o'),(2,'o'),(0,'b'),(1,'a'),(2,'r')]
>>> B.first getSum <$> ixtoListOf (ix (Sum 1) traversed . noix traversed) ["foo", "bar"]
[(0,'f'),(0,'o'),(0,'o'),(0,'b'),(0,'a'),(0,'r')]

Since: 0.0.3

cloneTraversalVl :: Applicative f => ATraversal f s t a b -> (a -> f b) -> s -> f t

Extract the Van Laarhoven function that characterizes a Traversal.

Since: 0.0.3

cloneTraversal1Vl :: Apply f => ATraversal1 f s t a b -> (a -> f b) -> s -> f t

Extract the Van Laarhoven function that characterizes a Traversal1.

Since: 0.0.3

Traversal0, Ixtraversal0

type Traversal0 s t a b = forall p. Affine p => Optic p s t a b

type Traversal0' s a = Traversal0 s s a a

type Ixtraversal0 k s t a b = forall p. Affine p => Ixoptic p k s t a b

type Ixtraversal0' k s a = Ixtraversal0 k s s a a

traversal0 :: (s -> t + a) -> (s -> b -> t) -> Traversal0 s t a b

Obtain a Traversal0 from match and constructor functions.

Caution: In order for the Traversal0 to be well-defined, you must ensure that the input functions satisfy the following properties:

sta (sbt a s) ≡ either (Left . const a) Right (sta s)

either id (sbt s) (sta s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

More generally, a profunctor optic must be monoidal as a natural transformation:

o id ≡ id

o (Procompose p q) ≡ Procompose (o p) (o q)

See Property.

ixtraversal0 :: (s -> t + (k, a)) -> (s -> b -> t) -> Ixtraversal0 k s t a b

TODO: Document

Since: 0.0.3

traversal0' :: (s -> Maybe a) -> (s -> b -> s) -> Traversal0 s s a b

Obtain a Traversal0' from match and constructor functions.

ixtraversal0' :: (s -> Maybe (k, a)) -> (s -> a -> s) -> Ixtraversal0' k s a

TODO: Document

Since: 0.0.3

traversalVl0 :: (forall f. Functor f => (forall c. c -> f c) -> (a -> f b) -> s -> f t) -> Traversal0 s t a b

Transform a Van Laarhoven Traversal0 into a profunctor Traversal0.

ixtraversalVl0 :: (forall f. Functor f => (forall c. c -> f c) -> (k -> a -> f b) -> s -> f t) -> Ixtraversal0 k s t a b

Transform an indexed Van Laarhoven Traversal0 into an indexed profunctor Traversal0.

Since: 0.0.3

cloneTraversal0 :: ATraversal0 s t a b -> Traversal0 s t a b

Clone a Traversal0.

Since: 0.0.3

Traversal1, Ixtraversal1

type Traversal1 s t a b = forall p. (Strong p, Traversing1 p) => Optic p s t a b

type Traversal1' s a = Traversal1 s s a a

type Ixtraversal1 k s t a b = forall p. (Strong p, Traversing1 p) => Ixoptic p k s t a b

type Ixtraversal1' k s a = Ixtraversal1 k s s a a

traversing1 :: Traversable1 f => (s -> a) -> (s -> b -> t) -> Traversal1 (f s) (f t) a b

Obtain a Traversal by lifting a lens getter and setter into a Traversable functor.

 withLens o traversing ≡ traversed . o

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions constitute a legal lens:

sa (sbt s a) ≡ a

sbt s (sa s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

See Property.

The resulting optic can detect copies of the lens structure inside any Traversable container. For example:

>>> toListOf (traversing snd $ \(s,_) b -> (s,b)) [(0,'f'),(1,'o'),(2,'o'),(3,'b'),(4,'a'),(5,'r')]
"foobar"

Compare folding.

ixtraversing1 :: Monoid k => Traversable1 f => (s -> (k, a)) -> (s -> b -> t) -> Ixtraversal1 k (f s) (f t) a b

Obtain a Ixtraversal by lifting an indexed lens getter and setter into a Traversable functor.

 withIxlens o ixtraversing ≡ ixtraversed . o

Caution: In order for the generated optic to be well-defined, you must ensure that the input functions constitute a legal indexed lens:

snd . sia (sbt s a) ≡ a

sbt s (snd $ sia s) ≡ s

sbt (sbt s a1) a2 ≡ sbt s a2

See Property.

Since: 0.0.3

traversalVl1 :: (forall f. Apply f => (a -> f b) -> s -> f t) -> Traversal1 s t a b

Obtain a profunctor Traversal1 from a Van Laarhoven Traversal1.

Caution: In order for the generated family to be well-defined, you must ensure that the traversal1 law holds for the input function:

fmap (abst f) . abst g ≡ getCompose . abst (Compose . fmap f . g)

See Property.

ixtraversalVl1 :: (forall f. Apply f => (k -> a -> f b) -> s -> f t) -> Ixtraversal1 k s t a b

Obtain a profunctor Ixtraversal1 from a Van Laarhoven Ixtraversal1.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

kabst (const Identity) ≡ Identity

fmap (kabst $ const f) . (kabst $ const g) ≡ getCompose . kabst (const $ Compose . fmap f . g)

See Property.

Since: 0.0.3

beside1 :: Bitraversable1 r => Traversal1 s1 t1 a b -> Traversal1 s2 t2 a b -> Traversal1 (r s1 s2) (r t1 t2) a b

TODO: Document

Dual Constructors

Cotraversal, Cxtraversal

type Cotraversal s t a b = forall p. (Coaffine p, Cotraversing p) => Optic p s t a b

type Cotraversal' t b = Cotraversal t t b b

type Cxtraversal k s t a b = forall p. (Coaffine p, Cotraversing p) => Cxoptic p k s t a b

type Cxtraversal' k t b = Cxtraversal k t t b b

acotraversal :: ((f a -> b) -> f s -> t) -> ACotraversal f s t a b

TODO: Document

cotraversing :: Distributive g => (((s -> a) -> b) -> t) -> Cotraversal (g s) (g t) a b

Obtain a Cotraversal by embedding a continuation into a Distributive functor.

 withColens o cotraversing ≡ cotraversed . o

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

sabt ($ s) ≡ s

sabt (k -> f (k . sabt)) ≡ sabt (k -> f ($ k))

retraversing :: Distributive g => (b -> t) -> (b -> s -> a) -> Cotraversal (g s) (g t) a b

Obtain a Cotraversal by embedding a reversed lens getter and setter into a Distributive functor.

 withLens (re o) retraversing ≡ cotraversed . o

cotraversalVl :: (forall f. Coapplicative f => (f a -> b) -> f s -> t) -> Cotraversal s t a b

Obtain a profunctor Cotraversal from a Van Laarhoven Cotraversal.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

abst copure ≡ copure

abst f . fmap (abst g) ≡ abst (f . fmap g . getCompose) . Compose

The cotraversal laws can be restated in terms of cotraverseOf:

cotraverseOf o (f . copure) ≡  fmap f . copure

cotraverseOf o f . fmap (cotraverseOf o g) == cotraverseOf o (f . fmap g . getCompose) . Compose

See Property.

cxtraversalVl :: (forall f. Coapplicative f => (f a -> k -> b) -> f s -> t) -> Cxtraversal k s t a b

Lift a coindexed VL cotraversal into a coindexed profunctor cotraversal.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

aibst (const . copure) ≡ copure

(aibst $ const . f) . fmap (aibst $ const . g) ≡ aibst (const . f . fmap g . getCompose) . Compose

See Property.

Since: 0.0.3

cloneCotraversalVl :: Coapplicative f => ACotraversal f s t a b -> (f a -> b) -> f s -> t

Extract the Van Laarhoven function that characterizes a Cotraversal.

Since: 0.0.3

Cotraversal0, Cxtraversal0

type Cotraversal0 s t a b = forall p. Coaffine p => Optic p s t a b

type Cotraversal0' t b = Cotraversal0 t t b b

cotraversal0 :: (((s -> t + a) -> b) -> t) -> Cotraversal0 s t a b

Obtain a Cotraversal0 from its CPS representation.

The construction uses closed . right' (the dual of second' . right' used by traversal0). After closed, the profunctor type is p ((s -> t + a) -> t + a) ((s -> t + a) -> t + b). The right dimap uses stabt to reconstruct t, relying on the CPS guarantee that when sta s = Left t, stabt short-circuits without evaluating the callback result (lazy evaluation).

Since: 0.0.3

cloneCotraversal0 :: ACotraversal0 s t a b -> Cotraversal0 s t a b

Clone a Cotraversal0.

Since: 0.0.3

Cotraversal1, Cxtraversal1

type Cotraversal1 s t a b = forall p. (Closed p, Cotraversing1 p) => Optic p s t a b

type Cotraversal1' t b = Cotraversal1 t t b b

type Cxtraversal1 k s t a b = forall p. (Closed p, Cotraversing1 p) => Cxoptic p k s t a b

type Cxtraversal1' k t b = Cxtraversal1 k t t b b

cotraversing1 :: Distributive1 g => (((s -> a) -> b) -> t) -> Cotraversal1 (g s) (g t) a b

Obtain a Cotraversal1 by embedding a continuation into a Distributive1 functor.

 withColens o cotraversing1 ≡ cotraversed1 . o

Caution: In order for the generated optic to be well-defined, you must ensure that the input function satisfies the following properties:

sabt ($ s) ≡ s

sabt (k -> f (k . sabt)) ≡ sabt (k -> f ($ k))

retraversing1 :: Distributive1 g => (b -> t) -> (b -> s -> a) -> Cotraversal1 (g s) (g t) a b

Obtain a Cotraversal1 by embedding a reversed lens getter and setter into a Distributive1 functor.

 withLens (re o) retraversing1 ≡ cotraversed1 . o

cotraversalVl1 :: (forall f. Coapply f => (f a -> b) -> f s -> t) -> Cotraversal1 s t a b

Obtain a profunctor Cotraversal1 from a Van Laarhoven Cotraversal1.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

abst runIdentity ≡ runIdentity

abst f . fmap (abst g) ≡ abst (f . fmap g . getCompose) . Compose

The cotraversal1 laws can be restated in terms of cotraverseOf:

cotraverseOf o (f . runIdentity) ≡  fmap f . runIdentity

cotraverseOf o f . fmap (cotraverseOf o g) == cotraverseOf o (f . fmap g . getCompose) . Compose

See Property.

cxtraversalVl1 :: (forall f. Coapply f => (f a -> k -> b) -> f s -> t) -> Cxtraversal1 k s t a b

Obtain a profunctor Cxtraversal1 from a Van Laarhoven Cxtraversal1.

Caution: In order for the generated optic to be well-defined, you must ensure that the input satisfies the following properties:

aibst (const . runIdentity) ≡ runIdentity

(aibst $ const . f) . fmap (aibst $ const . g) ≡ aibst (const . f . fmap g . getCompose) . Compose

See Property.

Since: 0.0.3

cloneCotraversal1Vl :: Coapply f => ACotraversal1 f s t a b -> (f a -> b) -> f s -> t

Extract the Van Laarhoven function that characterizes a Cotraversal1.

Since: 0.0.3

Optics

Traversal, Ixtraversal

traversed :: Traversable f => Traversal (f a) (f b) a b

TODO: Document

itraversedRep :: Representable f => Traversable f => Ixtraversal (Rep f) (f a) (f b) a b

TODO: Document

bitraversed :: Bitraversable f => Traversal (f a a) (f b b) a b

Traverse both parts of a Bitraversable container with matching types.

>>> traverseOf bitraversed (pure . length) (Right "hello")
Right 5
>>> traverseOf bitraversed (pure . length) ("hello","world")
(5,5)
>>> ("hello","world") ^. bitraversed
"helloworld"

bitraversed :: Traversal (a , a) (b , b) a b
bitraversed :: Traversal (a + a) (b + b) a b

anulled :: Traversal0' s a

TODO: Document

selected :: (a -> Bool) -> Traversal0' (a, b) b

TODO: Document

duplicated :: Traversal1 a b a b

Duplicate the results of a Traversal.

>>> toListOf (bitraversed . duplicated) ("hello","world")
["hello","hello","world","world"]

repeated :: Traversal1' a a

Obtain a Traversal1' by repeating the input forever.

repeat ≡ toListOf repeated

>>> take 5 $ 5 ^.. repeated
[5,5,5,5,5]

repeated :: Fold1 a a

iterated :: (a -> a) -> Traversal1' a a

x ^. iterated f returns an infinite Traversal1' of repeated applications of f to x.

toListOf (iterated f) a ≡ iterate f a

>>> take 3 $ (1 :: Int) ^.. iterated (+1)
[1,2,3]

cycling :: Apply f => ATraversal' f s a -> ATraversal' f s a

Transform a Traversal1' into a Traversal1' that loops over its elements repeatedly.

>>> take 7 $ (1 :| [2,3]) ^.. cycling traversed1
[1,2,3,1,2,3,1]

Traversal1, Ixtraversal1

traversed1 :: Traversable1 t => Traversal1 (t a) (t b) a b

Obtain a Traversal1 from a Traversable1 functor.

bitraversed1 :: Bitraversable1 r => Traversal1 (r a a) (r b b) a b

Traverse both parts of a Bitraversable1 container with matching types.

>>> ('h' :| "ello", 'w' :| "orld") & bitraversed1 **~ pure . NE.length
(5,5)

Dual Optics

Cotraversal, Cxtraversal

cotraversed :: Distributive f => Cotraversal (f a) (f b) a b

TODO: Document

unforked :: Cotraversal1 (a + a) (b + b) a b

TODO: Document

>>> cotraverseOf unforked (foldMap id) $ Left "foo" :| [Right "bar"]
Left "foo"
>>> cotraverseOf unforked (foldMap id) $ Right "foo" :| [Right "bar"]
Right "foobar"

Since: 0.0.3

Cotraversal1, Cxtraversal1

cotraversed1 :: Distributive1 f => Cotraversal1 (f a) (f b) a b

TODO: Document

'cotraversed1' :: 'Cotraversal1' [a] [b] a b

Operators

Traversal, Ixtraversal

traverseOf :: ATraversal f s t a b -> (a -> f b) -> s -> f t

Traverse over a Traversal.

Benchmark: 0.89x vs direct fmap — GHC optimizes Star carrier well. See Data.Profunctor.Optic.Bench.

ixtraverseOf :: Monoid k => AIxtraversal f k s t a b -> (k -> a -> f b) -> s -> f t

Traverse over an Ixtraversal.

ixtraverseOf o f = curry (traverseOf o $ uncurry f) mempty

Since: 0.0.3

sequenceOf :: Applicative f => ATraversal f s t (f a) a -> s -> f t

TODO: Document

ixsequenceOf :: Monoid k => Applicative f => AIxtraversal f k s t (f a) a -> s -> f t

Indexed sequenceOf.

Since: 0.0.3

reverseOf :: ATraversal (Backwards f) s t a b -> (a -> f b) -> s -> f t

This allows you to traverse the elements of a Traversing or Traversing1 optic in the opposite order.

This will preserve indexes on Indexed types and for example will give you the elements of a (finite) Fold or Traversal in the opposite order.

This has no practical effect on a View, Setter, Lens or Iso.

Since: 0.0.3

mapAccumLOf :: ATraversal (State r) s t a b -> (r -> a -> (r, b)) -> r -> s -> (r, t)

Generalize mapAccumL to a Traversing or Traversing1 optic.

mapAccumL ≡ mapAccumLOf traverse

mapAccumLOf accumulates State from left to right.

Since: 0.0.3

mapAccumROf :: ATraversal (Backwards (State r)) s t a b -> (r -> a -> (r, b)) -> r -> s -> (r, t)

Generalize mapAccumR to a Traversing or Traversing1 optic.

mapAccumR ≡ mapAccumROf traverse

mapAccumROf accumulates State from right to left.

Since: 0.0.3

ixmapAccumLOf :: Monoid k => AIxtraversal (State r) k s t a b -> (k -> r -> a -> (r, b)) -> r -> s -> (r, t)

Indexed mapAccumLOf. Accumulates state from left to right, threading the index.

Since: 0.0.3

ixmapAccumROf :: Monoid k => AIxtraversal (Backwards (State r)) k s t a b -> (k -> r -> a -> (r, b)) -> r -> s -> (r, t)

Indexed mapAccumROf. Accumulates state from right to left, threading the index.

Since: 0.0.3

scanl1Of :: ATraversal (State (Maybe a)) s t a a -> (a -> a -> a) -> s -> t

Scan left over a Traversing or Traversing1 optic.

scanl1 ≡ scanl1Of traverse

Since: 0.0.3

scanr1Of :: ATraversal (Backwards (State (Maybe a))) s t a a -> (a -> a -> a) -> s -> t

Scan left over a Traversing or Traversing1 optic.

scanr1 ≡ scanr1Of traverse

Since: 0.0.3

Traversal0, Ixtraversal0

matches :: ATraversal0 s t a b -> s -> t + a

Test whether the optic matches or not.

>>> matches just (Just 2)
Right 2
>>> matches just (Nothing :: Maybe Int) :: Either (Maybe Bool) Int
Left Nothing

ixmatches :: Monoid k => AIxtraversal0 k s t a b -> s -> t + (k, a)

Indexed matches. Returns the index along with the matched value.

Since: 0.0.3

Traversal1, Ixtraversal1

Dual Operators

Cotraversal, Cxtraversal

cotraverseOf :: ACotraversal f s t a b -> (f a -> b) -> f s -> t

Cotraverse over a Cotraversal.

cxtraverseOf :: Monoid k => ACxtraversal f k s t a b -> (k -> f a -> b) -> f s -> t

Cotraverse over a Cxtraversal.

cxtraverseOf o f = flip (cotraverseOf o $ flip f) mempty

Since: 0.0.3

collectOf :: Coapply f => ACotraversal f s t a (f a) -> f s -> t

TODO: Document

>>> collectOf cotraversed1 ["xxx","ooo"]
["xo","xo","xo"]
>>> collectOf left' (1, Left "foo") :: Either (Int8, String) String
Left (1,"foo")
>>> collectOf left' (1, Right "foo")
Right "foo"

cxcollectOf :: Monoid k => Coapply f => ACxtraversal f k s t a (f a) -> f s -> t

Coindexed collectOf.

Since: 0.0.3

Cotraversal0, Cxtraversal0

Cotraversal1, Cxtraversal1

Arrow-style combinators

arr :: Traversing p => (a -> b) -> p a b

TODO: Document

coarr :: Cotraversing p => (a -> b) -> p a b

TODO: Document

(***) :: Traversing1 p => p a1 b1 -> p a2 b2 -> p (a1, a2) (b1, b2) infixr 3

Profunctor variant of ***.

(&&&) :: Traversing1 p => p a b1 -> p a b2 -> p a (b1, b2) infixr 3

Profunctor variant of &&&.

(<<*>>) :: Traversing1 p => p a (b -> c) -> p a b -> p a c infixl 4

Profunctor variant of <*>.

liftR2 :: Traversing1 p => (b -> c -> d) -> p a b -> p a c -> p a d

(+++) :: Cotraversing1 p => p a1 b1 -> p a2 b2 -> p (a1 + a2) (b1 + b2) infixr 2

Profunctor variant of +++.

(|||) :: Cotraversing1 p => p a1 b -> p a2 b -> p (a1 + a2) b infixr 2

Profunctor variant of |||.

divide :: Traversing1 p => p a1 b -> p a2 b -> p (a1, a2) b

TODO: Document

divideWith :: Traversing1 p => (a -> (a1, a2)) -> p a1 b -> p a2 b -> p a b

Profunctor variant of divideWith.

cochoose :: Traversing1 p => p a b1 -> p a b2 -> p a (b1, b2)

TODO: Document

cochooseWith :: Traversing1 p => ((b1, b2) -> b) -> p a b1 -> p a b2 -> p a b

TODO: Document

choose :: Cotraversing1 p => p a1 b -> p a2 b -> p (a1 + a2) b

TODO: Document

chooseWith :: Cotraversing1 p => (a -> a1 + a2) -> p a1 b -> p a2 b -> p a b

Profunctor variant of chooseWith.

codivide :: Cotraversing1 p => p a b1 -> p a b2 -> p a (b1 + b2)

TODO: Document

codivideWith :: Cotraversing1 p => ((b1 + b2) -> b) -> p a b1 -> p a b2 -> p a b

TODO: Document

pappend :: Traversing1 p => p a b -> p a b -> p a b

TODO: Document

Reexports

class Profunctor p => Strong (p :: Type -> Type -> Type) where

Generalizing Star of a strong Functor

Note: Every Functor in Haskell is strong with respect to (,).

This describes profunctor strength with respect to the product structure of Hask.

http://www.riec.tohoku.ac.jp/~asada/papers/arrStrMnd.pdf

Minimal complete definition

first' | second'

Methods

first' :: p a b -> p (a, c) (b, c)

Laws:

first' ≡ dimap swap swap . second'
lmap fst ≡ rmap fst . first'
lmap (second' f) . first' ≡ rmap (second' f) . first'
first' . first' ≡ dimap assoc unassoc . first' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

second' :: p a b -> p (c, a) (c, b)

Laws:

second' ≡ dimap swap swap . first'
lmap snd ≡ rmap snd . second'
lmap (first' f) . second' ≡ rmap (first' f) . second'
second' . second' ≡ dimap unassoc assoc . second' where
  assoc ((a,b),c) = (a,(b,c))
  unassoc (a,(b,c)) = ((a,b),c)

class Profunctor p => Choice (p :: Type -> Type -> Type) where

The generalization of Costar of Functor that is strong with respect to Either.

Note: This is also a notion of strength, except with regards to another monoidal structure that we can choose to equip Hask with: the cocartesian coproduct.

Minimal complete definition

left' | right'

Methods

left' :: p a b -> p (Either a c) (Either b c)

Laws:

left' ≡ dimap swapE swapE . right' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Left ≡ lmap Left . left'
lmap (right f) . left' ≡ rmap (right f) . left'
left' . left' ≡ dimap assocE unassocE . left' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

right' :: p a b -> p (Either c a) (Either c b)

Laws:

right' ≡ dimap swapE swapE . left' where
  swapE :: Either a b -> Either b a
  swapE = either Right Left
rmap Right ≡ lmap Right . right'
lmap (left f) . right' ≡ rmap (left f) . right'
right' . right' ≡ dimap unassocE assocE . right' where
  assocE :: Either (Either a b) c -> Either a (Either b c)
  assocE (Left (Left a)) = Left a
  assocE (Left (Right b)) = Right (Left b)
  assocE (Right c) = Right (Right c)
  unassocE :: Either a (Either b c) -> Either (Either a b) c
  unassocE (Left a) = Left (Left a)
  unassocE (Right (Left b)) = Left (Right b)
  unassocE (Right (Right c)) = Right c

class Profunctor p => Closed (p :: Type -> Type -> Type) where

A strong profunctor allows the monoidal structure to pass through.

A closed profunctor allows the closed structure to pass through.

Methods

closed :: p a b -> p (x -> a) (x -> b)

Laws:

lmap (. f) . closed ≡ rmap (. f) . closed
closed . closed ≡ dimap uncurry curry . closed
dimap const ($()) . closed ≡ id

class (Sieve p (Rep p), Strong p) => Representable (p :: Type -> Type -> Type) where

A Profunctor p is Representable if there exists a Functor f such that p d c is isomorphic to d -> f c.

Associated Types

type Rep (p :: Type -> Type -> Type) :: Type -> Type

Methods

tabulate :: (d -> Rep p c) -> p d c

Laws:

tabulate . sieve ≡ id
sieve . tabulate ≡ id

class (Cosieve p (Corep p), Costrong p) => Corepresentable (p :: Type -> Type -> Type) where

A Profunctor p is Corepresentable if there exists a Functor f such that p d c is isomorphic to f d -> c.

Associated Types

type Corep (p :: Type -> Type -> Type) :: Type -> Type

Methods

cotabulate :: (Corep p d -> c) -> p d c

Laws:

cotabulate . cosieve ≡ id
cosieve . cotabulate ≡ id