Data.Sequence.Optic

Documentation

slicedTo :: Int -> Traversal' (Seq a) a

Traverse the first n elements of a Seq

>>> fromList [1,2,3,4,5] ^.. slicedTo 2
[1,2]

>>> fromList [1,2,3,4,5] & slicedTo 2 %~ (*10)
fromList [10,20,3,4,5]

>>> fromList [1,2,4,5,6] & slicedTo 10 .~ 0
fromList [0,0,0,0,0]

slicedFrom :: Int -> Traversal' (Seq a) a

Traverse all but the first n elements of a Seq

>>> fromList [1,2,3,4,5] ^.. slicedFrom 2
[3,4,5]

>>> fromList [1,2,3,4,5] & slicedFrom 2 %~ (*10)
fromList [1,2,30,40,50]

>>> fromList [1,2,3,4,5] & slicedFrom 10 .~ 0
fromList [1,2,3,4,5]

sliced :: Int -> Int -> Traversal' (Seq a) a

Traverse all the elements numbered from i to j of a Seq

>>> fromList [1,2,3,4,5] & sliced 1 3 %~ (*10)
fromList [1,20,30,4,5]

>>> fromList [1,2,3,4,5] ^.. sliced 1 3
[2,3]

>>> fromList [1,2,3,4,5] & sliced 1 3 .~ 0
fromList [1,0,0,4,5]

viewedl :: Iso (Seq a) (Seq b) (ViewL a) (ViewL b)

A Seq is isomorphic to a ViewL

viewl m ≡ m ^. viewedl

>>> Seq.fromList [1,2,3] ^. viewedl
1 :< fromList [2,3]

>>> Seq.empty ^. viewedl
EmptyL

>>> EmptyL ^. re viewedl
fromList []

>>> review viewedl $ 1 Seq.:< fromList [2,3]
fromList [1,2,3]

viewedr :: Iso (Seq a) (Seq b) (ViewR a) (ViewR b)

A Seq is isomorphic to a ViewR

viewr m ≡ m ^. viewedr

>>> Seq.fromList [1,2,3] ^. viewedr
fromList [1,2] :> 3

>>> Seq.empty ^. viewedr
EmptyR

>>> EmptyR ^. re viewedr
fromList []

>>> review viewedr $ fromList [1,2] Seq.:> 3
fromList [1,2,3]